### 2022-06-04 #4453 @editorialbot
[PRE REVIEW]: Eureka!: An End-to-End Pipeline for JWST Time-Series Observations
**Submitting author:** @taylorbell57 (<a href="http://orcid.org/0000-0003-4177-2149">Taylor Bell</a>)
**Repository:** https://github.com/kevin218/Eureka
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1
