### 2022-06-06 4bc5bee
Adding flake8 settings to setup.cfg
