+++ setup.cfg
@@ -1,0 +1,3 @@
+[flake8]
+exclude = ./build/**/*,./docs/**/*,.vscode/*,**/site-packages/**/*,**/*niriss*.py
+per-file-ignores = **/__init__.py:F401,**/manageevent.py:F841,./demos/**/run_*.py:E402,**/test_*.py:E402,**/*.py:E722,**/readECF.py:F401,**/readECF.py:E722
