### 2022-06-06 a033fb8
pushes call from s3_reduce to s3_reduce_niriss if appropriate
