### 2022-06-06 7ec2126
change default profile to median for opt extraction
