### 2022-06-06 8d5c124
add isplots command to niriss_class
