### 2022-06-06 744c874
Selecting relevant flake8 errors
