+++ setup.cfg
@@ -1,5 +1,5 @@
 [flake8]
 # Error codes explained here: https://pycodestyle.pycqa.org/en/latest/intro.html#error-codes
-select = E1,E2,E3,E4,E5,E7,W6,F
+select = E1,E3,E4,E5,E7,W6,F
 exclude = ./build/**/*,./docs/**/*,.vscode/*,**/site-packages/**/*,**/*niriss*.py
 per-file-ignores = **/__init__.py:F401,**/manageevent.py:F841,./demos/**/run_*.py:E402,**/test_*.py:E402,**/*.py:E722,**/readECF.py:F401,**/readECF.py:E722
