### 2022-06-06 21944d2
Adding back E2 to flake8, better per-file-ignores
