+++ setup.cfg
@@ -1,5 +1,18 @@
 [flake8]
 # Error codes explained here: https://pycodestyle.pycqa.org/en/latest/intro.html#error-codes
-select = E1,E3,E4,E5,E7,W6,F
+select = E1,E2,E3,E4,E5,E7,W6,F
+ignore = W504,W503,E704,E226,E722
 exclude = ./build/**/*,./docs/**/*,.vscode/*,**/site-packages/**/*,**/*niriss*.py
-per-file-ignores = **/__init__.py:F401,**/manageevent.py:F841,./demos/**/run_*.py:E402,**/test_*.py:E402,**/*.py:E722,**/readECF.py:F401,**/readECF.py:E722
+per-file-ignores = 
+	# Need to have unused imports in the init files
+	**/__init__.py:F401
+	# Using exec commands, so variable is actually used
+	**/manageevent.py:F841
+	# Need to allow prepending to PATH
+	./demos/**/run_*.py:E402
+	# Need to protect against crashes if jwst not installed
+	**/test_*.py:E402
+	# Keep the unused numpy so users can pass in numpy values (e.g. np.inf)
+	**/readECF.py:F401
+	# Keep the aligned values to explain what they are
+	**/hst_scan.py:E241
