### 2022-06-06 3b9bb1d
Only run flake8 test on PRs
