+++ .github/workflows/pr_testing.yml
@@ -1,4 +1,4 @@
-on: push
+on: pull_request
 name: PR Testing
 jobs:
   lint:
