+++ setup.cfg
@@ -2,17 +2,30 @@
 # Error codes explained here: https://pycodestyle.pycqa.org/en/latest/intro.html#error-codes
 select = E1,E2,E3,E4,E5,E7,W6,F
 ignore = W504,W503,E704,E226,E722
-exclude = ./build/**/*,./docs/**/*,.vscode/*,**/site-packages/**/*,**/*niriss*.py
-per-file-ignores = 
-	# Need to have unused imports in the init files
-	**/__init__.py:F401
-	# Using exec commands, so variable is actually used
-	**/manageevent.py:F841
-	# Need to allow prepending to PATH
-	./demos/**/run_*.py:E402
-	# Need to protect against crashes if jwst not installed
-	**/test_*.py:E402
-	# Keep the unused numpy so users can pass in numpy values (e.g. np.inf)
-	**/readECF.py:F401
-	# Keep the aligned values to explain what they are
-	**/hst_scan.py:E241
+exclude =
+    # Ignore built files
+    ./build/**/*
+    **/site-packages/**/*
+    # Ignore docs
+    ./docs/**/*
+    # Ignore VS Code files
+    .vscode/*
+    # Ignore run_eureka.py files in demos folder
+    **/run_eureka*.py
+    # Temporarily ignore NIRISS files
+    **/*niriss*.py
+per-file-ignores =
+    # Need to have unused imports in the init files
+    **/__init__.py:F401
+    # Using exec commands, so variable is actually used
+    **/manageevent.py:F841
+    # Need to allow prepending to PATH
+    ./demos/**/run_*.py:E402
+    # Need to protect against crashes if jwst not installed
+    **/test_*.py:E402
+    # Keep the unused numpy so users can pass in numpy values (e.g. np.inf)
+    **/readECF.py:F401
+    # Keep the aligned values to explain what they are
+    **/hst_scan.py:E241
+    # Temporarily ignore unused s1 error until we have an S1 test
+    **/test_MIRI.py:F401,E402
