### 2022-06-06 ad71120
Adding pytest to github action
