+++ .github/workflows/pr_testing.yml
@@ -1,5 +1,9 @@
-on: pull_request
 name: PR Testing
+
+on:
+  pull_request:
+    types: [opened,reopened,synchronize]
+
 jobs:
   lint:
     name: Flake8 Linting
@@ -8,5 +12,18 @@ jobs:
     - uses: actions/checkout@v2
     - uses: tonybajan/flake8-check-action@v1.2.0
       with:
+        select: E1,E2,E3,E4,E5,E7,W6,F
         maxlinelength: 100
         repotoken: ${{ secrets.GITHUB_TOKEN }}
+  test:
+    name: pytest Testing
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-python@v1
+        with:
+          python-version: 3.9
+      - name: Setup Environment
+        run: pip install .[jwst,hst,test]
+      - name: Run tests
+        run: pytest tests
