### 2022-06-06 82352d0
Fixing Flake8 ignoring setup.cfg settings
