### 2022-06-06 eb7a9f4
Setting CRDS settings in automated pytest
