+++ setup.cfg
@@ -29,3 +29,5 @@ per-file-ignores =
     **/hst_scan.py:E241
     # Temporarily ignore unused s1 error until we have an S1 test
     **/test_MIRI.py:F401,E402
+    # Temporarily ignore undefined mask2 error until we fix that
+    background.py:F821
