### 2022-06-06 ca2baa1
Removing gcc install from pytest action
