### 2022-06-06 29f874d
Trying csh's setenv instead of sh's export
