### 2022-06-06 e406eaf
Properly set env vars, add back select to f8 actn
