### 2022-06-06 83f89ec
Set GH action environment variables differently
