### 2022-06-06 5db675c
Separate flake8 and pytest, run action on releases
