+++ .github/workflows/flake8_linting.yml
@@ -1,0 +1,22 @@
+name: Flake8 Linting
+
+on:
+  pull_request:
+    types: [opened,reopened,synchronize]
+  push:
+    branches:    
+      - 'main'
+      - 'releases/**'
+
+jobs:
+  lint:
+    name: Flake8 Linting
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+    - uses: tonybajan/flake8-check-action@v1.2.0
+      with:
+        select: E1,E2,E3,E4,E5,E7,W6,F
+        ignore: W504,W503,E704,E226,E722
+        maxlinelength: 100
+        repotoken: ${{ secrets.GITHUB_TOKEN }}
