+++ .github/workflows/pytest_testing.yml
@@ -10,7 +10,6 @@ on:
 
 jobs:
   test:
-    name: Pytest Testing
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
@@ -19,9 +18,12 @@ jobs:
           python-version: 3.9
       - name: Setup Environment
         run: |
+          python -m pip install --upgrade pip
           pip install .[jwst,hst,test]
           mkdir $HOME/crds_cache
           echo "CRDS_PATH=$HOME/crds_cache" >> $GITHUB_ENV
           echo "CRDS_SERVER_URL=https://jwst-crds.stsci.edu" >> $GITHUB_ENV
       - name: Run tests
-        run: pytest tests
+        run: pytest tests --cov=. --cov-report xml
+#      - name: Upload coverage reports to Codecov with GitHub Action
+#        uses: codecov/codecov-action@v2
