+++ .github/workflows/flake8_linting.yml
@@ -10,7 +10,6 @@ on:
 
 jobs:
   lint:
-    name: Flake8 Linting
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v2
