### 2022-06-06 fb23327
Adding Eureka logo to README
