+++ .github/workflows/pytest_testing.yml
@@ -10,6 +10,7 @@ on:
 
 jobs:
   test:
+    name: pytest
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
