+++ .github/workflows/flake8_linting.yml
@@ -10,6 +10,7 @@ on:
 
 jobs:
   lint:
+    name: flake8
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v2
