### 2022-06-06 dc65a3e
Adding Eureka logo to docs
