### 2022-06-06 385b27f
Resolving mask2 undefined error in background.py
