### 2022-06-07 7df8002
Changing setup.py to setup.cfg
