+++ setup.cfg
@@ -1,3 +1,79 @@
+[metadata]
+name = eureka
+version = 0.0.1
+author = Eureka! pipeline developers
+author_email = kbstevenson@gmail.com
+description = A data reduction and analysis pipeline intended for time-series observations with the James Webb Space Telescope (JWST).
+long_description = file: README.md
+long_description_content_type = text/markdown
+license = MIT License
+license_file = LICENSE
+url = https://github.com/kevin218/Eureka
+project_urls =
+    Bug Tracker = https://github.com/kevin218/Eureka/issues
+    Source Code = https://github.com/kevin218/Eureka
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: BSD License
+    Operating System :: OS Independent
+
+
+[options]
+packages = find:
+zip_safe = True
+python_requires= >=3.8
+setup_requires =
+    cython
+install_requires =
+    astraeus @ git+https://github.com/kevin218/Astraeus@main
+    astropy
+    batman-package
+    bokeh
+    ccdproc
+    celerite # Needed for GP
+    corner
+    crds
+    cython
+    dynesty > 1.0
+    emcee > 3.0.0
+    george # Needed for GP
+    gwcs
+    h5py < 3.2
+    lmfit
+    matplotlib
+    numpy >= 1.20.0
+    pandas
+    photutils
+    scipy >= 1.4.0 # Needed for scipy.fft
+    stdatamodels
+    svo_filters
+    tqdm
+
+
+[options.extras_require]
+jwst =
+    jwst==1.3.3
+    stcal
+    asdf>=2.7.1,<2.11.0
+hst =
+    # Need the GitHub version as 0.2.6 is required for python>=3.10, but 0.2.6 is not yet on PyPI
+    image_registration @ git+https://github.com/keflavich/image_registration@master
+docs =
+    myst-parser
+    nbsphinx
+    numpydoc
+    recommonmark
+    sphinx
+    sphinx-automodapi
+    sphinx-rtd-theme
+test =
+    codecov
+    flake8
+    pytest
+    pytest-cov
+    pytest-doctestplus
+
+
 [flake8]
 # Error codes explained here: https://pycodestyle.pycqa.org/en/latest/intro.html#error-codes
 select = E1,E2,E3,E4,E5,E7,W6,F
