### 2022-06-07 1964aac
moving trace routine into niriss.py (and removing repeat functions)
