### 2022-06-07 3005829
trace niriss to own routine in niriss.py
