### 2022-06-07 73f5e37
better naming of trace tables (rather than tab1/tab2)
