+++ eureka/S3_data_reduction/miri.py
@@ -1,8 +1,8 @@
 import numpy as np
-import os
 from astropy.io import fits
 import astraeus.xarrayIO as xrio
 from . import nircam
+from ..lib.util import read_time
 
 
 def read(filename, data, meta):
@@ -70,12 +70,7 @@ def read(filename, data, meta):
 
     # Record integration mid-times in BJD_TDB
     if (hasattr(meta, 'time_file') and meta.time_file is not None):
-        fname = os.path.join(meta.topdir,
-                             os.sep.join(meta.time_file.split(os.sep)))
-        if meta.firstFile:
-            print('  Note: Using the time stamps from:\n'+fname)
-        time = np.loadtxt(fname).flatten()[data.attrs['intstart']-1:
-                                           data.attrs['intend']-1]
+        time = read_time(meta, data)
     elif len(int_times['int_mid_BJD_TDB']) == 0:
         if meta.firstFile:
             print('  WARNING: The timestamps for the simulated MIRI data are '
