+++ eureka/S3_data_reduction/nirspec.py
@@ -1,9 +1,9 @@
 # NIRSpec specific rountines go here
 import numpy as np
-import os
 from astropy.io import fits
 import astraeus.xarrayIO as xrio
 from . import nircam, sigrej
+from ..lib.util import read_time
 
 
 def read(filename, data, meta):
@@ -61,12 +61,7 @@ def read(filename, data, meta):
 
     # Record integration mid-times in BJD_TDB
     if (hasattr(meta, 'time_file') and meta.time_file is not None):
-        fname = os.path.join(meta.topdir,
-                             os.sep.join(meta.time_file.split(os.sep)))
-        if meta.firstFile:
-            print('  Note: Using the time stamps from:\n'+fname)
-        time = np.loadtxt(fname).flatten()[data.attrs['intstart']-1:
-                                           data.attrs['intend']-1]
+        time = read_time(meta, data)
     elif len(int_times['int_mid_BJD_TDB']) == 0:
         # There is no time information in the simulated NIRSpec data
         print('  WARNING: The timestamps for the simulated NIRSpec data are '
