+++ README.md
@@ -1,8 +1,9 @@
-<img src="docs/media/Eureka_logo.png" width="128px"/>
+<img src="docs/media/Eureka_logo.png" width="256px"/>
 
 [![Documentation Status](https://readthedocs.org/projects/eurekadocs/badge/?version=latest)](https://eurekadocs.readthedocs.io/en/latest/?badge=latest)
 [![pytest](https://github.com/kevin218/Eureka/actions/workflows/pytest_testing.yml/badge.svg)](https://github.com/kevin218/Eureka/actions/workflows/pytest_testing.yml)
 [![flake8](https://github.com/kevin218/Eureka/actions/workflows/flake8_linting.yml/badge.svg)](https://github.com/kevin218/Eureka/actions/workflows/flake8_linting.yml)
+[![codecov](https://codecov.io/gh/kevin218/Eureka/branch/pr_testing/graph/badge.svg?token=2Y2KK7DZWL)](https://codecov.io/gh/kevin218/Eureka)
 
 **Welcome to Eureka!**
 
@@ -22,7 +23,7 @@ Once Eureka! has been installed, you can learn how to use some aspects of the pa
 
 ## Issues and FAQ
 
-If you experience an issue when installing or using Eureka!, first checkout our
+If you experience an issue when installing or using Eureka!, first check out our
 [FAQ page](https://eurekadocs.readthedocs.io/en/latest/faq.html).
 If the answer to your question is not there, try looking searching for an open or closed issue on 
 [GitHub](https://github.com/kevin218/Eureka/issues?q=is%3Aissue).
