+++ .github/workflows/pytest_testing.yml
@@ -26,5 +26,5 @@ jobs:
           echo "CRDS_SERVER_URL=https://jwst-crds.stsci.edu" >> $GITHUB_ENV
       - name: Run tests
         run: pytest tests --cov=. --cov-report xml
-#      - name: Upload coverage reports to Codecov with GitHub Action
-#        uses: codecov/codecov-action@v2
+     - name: Upload coverage reports to Codecov with GitHub Action
+        uses: codecov/codecov-action@v3
