+++ .github/workflows/pytest_testing.yml
@@ -25,6 +25,10 @@ jobs:
           echo "CRDS_PATH=$HOME/crds_cache" >> $GITHUB_ENV
           echo "CRDS_SERVER_URL=https://jwst-crds.stsci.edu" >> $GITHUB_ENV
       - name: Run tests
-        run: pytest tests --cov=. --cov-report xml
+        run: pytest tests --cov=./ --cov-report=xml
      - name: Upload coverage reports to Codecov with GitHub Action
-        uses: codecov/codecov-action@v3
+       uses: codecov/codecov-action@v2
+       with:
+        env_vars: OS,PYTHON
+        files: ./coverage.xml
+        fail_ci_if_error: true
