### 2022-06-08 5b2ad76
Running pytest on macos and ubuntu
