+++ .github/workflows/pytest_testing.yml
@@ -11,7 +11,13 @@ on:
 jobs:
   test:
     name: pytest
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest]
+    env:
+      OS: ${{ matrix.os }}
+      PYTHON: '3.9'
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v1
@@ -24,7 +30,7 @@ jobs:
           mkdir $HOME/crds_cache
           echo "CRDS_PATH=$HOME/crds_cache" >> $GITHUB_ENV
           echo "CRDS_SERVER_URL=https://jwst-crds.stsci.edu" >> $GITHUB_ENV
-      - name: Run tests
+      - name: Run tests, generate coverage report
         run: pytest tests --cov=./ --cov-report=xml
      - name: Upload coverage reports to Codecov with GitHub Action
        uses: codecov/codecov-action@v2
