+++ .github/workflows/pytest_testing.yml
@@ -1,5 +1,4 @@
 name: Pytest Testing
-
 on:
   pull_request:
     types: [opened,reopened,synchronize]
@@ -7,7 +6,6 @@ on:
     branches:    
       - 'main'
       - 'releases/**'
-
 jobs:
   test:
     name: pytest
