### 2022-06-08 4f72bfe
Fixed "Invalid workflow file"
I had test rather than run under jobs
