### 2022-06-08 8376880
Removing slower macos pytest
Don't realistically need to test on ubuntu and macos, so removing slower
macos test. Also changing the codecov badge on the README to be correct
when merged into main.
