+++ .github/workflows/pytest_testing.yml
@@ -9,12 +9,9 @@ on:
 jobs:
   test:
     name: pytest
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest, macos-latest]
+    runs-on: 'ubuntu-latest'
     env:
-      OS: ${{ matrix.os }}
+      OS: 'ubuntu-latest'
       PYTHON: '3.9'
     steps:
       - uses: actions/checkout@v2
