### 2022-06-09 e8e1738
Restoring conda installation method
