### 2022-06-09 ce8655b
all the fixes requested
