### 2022-06-09 3e25707
these should be all the NIRISS specific doc strings
