+++ docs/source/ecf.rst
@@ -14,13 +14,13 @@ Stage 1
 .. include:: ../media/S1_template.ecf
    :literal:
 
-suffix 
+suffix
 ''''''
 Data file suffix (e.g. uncal).
 
 ramp_fit_algorithm
 ''''''''''''''''''
-Algorithm to use to fit a ramp to the frame-level images of uncalibrated files. Only default (i.e. the JWST pipeline) and mean can be used currently. 
+Algorithm to use to fit a ramp to the frame-level images of uncalibrated files. Only default (i.e. the JWST pipeline) and mean can be used currently.
 
 
 ramp_fit_max_cores
@@ -30,7 +30,7 @@ Fraction of processor cores to use to compute the ramp fits, options are ``none`
 
 skip_*
 ''''''
-If True, skip the named step. 
+If True, skip the named step.
 
 .. note::
    Note that some instruments and observing modes might skip a step either way! See `here <https://jwst-pipeline.readthedocs.io/en/latest/jwst/pipeline/calwebb_detector1.html>`__ for the list of steps run for each instrument/mode by the STScI's JWST pipeline.
@@ -54,16 +54,16 @@ Define the method by which individual frame pixels will be weighted during the d
 
 ``default``: Slope estimation using a least-squares algorithm with an "optimal" weighting, see `here <https://jwst-pipeline.readthedocs.io/en/latest/jwst/ramp_fitting/description.html#optimal-weighting-algorithm>`__.
 
-In short this weights each pixel, :math:`i`, within a slope following :math:`w_i = (i - i_{midpoint})^P`, where the exponent :math:`P` is selected depending on the estimated signal-to-noise ratio of each pixel (see link above). 
+In short this weights each pixel, :math:`i`, within a slope following :math:`w_i = (i - i_{midpoint})^P`, where the exponent :math:`P` is selected depending on the estimated signal-to-noise ratio of each pixel (see link above).
 
 
 ``fixed``: As with default, except the weighting exponent :math:`P` is fixed to a precise value through the ``default_ramp_fit_fixed_exponent`` entry
 
 
-``interpolated``: As with default, except the SNR to :math:`P` lookup table is converted to a smooth interpolation. 
+``interpolated``: As with default, except the SNR to :math:`P` lookup table is converted to a smooth interpolation.
 
 
-``flat``: As with default, except the weighting equation is no longer used, and all pixels are weighted equally. 
+``flat``: As with default, except the weighting equation is no longer used, and all pixels are weighted equally.
 
 
 ``custom``: As with default, except a custom SNR to :math:`P` lookup table can be defined through the ``default_ramp_fit_custom_snr_bounds`` and ``default_ramp_fit_custom_exponents`` (see example .ecf file).
@@ -236,7 +236,6 @@ Possible values:
 5. If MAD of the greatest background outlier is greater than 5, remove this background pixel from the background value calculation. Repeat from Step 2. and repeat as long as there is no 5*MAD outlier in the background column.
 6. Calculate the flux of the polynomial of degree  ``bg_deg`` (calculated in Step 2) at the spectrum and subtract it.
 
-
 p3thresh
 ''''''''
 Only important if ``bg_deg => 0`` (see above). # sigma threshold for outlier rejection during background subtraction which corresponds to step 3 of optimal spectral extraction, as defined by Horne (1986).
@@ -294,10 +293,56 @@ The path to the directory in which to output the Stage 3 JWST data and plots.
 
 
 
+Stage 3 - NIRISS
+----------------
 
+trace_method
+''''''''''''
[…]
