### 2022-06-09 028b903
Fixing double unit conversion for WCF3 e-/s files
