### 2022-06-09 6d6ea93
more docstrings in background.py
