+++ eureka/lib/simultaneous_order_fitting.py
@@ -1,11 +1,9 @@
 import numpy as np
 import scipy.optimize as so
-
-
 import pyximport
+import itertools
 pyximport.install()
 from ..S3_data_reduction import niriss_extraction as profiles
-
 
 __all__ = ['fit_orders_fast', 'fit_orders']
 
@@ -13,8 +11,8 @@ __all__ = ['fit_orders_fast', 'fit_orders']
 def fit_orders(data, tab):
     """
     Creates a 2D image optimized to fit the data. Currently
-    runs with a Gaussian profile, but will look into other 
-    more realistic profiles at some point. This routine  
+    runs with a Gaussian profile, but will look into other
+    more realistic profiles at some point. This routine
     is a bit slow, but fortunately, you only need to run it
     once per observations.
 
@@ -23,7 +21,7 @@ def fit_orders(data, tab):
     data : object
     which_table : int, optional
        Sets with table of initial y-positions for the
-       orders to use. Default is 2.  
+       orders to use. Default is 2.
 
     Returns
     -------
@@ -33,60 +31,64 @@ def fit_orders(data, tab):
     print("Go grab some food. This routing could take up to 30 minutes.")
 
     def construct_guesses(A, B, sig, length=10):
-        As   = np.linspace(A[0],   A[1],   length)  # amplitude of gaussian for first order
-        Bs   = np.linspace(B[0],   B[1],   length)  # amplitude of gaussian for second order
-        sigs = np.linspace(sig[0], sig[1], length)  # std of gaussian profile  
-        combos = np.array(list(itertools.product(*[As,Bs,sigs]))) # generates all possible combos
+        # amplitude of gaussian for first order
+        As = np.linspace(A[0], A[1], length)
+        # amplitude of gaussian for second order
+        Bs = np.linspace(B[0], B[1], length)
+        # std of gaussian profile
+        sigs = np.linspace(sig[0], sig[1], length)
+        # generates all possible combos
+        combos = np.array(list(itertools.product(*[As, Bs, sigs])))
         return combos
 
     pos1, pos2 = tab['order_1'], tab['order_2']
 
-    # Good initial guesses 
-    combos = construct_guesses([0.1,30], [0.1,30], [1,40])
+    # Good initial guesses
+    combos = construct_guesses([0.1, 30], [0.1, 30], [1, 40])
 
     # generates length x length x length number of images and fits to the data
     img1, sigout1 = profiles.build_gaussian_images(data.median,
-                                                   combos[:,0], combos[:,1],
-                                                   combos[:,2],
+                                                   combos[:, 0], combos[:, 1],
+                                                   combos[:, 2],
                                                    pos1, pos2)
-    
+
     # Iterates on a smaller region around the best guess
     best_guess = combos[np.argmin(sigout1)]
-    combos = construct_guesses( [best_guess[0]-0.5, best_guess[0]+0.5],
-                                [best_guess[1]-0.5, best_guess[1]+0.5],
-                                [best_guess[2]-0.5, best_guess[2]+0.5] )
+    combos = construct_guesses([best_guess[0]-0.5, best_guess[0]+0.5],
+                               [best_guess[1]-0.5, best_guess[1]+0.5],
+                               [best_guess[2]-0.5, best_guess[2]+0.5])
 
-    # generates length x length x length number of images centered around the previous
+    # generates length x length x length number of
+    #   images centered around the previous
     #   guess to optimize the image fit
     img2, sigout2 = profiles.build_gaussian_images(data.median,
-                                                   combos[:,0], combos[:,1],
-                                                   combos[:,2],
+                                                   combos[:, 0], combos[:, 1],
+                                                   combos[:, 2],
                                                    pos1, pos2)
 
[…]
