### 2022-06-10 c6bf387
Allowing S3 to batch files together
Working on resolving #283

Currently just testing this on WFC3, but will work on the other
instruments once WFC3 is working.

Eventually this will allow for different reference frames for WFC3
rather than just the first frame of each scan direction. It should also
speed up Stage 3 with parallel processing while using more RAM.
