+++ eureka/S3_data_reduction/wfc3.py
@@ -33,22 +33,20 @@ def preparation_step(meta, log):
     """
     meta.gain = 1
 
-    obstimes, CRPIX1, CRPIX2, postarg1, postarg2, ny, meta, log = \
-        separate_direct(meta, log)
-    meta, log = separate_scan_direction(obstimes, postarg2, meta, log)
+    meta, log = separate_direct(meta, log)
+    meta, log = separate_scan_direction(meta, log)
 
     # Calculate centroid of direct image(s)
     meta.centroid = hst.imageCentroid(meta.direct_list, meta.centroidguess,
-                                      meta.centroidtrim, ny, CRPIX1, CRPIX2,
-                                      postarg1, postarg2)
+                                      meta.centroidtrim, meta.ny, meta.CRPIX1,
+                                      meta.CRPIX2, meta.postarg1,
+                                      meta.postarg2)
 
     # Initialize listto hold centroid positions from later steps in this stage
     meta.centroids = []
     meta.subflat = []
     meta.flatmask = []
     meta.scanHeight = []
-    meta.diffmask = []
-    meta.subdiffmask = []
     meta.drift2D = []
     meta.drift2D_int = []
     meta.subdata_ref = []
@@ -78,8 +76,6 @@ def conclusion_step(meta, log):
     meta.subflat = np.array(meta.subflat)
     meta.flatmask = np.array(meta.flatmask)
     meta.scanHeight = np.array(meta.scanHeight)
-    meta.diffmask = np.array(meta.diffmask)
-    meta.subdiffmask = np.array(meta.subdiffmask)
     meta.drift2D = np.array(meta.drift2D)
     meta.drift2D_int = np.array(meta.drift2D_int)
     meta.subdata_ref = np.array(meta.subdata_ref)
@@ -205,10 +201,17 @@ def separate_direct(meta, log):
             meta.direct_index[i] = \
                 np.where(science_times[i] > direct_times)[0][-1]
 
-    return obstimes, CRPIX1, CRPIX2, postarg1, postarg2, ny, meta, log
+    meta.obstimes = obstimes
+    meta.CRPIX1 = CRPIX1
+    meta.CRPIX2 = CRPIX2
+    meta.postarg1 = postarg1
+    meta.postarg2 = postarg2
+    meta.ny = ny
+
+    return meta, log
 
 
-def separate_scan_direction(obstimes, postarg2, meta, log):
+def separate_scan_direction(meta, log):
     """Separate alternating scan directions.
 
     Parameters
@@ -239,12 +242,12 @@ def separate_scan_direction(obstimes, postarg2, meta, log):
         meta.scandir = np.zeros(meta.num_data_files, dtype=int)
         meta.n_scan0 = 0
         meta.n_scan1 = 0
-        scan0 = postarg2[0]
-        scan1 = postarg2[1]
+        scan0 = meta.postarg2[0]
+        scan1 = meta.postarg2[1]
         for m in range(meta.num_data_files):
-            if postarg2[m] == scan0:
+            if meta.postarg2[m] == scan0:
                 meta.n_scan0 += 1
-            elif postarg2[m] == scan1:
+            elif meta.postarg2[m] == scan1:
                 meta.scandir[m] = 1
                 meta.n_scan1 += 1
             else:
@@ -256,7 +259,18 @@ def separate_scan_direction(obstimes, postarg2, meta, log):
                  mute=(not meta.verbose))
 
     # Group frames into frame, batch, and orbit number
-    meta.framenum, meta.batchnum, meta.orbitnum = hst.groupFrames(obstimes)
+    meta.framenum, meta.batchnum, meta.orbitnum = \
+        hst.groupFrames(meta.obstimes)
+
[…]
