### 2022-06-10 6f786ac
photutils documentation links
