### 2022-06-10 b3152fa
more doc strings for niriss_class
