### 2022-06-10 d35cc81
meta & data doc string changes
