### 2022-06-10 f709901
rmv debugging plot
