+++ eureka/S3_data_reduction/niriss.py
@@ -176,9 +176,7 @@ def fit_bg(data, meta, log,
         box_mask = dirty_mask(data.medflux.values,
                               meta.trace_edge, booltype=True,
                               return_together=True)
-    import matplotlib.pyplot as plt
-    plt.imshow(box_mask)
-    plt.show()
+
     bkg, bkg_var, cr_mask = fitbg3(data, np.array(box_mask-1, dtype=bool),
                                    readnoise, sigclip, bkg_estimator=bkg_estimator,
                                    box=box, filter_size=filter_size,
