### 2022-06-10 496d18c
hasattr meta.trace_method
