+++ eureka/S3_data_reduction/background.py
@@ -586,4 +586,4 @@ def fitbg3(data, order_mask, readnoise=11,
 
         bkg_var[i] = np.nansum(v, axis=0)
 
-    return np.abs(bkg), bkg_var, rm_crs
+    return bkg, bkg_var, rm_crs
