### 2022-06-10 c291f20
interpolates & fills-in poor DQ pixels
