### 2022-06-10 b56fcbd
Fixing off-by-one error in batch files
