### 2022-06-10 c2f4d8a
Fixing meta.nreads for WFC3 with new batching
