### 2022-06-10 e3c16c5
WFC3 data['mask'] now same as meta.diffmask
