### 2022-06-11 d7ef145
Prints to logs, adding meta.use_estsig for flag_bg
Also saving WFC3 mask instead of reinitializing in s3_resduce.py.
Temporarily adding back meta.diffmask and meta.subdiffmask. Fixing
meta.nreads after differencing frames.

WFC3 files now work in batched or unbatched mode, but they need
use_estsig=True for some reason. Likely because of the change in the
masking. Without use_estsig, I get an error on line 150 of optspex.py
saying can't compute argmax of an empty slice

Also, ncpu>1 seems very slow for batched WFC3 files at present
