+++ eureka/S3_data_reduction/nirspec.py
@@ -6,7 +6,7 @@ from . import nircam, sigrej
 from ..lib.util import read_time
 
 
-def read(filename, data, meta):
+def read(filename, data, meta, log):
     '''Reads single FITS file from JWST's NIRCam instrument.
 
     Parameters
@@ -17,11 +17,17 @@ def read(filename, data, meta):
         The Dataset object in which the fits data will stored.
     meta : eureka.lib.readECF.MetaClass
         The metadata object.
+    log : logedit.Logedit
+        The current log.
 
     Returns
     -------
     data : Xarray Dataset
         The updated Dataset object with the fits data stored inside.
+    meta : eureka.lib.readECF.MetaClass
+        The metadata object
+    log : logedit.Logedit
+        The current log.
 
     Notes
     -----
@@ -89,7 +95,7 @@ def read(filename, data, meta):
     data['wave_2d'] = (['y', 'x'], wave_2d)
     data['wave_2d'].attrs['wave_units'] = wave_units
 
-    return data, meta
+    return data, meta, log
 
 
 def flag_bg(data, meta):
