### 2022-06-11 03dbb14
Fixing log formatting
