### 2022-06-11 3ec54c3
Fixing more log formatting
