### 2022-06-12 #4453 @danielskatz
@christinahedges - We can let the current @openjournals/joss-eics decide ...
