### 2022-06-13 c3b4eb6
add jump rejection threshold to S1
