### 2022-06-13 e6e5f83
add default when not a float/not defined
