### 2022-06-13 #4453 @taylorbell57
During our Eureka team meeting today, we were wondering if it would be possible to get a more rapid review of our paper. With the first JWST Early Release Science data becoming public on July 12, our hope is to have a paper that users of our pipeline can cite when publishing those first observations so that our work gets properly recognized. It seems unfortunate to us to have most of the time between now and then lost while on a waitlist
