### 2022-06-13 c424b0f
Fixing fractional masking for WFC3
