### 2022-06-14 a31a4c3
data.mask should be integers
