### 2022-06-14 #4453 @arfon
Hi @taylorbell57 – I totally understand the time pressure here but I also want to gently point out that you're asking a volunteer editorial team to do more work. We've waitlisted this for a good reason (limited capacity) and will get to this paper when we can.

Also, as a workaround (to provide something citable before the JOSS process is complete), you could post a pre-print to the arXiv which will give the community something to cite in the interim, and ADS should connect the dots between the pre-print and the JOSS paper (see [this example](https://ui.adsabs.harvard.edu/abs/2021JOSS....6.3285F/abstract)).
