### 2022-06-14 bee1797
Fix WFC3 masking, allow time-axis clipping
