+++ eureka/S3_data_reduction/wfc3.py
@@ -6,7 +6,7 @@ from astropy.io import fits
 import scipy.interpolate as spi
 import scipy.ndimage as spni
 import astraeus.xarrayIO as xrio
-from . import nircam
+from . import nircam, sigrej
 from . import hst_scan as hst
 from ..lib import suntimecorr, utc_tt
 
@@ -47,10 +47,6 @@ def preparation_step(meta, log):
     meta.subflat = []
     meta.flatmask = []
     meta.scanHeight = []
-    meta.diffmask = []
-    meta.subdiffmask = []
-    meta.drift2D = []
-    meta.drift2D_int = []
     meta.subdata_ref = []
     meta.subdiffmask_ref = []
 
@@ -78,10 +74,6 @@ def conclusion_step(meta, log):
     meta.subflat = np.array(meta.subflat)
     meta.flatmask = np.array(meta.flatmask)
     meta.scanHeight = np.array(meta.scanHeight)
-    meta.diffmask = np.array(meta.diffmask)
-    meta.subdiffmask = np.array(meta.subdiffmask)
-    meta.drift2D = np.array(meta.drift2D)
-    meta.drift2D_int = np.array(meta.drift2D_int)
     meta.subdata_ref = np.array(meta.subdata_ref)
     meta.subdiffmask_ref = np.array(meta.subdiffmask_ref)
 
@@ -330,18 +322,16 @@ def read(filename, data, meta, log):
         # Determine if we are using IMA or FLT files
         if filename.endswith('flt.fits'):
             # FLT files subtract first from last, 2 reads
-            meta.nreads = 2
+            data.attrs['nreads'] = 2
         else:
-            meta.nreads = data.attrs['shdr']['SAMPNUM']
-        
-        data.attrs['nreads'] = meta.nreads
+            data.attrs['nreads'] = data.attrs['shdr']['SAMPNUM']
 
-        sci = np.zeros((meta.nreads, meta.ny, meta.nx))  # Flux
-        err = np.zeros((meta.nreads, meta.ny, meta.nx))  # Error
-        dq = np.zeros((meta.nreads, meta.ny, meta.nx))  # Flags
+        sci = np.zeros((data.attrs['nreads'], meta.ny, meta.nx))  # Flux
+        err = np.zeros((data.attrs['nreads'], meta.ny, meta.nx))  # Error
+        dq = np.zeros((data.attrs['nreads'], meta.ny, meta.nx))  # Flags
         jd = []
         j = 0
-        for rd in range(meta.nreads, 0, -1):
+        for rd in range(data.attrs['nreads'], 0, -1):
             sci[j] = hdulist['SCI', rd].data
             err[j] = hdulist['ERR', rd].data
             dq[j] = hdulist['DQ', rd].data
@@ -382,7 +372,7 @@ def read(filename, data, meta, log):
                                      name='dq')
 
     # Calculate centroids for each frame
-    centroids = np.zeros((meta.nreads, 2))
+    centroids = np.zeros((data.attrs['nreads'], 2))
     # Figure out which direct image is the relevant one for this observation
     image_number = np.where(meta.segment_list == filename)[0][0]
     centroid_index = meta.direct_index[image_number]
@@ -432,8 +422,8 @@ def read(filename, data, meta, log):
     diffdata['wave_2d'].attrs['wave_units'] = wave_units
 
     # Figure out which read this file starts and ends with
-    diffdata.attrs['intstart'] = image_number*(meta.nreads-1)
-    diffdata.attrs['intend'] = (image_number+1)*(meta.nreads-1)
+    diffdata.attrs['intstart'] = image_number*(data.attrs['nreads']-1)
[…]
