+++ eureka/S3_data_reduction/hst_scan.py
@@ -797,7 +797,7 @@ def replacePixels(shiftdata, shiftmask, m, n, i, j, k, ktot, ny, nx, sy, sx):
     return shift, m, n, i, j
 
 
-def drift_fit2D(meta, data, validRange=9):
+def drift_fit2D(meta, data, nreads, validRange=9):
     '''Measures the spectrum drift over all frames and all non-destructive reads.
 
     Parameters
@@ -806,6 +806,8 @@ def drift_fit2D(meta, data, validRange=9):
         Event object.
     data : ndarray
         4D data frames.
+    nreads : int
+        The number of reads per file.
     validRange : int
         Trim spectra by +/- pixels to compute valid region of
         cross correlation.
@@ -824,14 +826,14 @@ def drift_fit2D(meta, data, validRange=9):
     '''
     # if postclip is not None:
     #    postclip = -postclip
-    if meta.nreads > 2:
+    if nreads > 2:
         istart = 1
     else:
         istart = 0
-    drift = np.zeros((meta.num_data_files, meta.nreads-1))
-    # model = np.zeros((ev.n_files, ev.n_reads-1))
-    # goodmask = np.zeros((ev.n_files, ev.n_reads-1),dtype=int)
-    for n in range(istart, meta.nreads-1):
+    drift = np.zeros((meta.num_data_files, nreads-1))
+    # model = np.zeros((meta.num_data_files, n_reads-1))
+    # goodmask = np.zeros((meta.num_data_files, n_reads-1), dtype=int)
+    for n in range(istart, nreads-1):
         ref_data = np.copy(data[-1, n])
         ref_data[np.where(np.isnan(ref_data))] = 0
         for m in range(meta.num_data_files):
@@ -861,10 +863,10 @@ def drift_fit2D(meta, data, validRange=9):
                 subvals = vals[argmax-width:argmax+width+1]
                 params, err = g.fitgaussian(subvals/subvals.max(),
                                             guess=[width/5., width*1., 1])
-                drift[n,m,i]= len(vals)/2 - params[1] - argmax + width
-                goodmask[n,m,i] = 1
+                drift[n, m, i]= len(vals)/2 - params[1] - argmax + width
+                goodmask[n, m, i] = 1
             except:
-                print('Spectrum ' +str(n)+','+str(m)+','+str(i)
+                print('Spectrum '+str(n)+','+str(m)+','+str(i)
                       ' marked as bad.')
             '''
 
