### 2022-06-14 9ebacc7
Updaing S3 ECFs to allow batching
Changing to zero counting for xarray save file names. Also, for some
reason it seems that the FluxData save file is larger for batched data
than the sum of the individual files.

Tested on full datasets for WFC3 and MIRI
