### 2022-06-15 3946ac8
Fixing NIRISS flake8 issues
