+++ eureka/niriss_class.py
@@ -1,19 +1,13 @@
 import os
 import numpy as np
-from astropy import units
 from astropy.io import fits
-import matplotlib.pyplot as plt
-from astropy.table import Table
 
 from .lib.tracing_niriss import mask_method_edges, mask_method_ears, ref_file
-from .lib.masking        import (interpolating_row, data_quality_mask,
-                              interpolating_image)
-from .lib.clipping       import time_removal
-from .S3_data_reduction.background     import bkg_sub, fitbg3
-from .S3_data_reduction.niriss_extraction   import (dirty_mask, box_extract,
-                                              optimal_extraction_routine)
-#from .S3_data_reduction.niriss  import wave_NIRISS as wavelength
-from .lib.simultaneous_order_fitting import fit_orders, fit_orders_fast
+from .lib.masking import data_quality_mask, interpolating_image
+from .S3_data_reduction.background import fitbg3
+from .S3_data_reduction.niriss_extraction import (dirty_mask, box_extract,
+                                                  optimal_extraction_routine)
+# from .S3_data_reduction.niriss import wave_NIRISS as wavelength
 
 
 __all__ = ['NIRISS_S3']
@@ -23,8 +17,7 @@ class NIRISS_S3(object):
 
     def __init__(self, filename, f277_filename=None,
                  data_dir=None, output_dir=None):
-        """
-        Initializes the NIRISS S3 data reduction class.
+        """Initializes the NIRISS S3 data reduction class.
 
         Parameters
         ----------
@@ -90,25 +83,24 @@ class NIRISS_S3(object):
             self.setup_f277(f277_filename)
         else:
             self.f277 = None
-            print('Without F277W filter image, some functions may not be available.')
+            print('Without F277W filter image, some functions may not be '
+                  'available.')
 
-        self.trace_ear     = None
-        self.trace_edge     = None
-        self.bkg      = None
+        self.trace_ear = None
+        self.trace_edge = None
+        self.bkg = None
         self.box_mask = None
         self.box_var1 = None
         self.box_var2 = None
-        self.bkg_removed  = None
+        self.bkg_removed = None
         self.box_spectra1 = None
         self.box_spectra2 = None
         self.box_mask_separate = None
 
         return
 
-
     def setup(self):
-        """
-        Sets up all proper attributes from the FITS file.
+        """Sets up all proper attributes from the FITS file.
 
         Attributes
         ----------
@@ -143,47 +135,42 @@ class NIRISS_S3(object):
         median : np.ndarray
            Median frame of all science images.
         """
-        hdu = fits.open(os.path.join(self.data_dir, self.filename))
+        with fits.open(os.path.join(self.data_dir, self.filename)) as hdu:
+            self.mhdr = hdu[0].header        # Sets in the meta data header
+            self.shdr = hdu['SCI', 1].header  # Sets the science data header
 
-        self.mhdr = hdu[0].header        # Sets in the meta data header
-        self.shdr = hdu['SCI',1].header  # Sets the science data header
+            self.intend = np.copy(hdu[0].header['NINTS'])
+            self.time = np.linspace(self.mhdr['EXPSTART'],
+                                    self.mhdr['EXPEND'],
+                                    int(self.intend))
 
-        self.intend = np.copy(hdu[0].header['NINTS'])
-        self.time   = np.linspace(self.mhdr['EXPSTART'],
-                                  self.mhdr['EXPEND'],
-                                  int(self.intend) )
+            self.time_units = 'BJD_TDB'
+            self.inttime = hdu[0].header['EFFINTTM']
 
-        self.time_units = 'BJD_TDB'
-        self.inttime  = hdu[0].header['EFFINTTM']
[…]
