+++ eureka/S3_data_reduction/s3_reduce_niriss.py
@@ -30,8 +30,7 @@ from tqdm import tqdm
 
 from . import niriss_extraction
 from .niriss_extraction import optimal_extraction_routine
-from . import plots_s3, source_pos
-from . import background as bg
+from . import plots_s3
 from . import bright2flux as b2f
 from . import niriss as inst
 
@@ -39,7 +38,7 @@ from ..lib import logedit
 from ..lib import readECF
 from ..lib import manageevent as me
 from ..lib import util
-from ..lib import tracing_niriss
+from ..lib.masking import interpolating_image
 
 
 def reduce(eventlabel, ecf_path=None, s2_meta=None):
@@ -108,12 +107,11 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
     meta.x1d_segment_list = x1d_files
 
     # Get the NIRISS traces
-    meta = inst.define_traces(meta)
+    meta = inst.define_traces(meta, log)
 
-    ## TO DO : RECORD THE TRACES IN THE DATA OBJECT
+    # TO DO : RECORD THE TRACES IN THE DATA OBJECT
     # want to record the trace in the data object via Astreaus
     # make flux like data command
-
 
     # create directories to store data
     # run_s3 used to make sure we're always looking at the right run for
@@ -148,7 +146,6 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
         log.writelog(f'\nFound {meta.num_data_files} data file(s) '
                      f'ending in {meta.suffix}.fits',
                      mute=(not meta.verbose))
-
 
     datasets = []
     # Loop over each segment
@@ -194,8 +191,7 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
                 soln = hdulist[ext].data['WAVELENGTH'] + 0.0
                 wave_soln[e, :len(soln)] = soln
 
-        data['wave_1d'] = (['order', 'x'],
-                            wave_soln)
+        data['wave_1d'] = (['order', 'x'], wave_soln)
         data['wave_1d'].attrs['wave_units'] = 'micron'
 
         # Convert flux units to electrons
@@ -210,11 +206,11 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
 
         # Interpolating over bad pixels from the data quality map
         data['data'] = interpolating_image(data['data'], mask=data['DQ'])
-        data['err']  = interpolating_image(data['err'],  mask=data['DQ'])
-        data['v0']  = interpolating_image(data['v0'],  mask=data['DQ'])
-        data['medflux'] = interpolating_image(np.nanmedian(data['data'], axis=0),
-                                              mask=np.nanmedian(mask=data['DQ'],
-                                                                axis=0))
+        data['err'] = interpolating_image(data['err'], mask=data['DQ'])
+        data['v0'] = interpolating_image(data['v0'], mask=data['DQ'])
+        data['medflux'] = interpolating_image(
+            np.nanmedian(data['data'], axis=0),
+            mask=np.nanmedian(mask=data['DQ'], axis=0))
 
         # Create bad pixel mask (1 = good, 0 = bad)
         # FINDME: Will want to use DQ array in the future
@@ -242,10 +238,8 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
         log.writelog('  Performing background outlier rejection',
                      mute=(not meta.verbose))
 
-        data = inst.fit_bg(data, meta, log,
-                           readnoise=meta.readnoise,
-                           sigclip=meta.sigclip,
-                           box=meta.box,
+        data = inst.fit_bg(data, meta, readnoise=meta.readnoise,
+                           sigclip=meta.sigclip, box=meta.box,
                            filter_size=meta.filter_size,
                            bkg_estimator=meta.bkg_estimator,
                            testing=meta.testing_S3,
[…]
