+++ eureka/S4_generate_lightcurves/s4_genLC_niriss.py
@@ -77,7 +77,6 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
     meta = me.mergeevents(meta, s3_meta)
 
     # Create directories for Stage 5 outputs
-    meta.run_s4 = None
     meta.run_s4 = util.makedirectory(meta, 'S4', meta.run_s4,
                                      order=meta.order)
 
@@ -243,14 +242,11 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
     # Need to look into these issues.
     optspec_ma = np.ma.masked_invalid(optspec_ma)
     opterr_ma = np.ma.masked_array(opterr_ma, optspec_ma.mask)
-    # spec['optspec_drift']
 
     # Compute MAD alue
     meta.mad_s4 = util.get_mad(meta, spec.wave_1d.values, optspec_ma,
                                meta.wave_min, meta.wave_max)
     log.writelog(f"Stage 4 MAD = {str(np.round(meta.mad_s4, 2))} ppm")
-
-    print(meta.outputdir)
 
     if meta.isplots_S4 >= 1:
         plots_s4.lc_driftcorr(meta, spec.wave_1d[index][nonans], optspec_ma)
