+++ eureka/S3_data_reduction/niriss.py
@@ -11,9 +11,9 @@ __all__ = ['read', 'define_traces',
            'fit_bg']
 
 
-def read(filename, data, meta, f277_filename=None):
-    """
-    Reads a single FITS file from JWST's NIRISS instrument.
+def read(filename, data, meta):
+    """Reads a single FITS file from JWST's NIRISS instrument.
+    
     This takes in the Stage 2 processed files.
 
     Parameters
@@ -24,8 +24,6 @@ def read(filename, data, meta, f277_filename=None):
         The Dataset object in which the fits data will stored.
     meta : eureka.lib.readECF.MetaClass
         The metadata object.
-    f277_filename : str, optional
-       Single filename for the F277W filtered images.
 
     Returns
     -------
@@ -34,52 +32,51 @@ def read(filename, data, meta, f277_filename=None):
     meta : eureka.lib.readECF.MetaClass
        Metadata stored in the FITS file.
     """
-    hdulist = fits.open(filename)
+    with fits.open(filename) as hdulist:
+        # Load master and science headers
+        data.attrs['filename'] = filename
+        data.attrs['mhdr'] = hdulist[0].header
+        data.attrs['shdr'] = hdulist['SCI', 1].header
+        data.attrs['NINTS'] = data.attrs['mhdr']['NINTS']
 
-    # Load master and science headers
-    data.attrs['filename'] = filename
-    data.attrs['mhdr'] = hdulist[0].header
-    data.attrs['shdr'] = hdulist['SCI', 1].header
-    data.attrs['NINTS'] = data.attrs['mhdr']['NINTS']
+        # need some placeholder right now for testing
+        data.attrs['intstart'] = 0
+        data.attrs['intend'] = 3
 
-    # need some placeholder right now for testing
-    data.attrs['intstart'] = 0
-    data.attrs['intend'] = 3
+        if hasattr(meta, 'f277_filename') and meta.f277_filename is not None:
+            with fits.open(meta.f277_filename) as f277:
+                data.attrs['f277'] = f277[1].data
 
-    if f277_filename is not None:
-        with fits.open(f277_filename) as f277:
-            data.attrs['f277'] = f277[1].data
+        # Load data
+        sci = hdulist['SCI', 1].data
+        err = hdulist['ERR', 1].data
+        dq = hdulist['DQ', 1].data
+        v0 = hdulist['VAR_RNOISE', 1].data
+        # var  = hdulist['VAR_POISSON',1].data
+        wave_2d = hdulist['WAVELENGTH', 1].data
+        # int_times = hdulist['INT_TIMES', 1].data[data.attrs['intstart']-1:
+        #                                          data.attrs['intend']]
 
-    # Load data
-    sci = hdulist['SCI', 1].data
-    err = hdulist['ERR', 1].data
-    dq = hdulist['DQ', 1].data
-    v0 = hdulist['VAR_RNOISE', 1].data
-    # var  = hdulist['VAR_POISSON',1].data
-    wave_2d = hdulist['WAVELENGTH', 1].data
-    # int_times = hdulist['INT_TIMES', 1].data[data.attrs['intstart']-1:
-    #                                          data.attrs['intend']]
+        # Record integration mid-times in BJD_TDB
+        try:
+            int_times = hdulist['INT_TIMES', 1].data
+            time = int_times['int_mid_BJD_TDB']
+            time_units = 'BJD_TDB'
+        except:
+            # This exception is (hopefully) only for simulated data
+            print("WARNING: INT_TIMES not found."
+                  "Using EXPSTART and EXPEND in UTC.")
+            time = np.linspace(data.attrs['mhdr']['EXPSTART'],
+                               data.attrs['mhdr']['EXPEND'],
+                               int(data.attrs['NINTS']))
 
[…]
