+++ eureka/S3_data_reduction/s3_reduce_niriss.py
@@ -76,6 +76,7 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
     meta = readECF.MetaClass(ecf_path, ecffile)
     meta.eventlabel = eventlabel
     meta.inst = 'niriss'
+    meta.datetime = time_pkg.strftime('%Y-%m-%d')
 
     if s2_meta is None:
         # Locate the old MetaClass savefile, and load new ECF into
