### 2022-06-15 2dd0777
Allowing arbitrary reference frames for WFC3
