### 2022-06-16 93abc40
Changing MIRI print statements to logs
