### 2022-06-16 5cfda04
Modularizing code from s3_reduce.py
s3_reduce.py has a lot of code in it, but ideally it would pretty well
only be calling other functions for readability. This commit gets us
closer to that.
