### 2022-06-17 596e789
Clarifying S3 starting file log statements
