### 2022-06-18 f9e70d2
S3 now works well for WFC3 data (if batched)
Code has been tested against all instruments successfully. WFC3 code
currently does not work correctly without batching.

Still need to update documentation.
