+++ environment.yml
@@ -1,0 +1,42 @@
+name: eureka
+channels:
+    - conda-forge
+    - defaults
+dependencies:
+    - python==3.9.7
+    - asdf[version='>=2.7.1,<2.11.0'] # Needed for jwst==1.3.3
+    - astropy
+    - batman-package
+    - bokeh
+    - ccdproc
+    - celerite
+    - corner
+    - cython
+    - dynesty[version='>1.0']
+    - emcee[version='>3.0.0']
+    - george
+    - gwcs
+    - h5py[version='<3.2']
+    - ipython
+    - lmfit
+    - matplotlib
+    - nbsphinx # Needed for documentation
+    - numpy[version='>=1.20.0']
+    - pandas
+    - photutils
+    - pip # required for the pip installs below
+    - pytest
+    - recommonmark # Needed for documentation
+    - requests
+    - scipy[version='>=1.4.0'] # Needed for scipy.fft
+    - tqdm
+    - pip:
+        - astraeus @ git+https://github.com/kevin218/Astraeus@main#egg=astraeus
+        - crds
+        - image_registration @ git+https://github.com/keflavich/image_registration.git@master#egg=image_registration
+        - jwst==1.3.3
+        - myst-parser # Needed for documentation
+        - sphinx-rtd-theme # Needed for documentation
+        - stcal
+        - stdatamodels
+        - svo_filters
