### 2022-06-20 c168328
updated installation.rst in docs
