### 2022-06-20 8574569
added -e to pip install instruction
