+++ docs/source/installation.rst
@@ -64,13 +64,13 @@ Once in your new conda environment, you can install the ``Eureka!`` package with
 
 .. code-block:: bash
 
-	pip install git+https://github.com/kevin218/Eureka.git#egg=eureka[jwst]
+	pip install -e git+https://github.com/kevin218/Eureka.git#egg=eureka[jwst]
 
 where specific branches can be installed using:
 
 .. code-block:: bash
 	
-	pip install git+https://github.com/kevin218/Eureka.git@mybranchname#egg=eureka[jwst]
+	pip install -e git+https://github.com/kevin218/Eureka.git@mybranchname#egg=eureka[jwst]
 
 If you desire any of the files in the `demos folder <https://github.com/kevin218/Eureka/tree/main/demos>`_, you will have to download these from
 GitHub following the method described below.
