### 2022-06-20 3c96b74
Fix missing traces variable from s3_reduce_niriss
traces currently gets saved in meta.trace_edge
