### 2022-06-20 b267617
Fixing issues with negating arrays
