### 2022-06-20 294e636
Fixing issue with empty NIRISS time array
