### 2022-06-20 26069c6
Making Stage3 folder before opening new log
