### 2022-06-20 6592ac0
data['data'] -> data['flux']
