+++ eureka/S3_data_reduction/s3_reduce_niriss.py
@@ -205,11 +205,11 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
             data.flux.attrs['flux_units']
 
         # Interpolating over bad pixels from the data quality map
-        data['data'] = interpolating_image(data['data'], mask=data['DQ'])
+        data['flux'] = interpolating_image(data['flux'], mask=data['DQ'])
         data['err'] = interpolating_image(data['err'], mask=data['DQ'])
         data['v0'] = interpolating_image(data['v0'], mask=data['DQ'])
         data['medflux'] = interpolating_image(
-            np.nanmedian(data['data'], axis=0),
+            np.nanmedian(data['flux'], axis=0),
             mask=np.nanmedian(mask=data['DQ'], axis=0))
 
         # Create bad pixel mask (1 = good, 0 = bad)
