### 2022-06-20 d19a85f
Fixing issues with interpolating_image calls
