### 2022-06-20 98672e6
Adding NIRISS test
The NIRISS code is currently broken, but this test can be used for
debugging
