### 2022-06-20 a2f20e1
Removing outdated wavelength call in niriss_class
