+++ eureka/S3_data_reduction/s3_reduce.py
@@ -416,9 +416,9 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
             if meta.isplots_S3 >= 1:
                 log.writelog('Generating figure')
                 # 2D light curve without drift correction
-                plots_s3.lc_nodriftcorr(data, meta, spec.wave_1d, spec.optspec)
+                plots_s3.lc_nodriftcorr(spec, meta, spec.wave_1d, spec.optspec)
                 if meta.inst == 'wfc3':
-                    plots_s3.drift_2D(data, meta)
+                    plots_s3.drift_2D(spec, meta)
 
             # Save results
             if meta.save_output:
