### 2022-06-20 b56da70
Fixing flake8 issues with PyMC3Models.py
