+++ src/eureka/S5_lightcurve_fitting/models/PyMC3Models.py
@@ -45,7 +45,9 @@ class StarryModel(pm.Model):
         required = np.array(['Ms', 'Mp', 'Rs'])
         missing = np.array([name not in self.paramtitles for name in required])
         if np.any(missing):
-            raise AssertionError(f'Missing required params {required[missing]} in your EPF.')
+            message = (f'Missing required params {required[missing]} in your '
+                       f'EPF.')
+            raise AssertionError(message)
 
         with self:
             for parname in self.paramtitles:
@@ -55,31 +57,55 @@ class StarryModel(pm.Model):
                 elif param.ptype == 'fixed':
                     setattr(self, parname, tt.constant(param.value))
                 elif param.ptype not in ['free', 'shared']:
-                    raise ValueError(f'ptype {param.ptype} for parameter {param.name} is not recognized.')
+                    message = (f'ptype {param.ptype} for parameter '
+                               f'{param.name} is not recognized.')
+                    raise ValueError(message)
                 else:
                     for c in range(self.nchan):
-                        if c!=0:
+                        if c != 0:
                             parname_temp = parname+'_'+str(c)
                         else:
                             parname_temp = parname
 
-                        if param.ptype=='free' or c==0:
+                        if param.ptype == 'free' or c == 0:
                             if param.prior == 'U':
-                                setattr(self, parname_temp, pm.Uniform(parname_temp, lower=param.priorpar1, upper=param.priorpar2))
+                                setattr(self, parname_temp,
+                                        pm.Uniform(parname_temp,
+                                                   lower=param.priorpar1,
+                                                   upper=param.priorpar2))
                             elif param.prior == 'N':
-                                if parname in ['rp', 'per', 'ecc', 'scatter_mult', 'scatter_ppm', 'c0']:
-                                    setattr(self, parname_temp, BoundedNormal_0(parname_temp, mu=param.priorpar1, sigma=param.priorpar2))
+                                if parname in ['rp', 'per', 'ecc',
+                                               'scatter_mult', 'scatter_ppm',
+                                               'c0']:
+                                    setattr(self, parname_temp,
+                                            BoundedNormal_0(
+                                                parname_temp,
+                                                mu=param.priorpar1,
+                                                sigma=param.priorpar2))
                                 elif parname in ['inc']:
-                                    setattr(self, parname_temp, BoundedNormal_90(parname_temp, mu=param.priorpar1, sigma=param.priorpar2))
+                                    setattr(self, parname_temp,
+                                            BoundedNormal_90(
+                                                parname_temp,
+                                                mu=param.priorpar1,
+                                                sigma=param.priorpar2))
                                 else:
-                                    setattr(self, parname_temp, pm.Normal(parname_temp, mu=param.priorpar1, sigma=param.priorpar2))
+                                    setattr(self, parname_temp,
+                                            pm.Normal(parname_temp,
+                                                      mu=param.priorpar1,
+                                                      sigma=param.priorpar2))
                             elif param.prior == 'LU':
-                                setattr(self, parname_temp, tt.exp(pm.Uniform(parname_temp, lower=param.priorpar1, upper=param.priorpar2)))
+                                setattr(self, parname_temp,
[…]
