### 2022-06-21 1389edb
Fixing issues caused by update to Astraeus
