### 2022-06-21 278a260
Fixing NIRISS S3 log save position
