### 2022-06-21 ec7b7a3
Fixing 2D MAD wavelength trimming
Before, normalization was done before wavelength trimming which
gives worse results.
