### 2022-06-22 37803dd
Tweaking Stage 4 for WFC3 normalization
Currently, the effect of summing the reads together is not saved, so I
needed to add some extra code to S4 that was already written in S3.
This will be addressed soon
