### 2022-06-22 fb38f77
Fixing memory overflow with S3 NIRISS
Also commenting out interpolating_image calls which are extremely slow.
We should discuss whether we want to keep this function and whether
there's anything we can do to speed it up.

The util.check_nans feature doesn't seem to be working as expected
either as there are 1449536 NaNs in the simualted data for the ERR and
V0 arrays.

And after that, I eventually get:
```
Traceback (most recent call last):
  File "/Users/tjbell1/Research/JWST/ERS/Eureka/ecfs_NIRISS/run_eureka.ecf", line 23, in <module>
    s3_meta = s3.reduce(eventlabel)
  File "/Users/tjbell
[…]
