### 2022-06-22 04f5fd4
Partially handling read summing outputs in S3
