+++ src/eureka/S3_data_reduction/plots_s3.py
@@ -34,27 +34,15 @@ def lc_nodriftcorr(data, meta, wave_1d, optspec):
 
     if meta.inst == 'wfc3':
         normspec = np.copy(optspec)
-        if meta.sum_reads:
-            # Sum each read from a scan together
-            nreads = meta.nreads
-            # Reshape to get (nfiles, nreads, nwaves)
-            normspec = normspec.reshape(-1, nreads, normspec.shape[1])
-            # Average together the reads to get (nfiles, nwaves)
-            normspec = normspec.sum(axis=1)
-            scandir = data.scandir.values[::nreads]
-            nreads = 1
-        else:
-            # Just leave as (nfiles*nreads, nwaves)
-            scandir = data.scandir.values
-            nreads = meta.nreads
+        scandir = np.repeat(meta.scandir, meta.nreads)
         
         # Normalize the data
         for p in range(2):
             iscans = np.where(scandir == p)[0]
             if len(iscans) > 0:
-                for r in range(nreads):
-                    normspec[iscans[r::nreads]] /= np.ma.mean(
-                        normspec[iscans[r::nreads]], axis=0)
+                for r in range(meta.nreads):
+                    normspec[iscans[r::meta.nreads]] /= np.ma.mean(
+                        normspec[iscans[r::meta.nreads]], axis=0)
     else:
         # Normalize the data
         normspec = optspec / np.ma.mean(optspec, axis=0)
