### 2022-06-22 d517b14
Documenting WFC3 edits to S3
