### 2022-06-22 f8cb4ca
Properly handling S3 sum_reads for WFC3
