### 2022-06-22 bdeea18
Fixing WFC3 fig4102
