### 2022-06-22 38f83b2
Fixing error calculation bug in S4f
