### 2022-06-23 1d81f21
changed try/except to if/else
