+++ src/eureka/S3_data_reduction/niriss_class.py
@@ -1,497 +1,0 @@
-import os
-import numpy as np
-from astropy.io import fits
-import time as time_pkg
-
-from .tracing_niriss import mask_method_edges, mask_method_ears, ref_file
-from ..lib.masking import data_quality_mask, interpolating_image
-from .background import fitbg3
-from .niriss_extraction import (dirty_mask, box_extract,
-                                optimal_extraction_routine)
-
-
-__all__ = ['NIRISS_S3']
-
-
-class NIRISS_S3(object):
-
-    def __init__(self, filename, f277_filename=None,
-                 data_dir=None, output_dir=None):
-        """Initializes the NIRISS S3 data reduction class.
-
-        Parameters
-        ----------
-        filename : str
-           The name of the FITS file with NIRISS observations.
-        data_path : str, optional
-           The path to where the input FITS files are stored. Default
-           is None. If None, will search the current working directory
-           for the files.
-        output_dir : str, optional
-           The path where output files will be saved. Default is None.
-           if None, will save all files to the current working directory.
-
-        Attributes
-        ----------
-        filename : str
-           The science file name.
-        data_path : str
-           The path to where the science file is stored.
-        output_dir : str
-           The path to where output files will be stored.
-        bkg : np.ndarray
-           Array of background values.
-        bkg_var : np.ndarray
-           Array of variance estimations on the background.
-        cr_mask : np.ndarray
-           Array of masks for bad pixels.
-        bkg_removed : np.ndarray
-           Data - background array.
-        trace_ear : astropy.table.Table
-           Astropy table with x,y coordinates for each order.
-           `trace_ear` is initialized when using the method `edges`.
-        trace_edge : astropy.table.Table
-           Astropy table with x,y coordinates for each order.
-           `trace_edge` is initialized when using the method `centers`.
-        box_spectra1 : np.ndarray
-           Box extracted spectra for the first order.
-        box_spectra2 : np.ndarray
-           Box extracted spectra for the second order.
-        box_mask_separate : np.ndarray
-           Attribute for separate box masks per each order. Created
-           when `dirty_mask(return_together == False)`.
-        """
-        self.filename = filename
-        self.datetime = time_pkg.strftime('%Y-%m-%d')
-
-        if data_dir is not None:
-            self.data_dir = data_dir
-        else:
-            self.data_dir = os.getcwd()
-
-        if output_dir is not None:
-            self.output_dir = output_dir
-        else:
-            self.output_dir = os.getcwd()
-
-        # Opens the science FITS file and sets up all proper
-        #   data attributes for the reduction steps.
-        self.setup()
-
-        self.dq_mask = data_quality_mask(self.dq)
-
-        if f277_filename is not None:
-            self.setup_f277(f277_filename)
-        else:
-            self.f277 = None
-            print('Without F277W filter image, some functions may not be '
-                  'available.')
-
-        self.trace_ear = None
-        self.trace_edge = None
-        self.bkg = None
-        self.box_mask = None
-        self.box_var1 = None
-        self.box_var2 = None
-        self.bkg_removed = None
-        self.box_spectra1 = None
-        self.box_spectra2 = None
-        self.box_mask_separate = None
-
-        return
-
-    def setup(self):
-        """Sets up all proper attributes from the FITS file.
-
-        Attributes
-        ----------
-        mhdr : FITS header
-           The main header in the FITS file.
-        shdr : FITS header
-           The header associated with the science frame
-           extension.
-        intend : float
-        time : np.ndarray
-           Array of time values based on the exposure start
-           and stop times indicated in the FITS header.
-        time_units : str
[…]
