+++ src/eureka/S3_data_reduction/tracing_niriss.py
@@ -47,9 +47,13 @@ def image_filtering(img, radius=1, gf=4):
     # applies the mask to the main frame
     data = img*mask
     g = gaussian_filter(data, gf)
-    g[g > 4] = 10000
+    g[g > 4] = 10000  # Tends to be a good cut-off for what is part of the
+                      # the profile, and what is background. 10000 is simply
+                      # an easy number to identify later.
     edges = filters.sobel(g)
-    edges[edges > 0] = 10
+    edges[edges > 0] = 10  # This defines the edges. Edges will have values
+                           # > 0, as set by the filter. 10 is arbitrary and
+                           # simply an easy number to identify later.
 
     # turns edge array into a boolean array
     edges = (edges-np.nanmax(edges)) * -1
