+++ src/eureka/S3_data_reduction/niriss_cython.pyx
@@ -3,41 +3,41 @@ import numpy as np
 __all__ = ['build_gaussian_images', 'build_moffat_images']
 
 def build_gaussian_images(data, A, B, sig, mu1, mu2, return_together=True):
-  """
-  Builds the two order profiles simultaneously and approximates as a
-  Gaussian profile.
+    """
+    Builds the two order profiles simultaneously and approximates as a
+    Gaussian profile.
 
-  Parameters
-  ----------
-  data : np.ndarray
-     The images.
-  A : float
-     Sets the scaling of the profile for the first order.
-  B : float
-     Sets the scaling of the profile for the second order.
-  sig : float
-     Sets the width of the profile for the orders.
-  mu1 : float
-     Sets the mean of the profile for the first order.
-  mu2 : float
-     Sets the mean of the profile for the second order.
-  return_together : bool, optional
-     Allows users to return a combined profile for both orders, or individual
-     profiles. Default is `True`.
+    Parameters
+    ----------
+    data : np.ndarray
+       The images.
+    A : float
+       Sets the scaling of the profile for the first order.
+    B : float
+       Sets the scaling of the profile for the second order.
+    sig : float
+       Sets the width of the profile for the orders.
+    mu1 : float
+       Sets the mean of the profile for the first order.
+    mu2 : float
+       Sets the mean of the profile for the second order.
+    return_together : bool, optional
+       Allows users to return a combined profile for both orders, or individual
+       profiles. Default is `True`.
 
-  Returns
-  -------
-  model : np.ndarray
-     Best-fit profile with both the first and second orders.
-  sigma : np.ndarray
-     Describes how well-fit the profile is to the data.
-  f1x : np.ndarray
-     Best-fit profile with only the first order. Returns if
-     `return_together = False`.
-  f2x : np.ndarray
-     Best-fit profile with only the second order. Returns if
-     `return_together = False`.
-  """
+    Returns
+    -------
+    model : np.ndarray
+       Best-fit profile with both the first and second orders.
+    sigma : np.ndarray
+       Describes how well-fit the profile is to the data.
+    f1x : np.ndarray
+       Best-fit profile with only the first order. Returns if
+       `return_together = False`.
+    f2x : np.ndarray
+       Best-fit profile with only the second order. Returns if
+       `return_together = False`.
+    """
 
     y = np.transpose(np.full((len(A), data.shape[-1], data.shape[0]),
                          np.arange(0,data.shape[0],1)), axes=(0,2,1))
@@ -61,6 +61,7 @@ def build_gaussian_images(data, A, B, sig, mu1, mu2, return_together=True):
        return model, sigma
     else:
        return f1x, f2x, sigma
+
 
 def build_moffat_images(data, A, alpha, gamma, mu1, mu2, return_together=True):
     """
@@ -96,7 +97,6 @@ def build_moffat_images(data, A, alpha, gamma, mu1, mu2, return_together=True):
        Best-fit profile with only the second order. Returns if
        `return_together = False`.
     """
-
     def moffat(x, A, alpha, gamma):
        frac = 1 + (y - x)**2.0 / gamma**2.0
        return A * frac**(-alpha)
