### 2022-06-23 226c5ed
Fixing error calculation bug in S4 again
I previously was counting over the entire wavelength axis, but I should
have only been including the wavelengths in the current bin using index.
This is now fixed
