### 2022-06-23 ad3d3a6
Fixing error calculation bug in S4
At some point the error calculation was messed up in S4, and the error
estimates were far too low. The issue was that the errors were being
divided by the total number of unmasked data points rather than the
number of unmasked data points used for that wavelength channel
