### 2022-06-23 c2400d4
Adding nicer fig4102 1D LC for WFC3
