+++ setup.cfg
@@ -68,7 +68,6 @@ docs =
     myst-parser
     nbsphinx
     numpydoc
-    recommonmark
     sphinx
     sphinx-automodapi
     sphinx-rtd-theme
