### 2022-06-23 e32630d
Improving docstrings for data object
