+++ src/eureka/S3_data_reduction/wfc3.py
@@ -114,7 +114,7 @@ def conclusion_step(data, meta, log):
     Parameters
     ----------
     data : Xarray Dataset
-        The Dataset object in which the fits data will stored.
+        The Dataset object.
     meta : eureka.lib.readECF.MetaClass
         The current metadata object.
     log : logedit.Logedit
