### 2022-06-23 dba6314
Removing old argument to lc_nodriftcorr
