### 2022-06-23 b6408fd
Uncommenting image_and_background call
