### 2022-06-23 adff216
Adding new get_reference_frames function for WFC3
