+++ src/eureka/S3_data_reduction/wfc3.py
@@ -51,61 +51,79 @@ def preparation_step(meta, log):
     meta.subdata_ref = []
     meta.subdiffmask_ref = []
 
+    meta = get_reference_frames(meta)
+
+    return meta, log
+
+
+def get_reference_frames(meta):
+    """Process the reference frames for each scan direction and save them
+    in the meta object.
+
+    Parameters
+    ----------
+    meta : eureka.lib.readECF.MetaClass
+        The current metadata object.
+
+    Returns
+    -------
+    meta : eureka.lib.readECF.MetaClass
+        The updated metadata object.
+    """
     # Temporarily override some values
     verbose = meta.verbose
     meta.verbose = False
     ncpu = meta.ncpu
     meta.ncpu = 1
+
     # Set some default values
     meta.firstFile = False
     meta.firstInBatch = False
     meta.int_start = 0
-    # Save the reference frame for each scan direction
+    meta.files_per_batch = 1
+
+    # Use the first two files by default
     if not hasattr(meta, 'iref'):
-        # Use the first two files by default
         meta.iref = [0, 1]
+    
     # Make sure that the scan directions are in the right order
     if meta.iref[0] % 2 != 0:
         meta.iref = meta.iref[::-1]
+
+    # Save the reference frame for each scan direction
     for i in meta.iref:
         log.writelog(f"Preparing reference frame {i}...")
         data = xrio.makeDataset()
         data, meta, log = read(meta.segment_list[i], data, meta, log)
         meta.n_int, meta.ny, meta.nx = data.flux.shape
         data, meta = util.trim(data, meta)
-        # Need to add guess after trimming
-        meta.guess.append(data.guess)
         data, meta = b2f.convert_to_e(data, meta, log)
         meta.src_ypos = source_pos.source_pos(
             data, meta, i, header=('SRCYPOS' in data.attrs['shdr']))
         data['mask'] = (['time', 'y', 'x'],
                         np.ones(data.flux.shape, dtype=bool))
-        # Check if arrays have NaNs
         data['mask'] = util.check_nans(data['flux'], data['mask'],
                                        log, name='FLUX')
         data['mask'] = util.check_nans(data['err'], data['mask'],
                                        log, name='ERR')
         data['mask'] = util.check_nans(data['v0'], data['mask'],
                                        log, name='V0')
-        # Manually mask regions [colstart, colend, rowstart, rowend]
         if hasattr(meta, 'manmask'):
-            for i in range(len(meta.manmask)):
-                colstart, colend, rowstart, rowend = meta.manmask[i]
-                data['mask'][rowstart:rowend, colstart:colend] = 0
-        meta.bg_y2 = meta.src_ypos + meta.bg_hw
-        meta.bg_y1 = meta.src_ypos - meta.bg_hw
+            util.manmask(data, meta, log)
         data = flag_bg(data, meta, log)
         data = background.BGsubtraction(data, meta, log, meta.isplots_S3)
-        meta.files_per_batch = 1
-        _ = cut_aperture(data, meta, log)
+        cut_aperture(data, meta, log)
+        
+        # Save the reference values
+        meta.guess.append(data.guess)
         meta.subdata_ref.append(data.flux)
         meta.subdiffmask_ref.append(data.flatmask)
-    
+
     # Restore input values
     meta.verbose = verbose
     meta.ncpu = ncpu
 
-    return meta, log
+    return meta
 
 
 def conclusion_step(data, meta, log):
