### 2022-06-23 f352144
Adding new sum_reads function for WFC3
