+++ src/eureka/S3_data_reduction/wfc3.py
@@ -51,12 +51,12 @@ def preparation_step(meta, log):
     meta.subdata_ref = []
     meta.subdiffmask_ref = []
 
-    meta = get_reference_frames(meta)
+    meta, log = get_reference_frames(meta, log)
 
     return meta, log
 
 
-def get_reference_frames(meta):
+def get_reference_frames(meta, log):
     """Process the reference frames for each scan direction and save them
     in the meta object.
 
@@ -123,7 +123,7 @@ def get_reference_frames(meta):
     meta.verbose = verbose
     meta.ncpu = ncpu
 
-    return meta
+    return meta, log
 
 
 def conclusion_step(data, meta, log):
@@ -141,6 +141,8 @@ def conclusion_step(data, meta, log):
 
     Returns
     -------
+    data : Xarray Dataset
+        The updated Dataset object.
     meta : eureka.lib.readECF.MetaClass
         The updated metadata object.
     log : logedit.Logedit
@@ -154,62 +156,89 @@ def conclusion_step(data, meta, log):
 
     if meta.sum_reads:
         # Sum each read from a scan together
-        # Get a copy of the flux and time arrays
-        flux = np.copy(data.optspec.values)
-        err = np.copy(data.opterr.values)
-        optmask = np.copy(data.optmask.values)
-        std_flux = np.copy(data.stdspec.values)
-        std_var = np.copy(data.stdvar.values)
-        time = np.copy(data.optspec.time.values)
-        # Reshape to get (nfiles, nreads, nwaves)
-        flux = flux.reshape(-1, meta.nreads, flux.shape[1])
-        err = err.reshape(-1, meta.nreads, err.shape[1])
-        optmask = optmask.reshape(-1, meta.nreads, optmask.shape[1])
-        std_flux = std_flux.reshape(-1, meta.nreads, std_flux.shape[1])
-        std_var = std_var.reshape(-1, meta.nreads, std_var.shape[1])
-        time = time.reshape(-1, meta.nreads)
-        # Sum together the reads to get (nfiles, nwaves)
-        flux = flux.sum(axis=1)
-        std_flux = std_flux.sum(axis=1)
-        # Add errors in quadrature
-        err = np.sqrt(np.sum(err**2, axis=1))
-        std_var = np.sqrt(np.sum(std_var**2, axis=1))
-        # Mask if any of the reads were masked
-        optmask = np.any(optmask, axis=1)
-        # Average together the reads' times
-        time = time.mean(axis=1)
-        
-        # Take every nread value out of the data object
-        data = data.isel(time=np.arange(0, len(data.optspec.time),
-                                        meta.nreads))
-        # FINDME: Need to handle all the other time-dependent arrays
-        # since right now we just take the value from the first read
-        
-        # Update values based on those we've calculated above
-        data.optspec.values = flux
-        data.optspec['time'] = time
-        data.opterr.values = err
-        data.opterr['time'] = time
-        data.optmask.values = optmask
-        data.optmask['time'] = time
-        data.stdspec.values = std_flux
-        data.stdspec['time'] = time
-        data.stdvar.values = std_var
-        data.stdvar['time'] = time
-        # For these values, just use the first read's value
-        data.drift2D['time'] = time
-        data.guess['time'] = time
-        data.scandir['time'] = time
-        data.wave['time'] = time
-
-        # Update nreads
-        meta.nreads = 1
+        data, meta = sum_reads(data, meta)
 
     return data, meta, log
 
 
+def sum_reads(data, meta):
+    """Sum together the non-destructive reads from each file to reduce noise
+    and data volume.
+
[…]
