### 2022-06-23 995ec99
Moving duplicated flux normalization to lib.util
