### 2022-06-23 efe0106
Fixing bug with WCF3 reference frame preparation
