### 2022-06-23 079f315
normalize_spectrum now uses mean
