### 2022-06-24 4d0185c
Better explaining horizonsfile in ecf.rst
