### 2022-06-24 f40108f
record width of drift ccf and plot it
