+++ eureka/S4_generate_lightcurves/plots_s4.py
@@ -78,6 +78,37 @@ def drift1d(meta, lc):
     plt.savefig(meta.outputdir+fname, bbox_inches='tight', dpi=300)
     if not meta.hide_plots:
         plt.pause(0.2)
+        
+def driftwidth(meta, lc):
+    '''Plot the 1D drift width results. (Fig 4104)
+
+    Parameters
+    ----------
+    meta : eureka.lib.readECF.MetaClass
+        The metadata object.
+    lc : Xarray Dataset
+        The light curve object containing drift arrays.
+
+    Returns
+    -------
+    None
+    '''
+    plt.figure(4104, figsize=(8, 4))
+    plt.clf()
+    plt.plot(np.arange(meta.n_int)[np.where(~lc.driftmask)],
+             lc.driftwidth[np.where(~lc.driftmask)], '.',
+             label='Good Drift Points')
+    plt.plot(np.arange(meta.n_int)[np.where(lc.driftmask)],
+             lc.driftwidth[np.where(lc.driftmask)], '.',
+             label='Interpolated Drift Points')
+    plt.ylabel('Spectrum Drift CC Width Along x')
+    plt.xlabel('Frame Number')
+    plt.legend(loc='best')
+    plt.tight_layout()
+    fname = 'figs'+os.sep+'fig4104_Drift'+figure_filetype
+    plt.savefig(meta.outputdir+fname, bbox_inches='tight', dpi=300)
+    if not meta.hide_plots:
+        plt.pause(0.2)
 
 
 def lc_driftcorr(meta, wave_1d, optspec):
