### 2022-06-24 27364db
add recordDrift to test ecf files
