### 2022-06-27 3ec1d5c
initial commit
generate LD coefficients with exotic-ld
