### 2022-06-27 e8cc3cd
add curvature correction for nirspec-g395h stage3
