+++ eureka/S3_data_reduction/s3_reduce.py
@@ -247,6 +247,14 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
                 data['medflux'].attrs['flux_units'] = \
                     data.flux.attrs['flux_units']
 
+                #correct G395H curvature
+                if meta.inst == 'nirspec' and data.mhdr['GRATING']=='G395H' and data.mhdr['FILTER']=='F290LP':
+                    log.writelog('  In NIRSpec G395H/F290LP setting:', mute=(not meta.verbose))
+                    log.writelog('  Correcting curvature and bringing trace in the center of the detector', mute=(not meta.verbose))
+                    log.writelog('  !!! Ensure that you are using meddata for the optimal extraction profile !!!', mute=(not meta.verbose))
+                    log.writelog('  Updating src_ypos and median frame accordingly.', mute=(not meta.verbose))
+                    data, meta = inst.straighten_trace(data, meta)
+
                 # Create bad pixel mask (1 = good, 0 = bad)
                 # FINDME: Will want to use DQ array in the future
                 # to flag certain pixels
