### 2022-06-28 a91218e
Limiting numpy < 1.23 for image_registration
The current version of image_registration doesn't work for numpy==1.23.0
so for the time being, I'm setting an upper limit on numpy until
image_registration gets patched
