+++ setup.cfg
@@ -43,7 +43,7 @@ install_requires =
     h5py < 3.2
     lmfit
     matplotlib
-    numpy >= 1.20.0
+    numpy >= 1.20.0, < 1.23  # Upper limit currently required by image_registration
     pandas
     photutils
     scipy >= 1.4.0 # Needed for scipy.fft
@@ -102,6 +102,8 @@ per-file-ignores =
     ./demos/**/run_*.py:E402
     # Need to protect against crashes if jwst not installed
     **/test_*.py:E402
+    # Need to protect against crashes if hst packages not installed
+    **/test_WFC3.py:F401,E402
     # Keep the unused numpy so users can pass in numpy values (e.g. np.inf)
     **/readECF.py:F401
     # Keep the aligned values to explain what they are
