### 2022-06-28 65f136e
Update installation docs for stable release v0.1-alpha
