### 2022-06-28 6747770
Remove np upper limit - image_registration patched
