+++ setup.cfg
@@ -43,7 +43,7 @@ install_requires =
     h5py < 3.2
     lmfit
     matplotlib
-    numpy >= 1.20.0, < 1.23  # Upper limit currently required by image_registration
+    numpy >= 1.20.0
     pandas
     photutils
     scipy >= 1.4.0 # Needed for scipy.fft
