### 2022-06-28 6f6a724
Useful error for WFC3 tests with missing packages
