### 2022-06-29 fcd7306
flake8 correction and new curvature keyword
