+++ src/eureka/S3_data_reduction/nirspec.py
@@ -157,8 +157,6 @@ def fit_bg(dataim, datamask, n, meta, isplots=0):
     return nircam.fit_bg(dataim, datamask, n, meta, isplots=isplots)
 
 
-
-#define functions that will compute the shifts
 def find_column_coms_and_shifts(data):
     '''
     Takes a data array (int, rows, columns) and find the 
@@ -176,33 +174,30 @@ def find_column_coms_and_shifts(data):
 
     '''
 
-    #get the dimensions of the data
-    nb_int = data.shape[0]
+    # get the dimensions of the data
     nb_rows = data.shape[1]
-    nc_cols = data.shape[-1]
 
-    #define a array of pixel positions (in their centers)
+    # define a array of pixel positions (in their centers)
     pix_centers = np.arange(nb_rows) + 0.5
 
-    #Compute the center of mass of each column and convert to integer (pixels)
-    column_coms = np.sum(pix_centers[None, :,None]*data, axis=1) / np.sum(data, axis=1)
+    # Compute the center of mass of each column and convert to integer (pixels)
+    column_coms = np.sum(pix_centers[None, :, None]*data, axis=1) / np.sum(data, axis=1)
     column_coms = np.around(column_coms).astype(int)
 
-    # #try using the max of each column
+    # # try using the max of each column
     # column_max = np.argmax(data, axis=1)
     # column_coms = column_max.astype(int)
 
-    #define the new center (where we will align the trace) in the middle of the detector
+    # define the new center (where we will align the trace) in the middle of the detector
     new_center = int(nb_rows/2)
 
-    #define an array containing the needed shift to bring the COMs to the
-    #new center for each column of each integration
+    # define an array containing the needed shift to bring the COMs to the
+    # new center for each column of each integration
     shifts = new_center - column_coms
 
-    #ensure columns of zeros or nans are not moved and dont induce bugs
-    shifts[column_coms<0] = 0
-    shifts[column_coms>nb_rows] = 0
-
+    # ensure columns of zeros or nans are not moved and dont induce bugs
+    shifts[column_coms < 0] = 0
+    shifts[column_coms > nb_rows] = 0
 
     return shifts, new_center
 
@@ -224,37 +219,34 @@ def find_column_median_shifts(data):
 
     '''
 
-    #get the dimensions of the data
+    # get the dimensions of the data
     nb_rows = data.shape[0]
-    nc_cols = data.shape[1]
 
-    #define a array of pixel positions (in their centers)
+    # define an array of pixel positions (in their centers)
     pix_centers = np.arange(nb_rows) + 0.5
 
-    #Compute the center of mass of each column and convert to integer (pixels)
-    column_coms = np.sum(pix_centers[:,None]*data, axis=0) / np.sum(data, axis=0)
+    # Compute the center of mass of each column and convert to integer (pixels)
+    column_coms = np.sum(pix_centers[:, None]*data, axis=0) / np.sum(data, axis=0)
     column_coms = np.around(column_coms).astype(int)
 
-    # #try using the max of each column
+    # # try using the max of each column
     # column_max = np.argmax(data, axis=1)
     # column_coms = column_max.astype(int)
 
-    #define the new center (where we will align the trace) in the middle of the detector
+    # define the new center (where we will align the trace) in the middle of the detector
     new_center = int(nb_rows/2)
 
-    #define an array containing the needed shift to bring the COMs to the
-    #new center for each column of each integration
+    # define an array containing the needed shift to bring the COMs to the
+    # new center for each column of each integration
     shifts = new_center - column_coms
 
[…]
