### 2022-06-29 69f1630
Make updates from Taylor's comments
