### 2022-06-29 b1abc39
Fixing centroidguess swapping bug
I'd accidentally been swapping guess for every file instead of just once
