### 2022-06-29 92d8840
Fixing S4 plotting bug
