### 2022-06-30 77e0125
Removed unneccesary update instruction
