### 2022-06-30 f9d8182
changed logs and updated new keyword in docs and test ecf
