+++ setup.cfg
@@ -36,6 +36,7 @@ install_requires =
     corner
     crds
     cython
+    dask < 2022.6.1
     dynesty > 1.0
     emcee > 3.0.0
     george # Needed for GP
