+++ src/eureka/S3_data_reduction/nirspec.py
@@ -157,68 +157,23 @@ def fit_bg(dataim, datamask, n, meta, isplots=0):
     return nircam.fit_bg(dataim, datamask, n, meta, isplots=isplots)
 
 
-def find_column_coms_and_shifts(data):
-    '''
-    Takes a data array (int, rows, columns) and find the 
-    center of mass (COM) in pixels for each column. It then returns the needed
-    shift to apply to each column to bring the COM to the center
-
-    Input:
-    data : (array) sequence of images. Shape: (Nb_integrations x Nb_rows x Nb_columns)
-
-    Return:
-    shifts : (array) shift to apply to each column of each integration to move
-                the trace to the middle of the detector and make it straight
-                Shape: (Nb_integrations x Nb_columns)
-
-
-    '''
-
-    # get the dimensions of the data
-    nb_rows = data.shape[1]
-
-    # define a array of pixel positions (in their centers)
-    pix_centers = np.arange(nb_rows) + 0.5
-
-    # Compute the center of mass of each column and convert to integer (pixels)
-    column_coms = np.sum(pix_centers[None, :, None]*data, axis=1) / np.sum(data, axis=1)
-    column_coms = np.around(column_coms).astype(int)
-
-    # # try using the max of each column
-    # column_max = np.argmax(data, axis=1)
-    # column_coms = column_max.astype(int)
-
-    # define the new center (where we will align the trace) in the middle of the detector
-    new_center = int(nb_rows/2)
-
-    # define an array containing the needed shift to bring the COMs to the
-    # new center for each column of each integration
-    shifts = new_center - column_coms
-
-    # ensure columns of zeros or nans are not moved and dont induce bugs
-    shifts[column_coms < 0] = 0
-    shifts[column_coms > nb_rows] = 0
-
-    return shifts, new_center
-
-
 def find_column_median_shifts(data):
-    '''
-    Takes the median frame (in time) and finds the 
+    '''Takes the median frame (in time) and finds the 
     center of mass (COM) in pixels for each column. It then returns the needed
     shift to apply to each column to bring the COM to the center
 
-    Input:
-    data : (array) median flux in time. Shape: (Nb_rows x Nb_columns)
+    Parameters
+    ----------
+    data : ndarray (2D)
+        The median of all data frames
 
-    Return:
-    shifts : (array) shift to apply to each column to move
-                the trace to the middle of the detector and make it straight
-                Shape: (Nb_columns)
-
-
+    Returns
+    -------
+    shifts : ndarray (2D)
+        The shifts to apply to each column to straighten the trace.
+    new_center : int
+        The central row of the detector where the trace is moved to.
     '''
-
     # get the dimensions of the data
     nb_rows = data.shape[0]
 
@@ -228,10 +183,6 @@ def find_column_median_shifts(data):
     # Compute the center of mass of each column and convert to integer (pixels)
     column_coms = np.sum(pix_centers[:, None]*data, axis=0) / np.sum(data, axis=0)
     column_coms = np.around(column_coms).astype(int)
-
-    # # try using the max of each column
-    # column_max = np.argmax(data, axis=1)
-    # column_coms = column_max.astype(int)
 
     # define the new center (where we will align the trace) in the middle of the detector
     new_center = int(nb_rows/2)
@@ -248,21 +199,21 @@ def find_column_median_shifts(data):
 
 
 def roll_columns(data, shifts):
-    '''
-    For each column of each integration, rolls the columns by the 
+    '''For each column of each integration, rolls the columns by the 
     values specified in shifts 
 
-    Input:
-    data   : (array) sequence of images. Shape: (Nb_integrations x Nb_rows x Nb_cols)
-    shifts : (array) shift to apply to each column of each integration
[…]
