### 2022-07-04 1fa5ebc
option to record y pos and width for each (file, integration)
