### 2022-07-04 63a946a
add option to choose integration + return Gaussian width
