+++ eureka/S3_data_reduction/source_pos.py
@@ -5,7 +5,7 @@ from scipy.optimize import curve_fit
 from . import plots_s3
 
 
-def source_pos(data, meta, m, header=False):
+def source_pos(data, meta, m, integ=0, header=False):
     '''Make image+background plot.
 
     Parameters
@@ -16,6 +16,9 @@ def source_pos(data, meta, m, header=False):
         The metadata object.
     m : int
         The file number.
+    integ : int
+        The integration number.
+        Default is 0 (first integration)
     header : bool; optional
         If True, use the source position in the FITS header.
         Defaults to False.
@@ -24,23 +27,31 @@ def source_pos(data, meta, m, header=False):
     -------
     src_ypos : int
         The central position of the star.
+    src_ypos_exact : float
+        The exact (not rounded) central position of the star.
+    src_ypos_width : float
+        If gaussian fit, the std of the Gaussian fitted to the image
+        Otherwise, array of zeros.
     '''
     if header:
         src_ypos = data.attrs['shdr']['SRCYPOS'] - meta.ywindow[0]
     elif meta.src_pos_type == 'weighted':
         # find the source location using a flux-weighted mean approach
-        src_ypos = source_pos_FWM(data.flux.values, meta, m)
+        src_ypos = source_pos_FWM(data.flux.values, meta, m, integ=integ)
     elif meta.src_pos_type == 'gaussian':
         # find the source location using a gaussian fit
-        src_ypos = source_pos_gauss(data.flux.values, meta, m)
+        src_ypos, src_ywidth = source_pos_gauss(data.flux.values, meta, m, integ=integ)
     else:
         # brightest row for source location
-        src_ypos = source_pos_max(data.flux.values, meta, m)
+        src_ypos = source_pos_max(data.flux.values, meta, m, integ=integ)
 
-    return round(src_ypos)
+    if meta.src_pos_type == 'gaussian':
+        return round(src_ypos), src_ypos, src_ywidth
+    else:
+        return round(src_ypos), src_ypos, np.zeros_like(src_ypos)
 
 
-def source_pos_max(flux, meta, m, plot=True):
+def source_pos_max(flux, meta, m, integ=0, plot=True):
     '''A simple function to find the brightest row for source location
 
     Parameters
@@ -51,6 +62,9 @@ def source_pos_max(flux, meta, m, plot=True):
         The metadata object.
     m : int
         The file number.
+    integ : int
+        The integration number.
+        Default is 0 (first integration)
     plot : bool; optional
         If True, plot the source position determination.
         Defaults to True.
@@ -72,7 +86,7 @@ def source_pos_max(flux, meta, m, plot=True):
 
     x_dim = flux.shape[1]
 
-    sum_row = np.sum(flux[0], axis=1)
+    sum_row = np.sum(flux[integ], axis=1)
     pos_max = np.argmax(sum_row)
 
     y_pixels = np.arange(0, x_dim)
@@ -85,7 +99,7 @@ def source_pos_max(flux, meta, m, plot=True):
     return pos_max
 
 
-def source_pos_FWM(flux, meta, m):
+def source_pos_FWM(flux, meta, m, integ=0):
     '''An alternative function to find the source location using a
     flux-weighted mean approach.
 
@@ -97,6 +111,9 @@ def source_pos_FWM(flux, meta, m):
         The metadata object.
     m : int
         The file number.
+    integ : int
+        The integration number.
+        Default is 0 (first integration)
 
     Returns
     -------
[…]
