### 2022-07-04 d7c0634
update to record ypos and widths in spec hdf5
