### 2022-07-05 8be87da
Implementing requested S3 changes
