### 2022-07-05 7e82832
No longer using masked arrays for S3-5
Rather than making a separate optspec_ma masked array throughout S4,
I've now tweaked it so that S4 uses spec.optmask instead. This had been
on our to-do list, and I wanted to do so to simplify moving the
sum_reads function from S3 to S4. While I was at it, I updated S3-5
