+++ src/eureka/S3_data_reduction/background.py
@@ -74,7 +74,7 @@ def BGsubtraction(data, meta, log, isplots):
         return
 
     # Compute background for each integration
-    log.writelog('  Performing background subtraction',
+    log.writelog('  Performing background subtraction...',
                  mute=(not meta.verbose))
     data['bg'] = (['time', 'y', 'x'], np.zeros(data.flux.shape))
     data['bg'].attrs['flux_units'] = data['flux'].attrs['flux_units']
