+++ src/eureka/lib/utc_tt.py
@@ -76,12 +76,13 @@ def leapdates(rundir, log):
         log.writelog('  NIST leap-second file not available. '
                      'Using stored table.')
 
-        return np.array([2441316.5, 2441682.5, 2442047.5, 2442412.5,
-                         2442777.5, 2443143.5, 2443508.5, 2443873.5,
-                         2444238.5, 2444785.5, 2445150.5, 2445515.5,
-                         2446246.5, 2447160.5, 2447891.5, 2448256.5,
-                         2448803.5, 2449168.5, 2449533.5, 2450082.5,
-                         2450629.5, 2451178.5, 2453735.5, 2454831.5])+1
+        return np.array([2441316.5, 2441498.5, 2441682.5, 2442047.5,
+                         2442412.5, 2442777.5, 2443143.5, 2443508.5,
+                         2443873.5, 2444238.5, 2444785.5, 2445150.5,
+                         2445515.5, 2446246.5, 2447160.5, 2447891.5,
+                         2448256.5, 2448803.5, 2449168.5, 2449533.5,
+                         2450082.5, 2450629.5, 2451178.5, 2453735.5,
+                         2454831.5, 2456108.5, 2457203.5, 2457753.5])+1
 
 
 def leapseconds(jd_utc, dates):
