### 2022-07-05 df97713
Changing hardcoded default iref for WFC3 S3
