+++ src/eureka/S3_data_reduction/wfc3.py
@@ -83,7 +83,10 @@ def get_reference_frames(meta, log):
 
     # Use the first two files by default
     if not hasattr(meta, 'iref'):
-        meta.iref = [2, 3]
+        raise AttributeError(
+            'You must set the meta.iref parameter in your ECF for WFC3 '
+            'observations. The recommended setting is [2, 3].'
+        )
     
     # Make sure that the scan directions are in the right order
     if meta.iref[0] % 2 != 0:
