### 2022-07-06 c5cc753
Adding hst_cal S3 ECF setting
