+++ docs/source/ecf.rst
@@ -168,14 +168,21 @@ As we want to do our own spectral extraction, we set this variable to ``calints`
 .. note::
 	Note that other Instruments might used different suffixes!
 
+hst_cal
+'''''''
+Only used for HST analyses. The fully qualified path to the folder containing HST calibration files.
 
 horizonsfile
 ''''''''''''
-Only used for HST analyses. This should be set to the fully qualified path to the horizons file you've downloaded from https://ssd.jpl.nasa.gov/horizons/app.html#/. On that website, 1. Select "Vector Table", 2. Select "HST", 3. Select "@ssb" (Solar System Barycenter), 4. Select a date range that spans the days relevant to your observations. Then click Generate Ephemeris and click Download Results.
+Only used for HST analyses. The path with respect to hst_cal to the horizons file you've downloaded from https://ssd.jpl.nasa.gov/horizons/app.html#/. To get a new horizons file on that website, 1. Select "Vector Table", 2. Select "HST", 3. Select "@ssb" (Solar System Barycenter), 4. Select a date range that spans the days relevant to your observations. Then click Generate Ephemeris and click Download Results.
 
 leapdir
 '''''''
-Only used for HST analyses. The folder where leapsecond calibration files will be saved
+Only used for HST analyses. The folder with respect to hst_cal where leapsecond calibration files will be saved.
+
+flatfile
+''''''''
+Only used for HST analyses. The path with respect to hst_cal to the flatfield file to use. A WFC3/G102 flat can be downloaded `here <http://www.stsci.edu/~WFC3/grism-resources/G102/WFC3.IR.G102.flat.2.fits.gz>`_ and a WFC3/G141 flat can be downloaded `here <http://www.stsci.edu/~WFC3/grism-resources/G141/WFC3.IR.G141.flat.2.fits.gz>`_; be sure to unzip the files after downloading them.
 
 ywindow & xwindow
 '''''''''''''''''
@@ -212,10 +219,6 @@ flatsigma
 '''''''''
 Only used for HST analyses. Used to sigma clip bad values from the flatfield image.
 
-flatfile
-''''''''
-Only used for HST analyses. The fully qualified path to the flatfield file to use. A WFC3/G102 flat can be downloaded `here <http://www.stsci.edu/~WFC3/grism-resources/G102/WFC3.IR.G102.flat.2.fits.gz>`_ and a WFC3/G141 flat can be downloaded `here <http://www.stsci.edu/~WFC3/grism-resources/G141/WFC3.IR.G141.flat.2.fits.gz>`_; be sure to unzip the files after downloading them.
-
 diffthresh
 ''''''''''
 Only used for HST analyses. Sigma theshold for bad pixel identification in the differential non-destructive reads (NDRs).
@@ -246,13 +249,11 @@ The plot below shows you which parts will be used for the background calculation
 
 .. image:: ../media/bg_hw.png
 
-
 bg_thresh
 '''''''''
 Double-iteration X-sigma threshold for outlier rejection along time axis.
 The flux of every background pixel will be considered over time for the current data segment.
 e.g: ``bg_thresh = [5,5]``: Two iterations of 5-sigma clipping will be performed in time for every background pixel. Outliers will be masked and not considered in the background flux calculation.
-
 
 bg_deg
 ''''''
@@ -279,26 +280,21 @@ p3thresh
 ''''''''
 Only important if ``bg_deg => 0`` (see above). # sigma threshold for outlier rejection during background subtraction which corresponds to step 3 of optimal spectral extraction, as defined by Horne (1986).
 
-
 p5thresh
 ''''''''
[…]
