### 2022-07-06 c3cc1aa
Add ability to download HST data using MAST API
