+++ src/eureka/lib/mastDownload.py
@@ -1,0 +1,248 @@
+
+import numpy as np
+import re, shutil, os
+from astroquery.mast import Observations
+import astropy.io.fits as pf
+
+def columnNames():
+    """
+    Print column names from MAST Observation table.
+
+    Notes
+    -----
+    History:
+
+    - June 2022 Kevin Stevenson
+        Initial version
+    """
+    meta_table = Observations.get_metadata("observations")
+    for val in meta_table['Column Name']:
+        print(val)
+    return
+
+def login(mast_token=None):
+    """
+    Log into the MAST portal.
+
+    Parameters
+    ----------
+    mast_token : string (optional)
+        The token to authenticate the user. Default is None.
+        This can be generated at https://auth.mast.stsci.edu/token.
+        If not supplied, it will be prompted for if not in the keyring
+        or set via $MAST_API_TOKEN.
+
+    Notes
+    -----
+    History:
+
+    - July 2022 Kevin Stevenson
+        Initial version
+    """
+    Observations.login(mast_token)
+    return
+
+def logout():
+    """
+    Log out of current MAST session.
+
+    Notes
+    -----
+    History:
+
+    - July 2022 Kevin Stevenson
+        Initial version
+    """
+    Observations.logout()
+    return
+
+def download(proposal_id, visit, inst='WFC3', download_dir='.', subgroup='IMA'):
+    """
+    Download observation visit number from specified proposal ID.
+
+    Parameters
+    ----------
+    proposal_id : string or int
+        HST proposal/program ID (e.g., 13467).
+
+    visit : string or int
+        HST visit number listed on the Visit Status Report (e.g., 60).
+        See https://www.stsci.edu/cgi-bin/get-visit-status?id=XXXXX,
+        where XXXXX is the proposal/program ID.
+
+    inst : string
+        HST instrument name, can be upper or lower case.
+        Supported options include: WFC3, STIS, COS, or FGS.
+
+    download_dir : string (optional)
+        Temporary download directory will be 'download_dir'/mastDownload/...
+
+    subgroup : string, (optional)
+        FITS file type (usually IMA, sometimes FLT)
+
+    Returns
+    -------
+    result : AstroPy Table
+        The manifest of files downloaded.
+
+    Notes
+    -----
+    History:
+
+    - June 2022 Kevin Stevenson
+        Initial version
+    """
+    # Convert to string
+    if type(proposal_id) is not str:
+        proposal_id = str(proposal_id).zfill(5)
+    if type(visit) is not str:
+        visit = str(visit).zfill(2)
+    # Determine instrument ID, as indicated by a single letter
+    # L=COS; I=WFC3; J=ACS; N=NICMOS; O=STIS; U=WFPC2; W=WFPC;
+    # X=FOC; Y=FOS; Z=GHRS; F=FGS; V=HSP;
+    if inst.casefold() == 'wfc3':
+        iid = 'i'
+    elif inst.casefold() == 'stis':
+        iid = 'o'
+    elif inst.casefold() == 'cos':
+        iid = 'l'
+    elif inst.casefold() == 'fgs':
+        iid = 'f'
+    else:
+        print(f"Unknown instrument: {inst}")
+        print("Supported options include: WFC3, STIS, COS, or FGS")
+        return None
+    # Specify obsid using wildcards
+    obsid = iid+'*'+visit+'*'
+
+    # Query MAST for requested visit
+    sci_table = Observations.query_criteria(proposal_id=proposal_id,
+                                            obs_id=obsid)
+
+    # AstroQuery doesn't support single character wildcards,
+    # so sometimes it identifies extra, unwanted files.
+    # Using regex to remove these unwanted files
+    counter = 0
[…]
