### 2022-07-06 07de87e
Tweaking log formatting
