+++ src/eureka/S4_generate_lightcurves/s4_genLC.py
@@ -20,7 +20,7 @@ import numpy as np
 import scipy.interpolate as spi
 import astraeus.xarrayIO as xrio
 from astropy.convolution import Box1DKernel
-from . import plots_s4, drift
+from . import plots_s4, drift, wfc3
 from ..lib import logedit
 from ..lib import readECF
 from ..lib import manageevent as me
@@ -174,7 +174,18 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
                                   spec.wave_1d.attrs['wave_units'],
                                   spec.optspec.attrs['time_units'],
                                   name='err')
-            lc = xrio.makeDataset({'data': lcdata, 'err': lcerr})
+            lcmask = xrio.makeLCDA(np.zeros((meta.nspecchan, meta.n_int),
+                                            dtype=bool),
+                                   meta.wave, spec.time.values, 'None',
+                                   spec.wave_1d.attrs['wave_units'],
+                                   spec.optspec.attrs['time_units'],
+                                   name='mask')
+            lc = xrio.makeDataset({'data': lcdata, 'err': lcerr,
+                                   'mask': lcmask})
+            if hasattr(spec, 'scandir'):
+                lc['scandir'] = spec.scandir
+            if hasattr(spec, 'drift2D'):
+                lc['drift2D'] = spec.drift2D
             lc['wave_low'] = (['wavelength'], meta.wave_low)
             lc['wave_hi'] = (['wavelength'], meta.wave_hi)
             lc['wave_mid'] = (lc.wave_hi + lc.wave_low)/2
@@ -188,24 +199,21 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
                 # The default value before this was added as an option
                 meta.boundary = 'extend'
 
-            # FINDME: The current implementation needs improvement,
-            # consider using optmask instead of masked arrays
-            # Create masked array for steps below
-            optspec_ma = np.ma.masked_array(spec.optspec, spec.optmask)
-            # Create opterr array with same mask as optspec
-            opterr_ma = np.ma.masked_array(spec.opterr, optspec_ma.mask)
-
             # Do 1D sigma clipping (along time axis) on unbinned spectra
             if meta.sigma_clip:
                 log.writelog('Sigma clipping unbinned optimal spectra along '
-                             'time axis')
+                             'time axis...')
                 outliers = 0
                 for w in range(meta.subnx):
-                    optspec_ma[:, w], nout = \
-                        clipping.clip_outliers(optspec_ma[:, w], log,
-                                               spec.wave_1d[w], meta.sigma,
-                                               meta.box_width, meta.maxiters,
-                                               meta.boundary, meta.fill_value,
+                    spec.optspec[:, w], spec.optmask[:, w], nout = \
+                        clipping.clip_outliers(spec.optspec[:, w], log,
+                                               spec.wave_1d[w],
+                                               mask=spec.optmask[:, w],
+                                               sigma=meta.sigma,
+                                               box_width=meta.box_width,
+                                               maxiters=meta.maxiters,
+                                               boundary=meta.boundary,
+                                               fill_value=meta.fill_value,
                                                verbose=meta.verbose)
                     outliers += nout
                 # Print summary if not verbose
@@ -221,7 +229,8 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
                 # This can take a long time, so always print this message
                 log.writelog('Applying drift/jitter correction')
                 # Compute drift/jitter
-                drift1d, driftmask = drift.spec1D(optspec_ma, meta, log)
+                drift1d, driftmask = drift.spec1D(spec.optspec, meta, log,
[…]
