+++ src/eureka/S3_data_reduction/s3_reduce.py
@@ -22,11 +22,13 @@
 # 16. Save Stage 3 data products
 # 17. Produce plots DONE
 
+import os
 import time as time_pkg
 import numpy as np
 import astraeus.xarrayIO as xrio
 from astropy.io import fits
 from tqdm import tqdm
+import psutil
 from . import optspex
 from . import plots_s3, source_pos
 from . import background as bg
@@ -185,7 +187,6 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
             else:
                 raise ValueError('Unknown instrument {}'.format(meta.inst))
 
-            datasets = []
             # Loop over each segment
             # Only reduce the last segment/file if testing_S3 is set to
             # True in ecf
@@ -193,25 +194,57 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
                 istart = meta.num_data_files - 1
             else:
                 istart = 0
-            for m in range(istart, meta.num_data_files):
-                # Initialize data object
-                data = xrio.makeDataset()
 
-                # Keep track if this is the first file - otherwise MIRI will
-                # keep swapping x and y windows
-                meta.firstFile = (m == istart and
-                                  meta.spec_hw == meta.spec_hw_range[0] and
-                                  meta.bg_hw == meta.bg_hw_range[0])
+            # Group files into batches
+            if not hasattr(meta, 'max_memory'):
+                meta.max_memory = 0.5
+            if not hasattr(meta, 'nfiles'):
+                meta.nfiles = 1
+            system_RAM = psutil.virtual_memory().total
+            filesize = os.path.getsize(meta.segment_list[istart])
+            maxfiles = max([1, int(system_RAM*meta.max_memory/filesize)])
+            meta.files_per_batch = min([maxfiles, meta.nfiles])
+            meta.nbatch = int(np.ceil((meta.num_data_files-istart) /
+                                      meta.files_per_batch))
+
+            datasets = []
+            for m in range(meta.nbatch):
+                first_file = m*meta.files_per_batch
+                last_file = min([meta.num_data_files,
+                                 (m+1)*meta.files_per_batch])
+                nfiles = last_file-first_file
+                               
                 # Report progress
-                if meta.verbose:
-                    log.writelog(f'Reading file {m + 1} of '
-                                 f'{meta.num_data_files}')
+                if meta.files_per_batch > 1:
+                    message = (f'Starting batch {m + 1} of {meta.nbatch} '
+                               f'with {nfiles} files')
                 else:
-                    log.writelog(f'Reading file {m + 1} of '
-                                 f'{meta.num_data_files}', end='\r')
+                    message = f'Starting file {m + 1} of {meta.num_data_files}'
+                if meta.verbose:
+                    log.writelog(message)
+                else:
+                    log.writelog(message, end='\r')
 
                 # Read in data frame and header
-                data, meta = inst.read(meta.segment_list[m], data, meta)
+                batch = []
+                for i in range(first_file, last_file):
+                    # Keep track if this is the first file - otherwise
+                    # MIRI will keep swapping x and y windows
+                    meta.firstFile = m == 0 and i == 0
+                    meta.firstInBatch = i == 0
+                    # Initialize a new data object
+                    data = xrio.makeDataset()
+                    data, meta, log = inst.read(meta.segment_list[i], data, 
+                                                meta, log)
+                    batch.append(data)
+
+                # Combine individual datasets
+                if meta.files_per_batch > 1:
+                    log.writelog('  Concatenating files...',
[…]
