### 2022-07-06 9ca97bb
Fixing masking issue in computeRMS
The allan plot before sometimes had no stderr line - this is now fixed
