### 2022-07-06 01602dc
Fixing intstart off-by-one error
Also adding logging to util.read_time
