### 2022-07-06 9a8aa65
Tweaking S4 log formatting
