### 2022-07-06 0455853
Properly handling new returns from inst.read
