### 2022-07-06 92a7ff6
Fixing more NIRSpec logs
