### 2022-07-06 e166333
Logging S4 mad for each binned wavelength
