### 2022-07-06 7063438
Logging S4 mad for each binned wavelength
