+++ src/eureka/S4_generate_lightcurves/s4_genLC.py
@@ -279,6 +279,7 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
             log.writelog("Generating light curves")
 
             # Loop over spectroscopic channels
+            meta.mad_s4_binned = []
             for i in range(meta.nspecchan):
                 log.writelog(f"  Bandpass {i} = {lc.wave_low.values[i]:.3f} - "
                              f"{lc.wave_hi.values[i]:.3f}")
@@ -313,7 +314,7 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
 
                 # Plot each spectroscopic light curve
                 if meta.isplots_S4 >= 3:
-                    plots_s4.binned_lightcurve(meta, lc, i)
+                    plots_s4.binned_lightcurve(meta, log, lc, i)
 
             # Calculate total time
             total = (time_pkg.time() - t0) / 60.
