### 2022-07-07 3f2610a
Removing old debug print statement
