### 2022-07-07 005e1ff
Fixing issues with xarrays and flux_white
