+++ src/eureka/S4_generate_lightcurves/s4_genLC.py
@@ -207,7 +207,8 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
                 for w in range(meta.subnx):
                     spec.optspec[:, w], spec.optmask[:, w], nout = \
                         clipping.clip_outliers(spec.optspec[:, w], log,
-                                               spec.wave_1d[w],
+                                               spec.wave_1d[w].values,
+                                               spec.wave_1d.attrs['wave_units'],
                                                mask=spec.optmask[:, w],
                                                sigma=meta.sigma,
                                                box_width=meta.box_width,
@@ -304,8 +305,9 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
                 if meta.sigma_clip:
                     lc['data'][i], lc['mask'][i], nout = \
                         clipping.clip_outliers(
-                            lc['data'][i], log, lc.wave_mid[i],
-                            mask=lc['mask'][i],
+                            lc.data[i], log, lc.data.wavelength[i].values,
+                            lc.data.attrs["wave_units"],
+                            mask=lc.mask[i],
                             sigma=meta.sigma, box_width=meta.box_width,
                             maxiters=meta.maxiters, boundary=meta.boundary,
                             fill_value=meta.fill_value, verbose=False)
@@ -320,57 +322,41 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
             if hasattr(meta, 'compute_white') and meta.compute_white:
                 log.writelog("Generating white-light light curve")
 
-                lcdata_white = xrio.makeLCDA(np.ma.zeros((1, meta.n_int)),
-                                             [np.mean(meta.wave), ],
-                                             spec.time.values,
-                                             spec.optspec.attrs['flux_units'],
-                                             spec.wave_1d.attrs['wave_units'],
-                                             spec.optspec.attrs['time_units'],
-                                             name='flux_white')
-                lcerr_white = xrio.makeLCDA(np.ma.zeros((1, meta.n_int)),
-                                            [np.mean(meta.wave), ],
-                                            spec.time.values,
-                                            spec.optspec.attrs['flux_units'],
-                                            spec.wave_1d.attrs['wave_units'],
-                                            spec.optspec.attrs['time_units'],
-                                            name='err_white')
-                mask_white = xrio.makeLCDA(np.zeros((1, meta.n_int),
-                                                    dtype=bool),
-                                           [np.mean(meta.wave), ],
-                                           spec.time.values, 'None',
-                                           spec.wave_1d.attrs['wave_units'],
-                                           spec.optspec.attrs['time_units'],
-                                           name='mask_white')
-                lc['flux_white'] = lcdata_white
-                lc['err_white'] = lcerr_white
-                lc['mask_white'] = mask_white
-
-                log.writelog(f"  White-light Bandpass = {meta.wave_min:.3f} - "
-                             f"{meta.wave_max:.3f}")
                 # Compute valid indeces within wavelength range
                 index = np.where((spec.wave_1d >= meta.wave_min) *
                                  (spec.wave_1d < meta.wave_max))[0]
+                central_wavelength = np.mean(spec.wave_1d[index].values)
+                lc.attrs['white_wavelength'] = central_wavelength
+                lc.attrs['flux_white'] = np.zeros((1, meta.n_int))
+                lc.attrs['err_white'] = np.zeros((1, meta.n_int))
+                lc.attrs['mask_white'] = np.zeros((1, meta.n_int), dtype=bool)
+                
+                log.writelog(f"  White-light Bandpass = {meta.wave_min:.3f} - "
+                             f"{meta.wave_max:.3f}")
                 # Make masked arrays for easy summing
[…]
