### 2022-07-07 9810f5c
Fixing more issues with xarrays and flux_white
