### 2022-07-07 638c360
return individual coeffs
