### 2022-07-07 2cf1204
Fixing masked array use in S5
