### 2022-07-07 d2dafba
Allow setting vmin, vmax in figs 3101, 4101
