### 2022-07-07 d3f7b76
new simpler bkg routine (should also throw no errors when testing)
