### 2022-07-07 9f8af53
should ignore 3rd order if boxsize3=None
