### 2022-07-07 cfccf03
Allow choosing time axis in figs 3101, 4101
Resolves #347
