+++ docs/source/ecf.rst
@@ -278,11 +278,15 @@ Sets how many plots should be saved when running Stage 3. A full description of 
 
 vmin
 ''''
-Optional. Sets the vmin of the color bar for Figure 3101.
+Optional. Sets the vmin of the color bar for Figure 3101. Defaults to 0.97.
 
 vmax
 ''''
-Optional. Sets the vmax of the color bar for Figure 3101.
+Optional. Sets the vmax of the color bar for Figure 3101. Defaults to 1.03.
+
+time_axis
+'''''''''
+Optional. Determines whether the time axis in Figure 3101 is along the y-axis ('left') or the x-axis ('bottom'). Defaults to 'left'.
 
 testing_S3
 ''''''''''
@@ -376,11 +380,15 @@ Sets how many plots should be saved when running Stage 4. A full description of 
 
 vmin
 ''''
-Optional. Sets the vmin of the color bar for Figure 4101.
+Optional. Sets the vmin of the color bar for Figure 4101. Defaults to 0.97.
 
 vmax
 ''''
-Optional. Sets the vmax of the color bar for Figure 4101.
+Optional. Sets the vmax of the color bar for Figure 4101. Defaults to 1.03.
+
+time_axis
+'''''''''
+Optional. Determines whether the time axis in Figure 4101 is along the y-axis ('left') or the x-axis ('bottom'). Defaults to 'left'.
 
 hide_plots
 ''''''''''
