### 2022-07-08 a8d4020
edit for using exotic-ld
to only allow for supported ld models when using exotic-ld generated ones
