### 2022-07-08 c39804b
overwriting fixed ld parameters
using generated ld model
