+++ src/eureka/S5_lightcurve_fitting/models/BatmanModels.py
@@ -45,9 +45,25 @@ class BatmanTransitModel(Model):
         self.paramtitles = kwargs.get('paramtitles')
 
         # Store the ld_profile
-        ld_func = ld_profile(self.parameters.limb_dark.value)
+        self.ld_from_S4 = kwargs.get('ld_from_S4')
+        ld_func = ld_profile(self.parameters.limb_dark.value, 
+                             use_gen_ld=self.ld_from_S4)
         len_params = len(inspect.signature(ld_func).parameters)
         self.coeffs = ['u{}'.format(n) for n in range(len_params)[1:]]
+        self.ld_S4_array = kwargs.get('ld_coeffs')[len_params-2]
+        
+        #Replace fixed u parameters with generated limb-darkening values
+        if self.ld_from_S4 != None:
+            for c in np.arange(self.nchan):
+                for u in self.coeffs:
+                    index = np.where(np.array(self.paramtitles) == u)[0]
+                    if len(index) != 0:
+                        item = self.longparamlist[c][index[0]]
+                        param = int(item[-1])
+                        if self.parameters.dict[item][1] == 'fixed':
+                            self.parameters.dict[item][0] = \
+                                self.ld_S4_array[c][param-1]
+            
 
     def eval(self, **kwargs):
         """Evaluate the function with the given values.
