+++ setup.cfg
@@ -38,6 +38,7 @@ install_requires =
     cython
     dynesty > 1.0
     emcee > 3.0.0
+    exotic-ld
     george # Needed for GP
     gwcs
     h5py < 3.2
