### 2022-07-08 6349536
add diagnostics plots for spatial position+width
