### 2022-07-08 7408ab1
if y position recorded, plot spatial position + width
