### 2022-07-08 4521927
Fixing a few breaking issues
meta had either trace_edge or trace_ears, but trace_edge was always used
in some places which broke the code. Also fixed missing import, missing
meta.time, missing log, and flake8 issue in niriss.py.
