### 2022-07-08 1a88b28
Resolving PR edits
Added astroquery as a requirement, fixing flake8 issues, fixed bug
caused by != None comparison
