### 2022-07-08 39a4a05
Allowing separate sci and cal folders for WFC3
