### 2022-07-11 a31adbc
Added initial ADS BibTeX citation.
