+++ docs/source/index.rst
@@ -23,7 +23,32 @@ The code is separated into six parts or "Stages":
 - Stage 5: Fits the light curves with noise and astrophysical models.
 - Stage 6: Creates a table and a figure summarizing the transmission and/or emission spectra from your many fits.
 
-The full code for ``Eureka!`` is available on `GitHub <http://github.com/kevin218/Eureka>`_
+The full code for ``Eureka!`` is available on `GitHub <http://github.com/kevin218/Eureka>`_.
+
+
+Citing ``Eureka!``
+~~~~~~~~~~~~~~~~~~
+
+If you wish to cite the use of ``Eureka!`` in published work, please use the following citation to the `JOSS paper <https://arxiv.org/abs/2207.03585>`_.
+
+.. code-block::
+
+    @ARTICLE{Bell2022,
+          author = {{Bell}, Taylor J. and {Ahrer}, Eva-Maria and {Brande}, Jonathan and {Carter}, Aarynn L. and {Feinstein}, Adina D. and {Guzman Caloca}, Giannina and {Mansfield}, Megan and {Zieba}, Sebastian and {Piaulet}, Caroline and {Benneke}, Bj{\"o}rn and {Filippazzo}, Joseph and {May}, Erin M. and {Roy}, Pierre-Alexis and {Kreidberg}, Laura and {Stevenson}, Kevin B.},
+            title = "{Eureka!: An End-to-End Pipeline for JWST Time-Series Observations}",
+          journal = {arXiv e-prints},
+        keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics},
+            year = 2022,
+            month = jul,
+              eid = {arXiv:2207.03585},
+            pages = {arXiv:2207.03585},
+    archivePrefix = {arXiv},
+          eprint = {2207.03585},
+    primaryClass = {astro-ph.IM},
+          adsurl = {https://ui.adsabs.harvard.edu/abs/2022arXiv220703585B},
+          adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+    }
+
 
 
 .. toctree::
