+++ src/eureka/S3_data_reduction/plots_s3.py
@@ -1,11 +1,13 @@
 import numpy as np
 import os
+from tqdm import tqdm
 import matplotlib.pyplot as plt
 from .source_pos import gauss
+from ..lib import util
 from ..lib.plots import figure_filetype
 
 
-def lc_nodriftcorr(meta, wave_1d, optspec):
+def lc_nodriftcorr(meta, wave_1d, optspec, optmask=None):
     '''Plot a 2D light curve without drift correction. (Fig 3101)
 
     Parameters
@@ -17,35 +19,37 @@ def lc_nodriftcorr(meta, wave_1d, optspec):
         which have been set in the S3 ecf
     optspec : ndarray
         The optimally extracted spectrum.
+    optmask : ndarray (1D), optional
+        A mask array to use if optspec is not a masked array. Defaults to None
+        in which case only the invalid values of optspec will be masked.
 
     Returns
     -------
     None
     '''
-    optspec = np.ma.masked_invalid(optspec)
-    plt.figure(3101, figsize=(8, 8))
-    plt.clf()
+    normspec = util.normalize_spectrum(meta, optspec, optmask=optmask)
     wmin = wave_1d.min()
     wmax = wave_1d.max()
-    n_int = optspec.shape[0]
     vmin = 0.97
     vmax = 1.03
-    normspec = optspec / np.ma.mean(optspec, axis=0)
+
+    plt.figure(3101, figsize=(8, 8))
+    plt.clf()
     plt.imshow(normspec, origin='lower', aspect='auto',
-               extent=[wmin, wmax, 0, n_int], vmin=vmin, vmax=vmax,
+               extent=[wmin, wmax, 0, meta.n_int], vmin=vmin, vmax=vmax,
                cmap=plt.cm.RdYlBu_r)
-    plt.title("MAD = " + str(np.round(meta.mad_s3, 0).astype(int)) + " ppm")
+    plt.title(f"MAD = {int(np.round(meta.mad_s3, 0))} ppm")
     plt.ylabel('Integration Number')
     plt.xlabel(r'Wavelength ($\mu m$)')
     plt.colorbar(label='Normalized Flux')
     plt.tight_layout()
-    fname = 'figs'+os.sep+'fig3101-2D_LC'+figure_filetype
+    fname = f'figs{os.sep}fig3101-2D_LC'+figure_filetype
     plt.savefig(meta.outputdir+fname, dpi=300)
     if not meta.hide_plots:
         plt.pause(0.2)
 
 
-def image_and_background(data, meta, n, m):
+def image_and_background(data, meta, log, m):
     '''Make image+background plot. (Figs 3301)
 
     Parameters
@@ -54,8 +58,8 @@ def image_and_background(data, meta, n, m):
         The Dataset object.
     meta : eureka.lib.readECF.MetaClass
         The metadata object.
-    n : int
-        The integration number.
+    log : logedit.Logedit
+        The current log.
     m : int
         The file number.
 
@@ -63,36 +67,75 @@ def image_and_background(data, meta, n, m):
     -------
     None
     '''
+    log.writelog('  Creating figures for background subtraction',
+                 mute=(not meta.verbose))
+
     intstart, subdata, submask, subbg = (data.attrs['intstart'],
-                                         data.flux.values, data.mask.values,
+                                         data.flux.values,
+                                         data.mask.values,
                                          data.bg.values)
     xmin, xmax = data.flux.x.min().values, data.flux.x.max().values
     ymin, ymax = data.flux.y.min().values, data.flux.y.max().values
 
-    plt.figure(3301, figsize=(8, 8))
[…]
