+++ src/eureka/lib/clipping.py
@@ -9,8 +9,9 @@ from astropy.stats import sigma_clip
 __all__ = ['clip_outliers', 'gauss_removal', 'time_removal']
 
 
-def clip_outliers(data, log, wavelength, sigma=10, box_width=5, maxiters=5,
-                  boundary='extend', fill_value='mask', verbose=False):
+def clip_outliers(data, log, wavelength, mask=None, sigma=10, box_width=5,
+                  maxiters=5, boundary='extend', fill_value='mask',
+                  verbose=False):
     '''Find outliers in 1D time series.
 
     Be careful when using this function on a time-series with known
@@ -20,12 +21,15 @@ def clip_outliers(data, log, wavelength, sigma=10, box_width=5, maxiters=5,
 
     Parameters
     ----------
-    data : ndarray (1D, float)
+    data : ndarray (1D)
         The input array in which to identify outliers
     log : logedit.Logedit
         The open log in which notes from this step can be added.
     wavelength : float
         The wavelength currently under consideration.
+    mask : ndarray (1D), optional
+        A mask array to use if data is not a masked array. Defaults to None
+        in which case only the invalid values of data will be masked.
     sigma : float; optional
         The number of sigmas a point must be from the rolling mean to be
         considered an outlier. Defaults to 10.
@@ -48,9 +52,11 @@ def clip_outliers(data, log, wavelength, sigma=10, box_width=5, maxiters=5,
 
     Returns
     -------
-    data : ndarray (1D, boolean)
+    data : ndarray (1D)
         An array with the same dimensions as the input array with outliers
         replaced with fill_value.
+    outliers : ndarray (1D)
+        A boolean array where True for values that were clipped.
     noutliers : int
         The number of outliers identified.
 
@@ -61,6 +67,9 @@ def clip_outliers(data, log, wavelength, sigma=10, box_width=5, maxiters=5,
     - Jan 29-31, 2022 Taylor Bell
         Initial version, added logging
     '''
+    data = np.ma.masked_invalid(np.ma.copy(data))
+    data = np.ma.masked_where(mask, data)
+        
     kernel = Box1DKernel(box_width)
     # Compute the moving mean
     bound_val = np.ma.median(data)  # Only used if boundary=='fill'
@@ -86,7 +95,7 @@ def clip_outliers(data, log, wavelength, sigma=10, box_width=5, maxiters=5,
     else:
         data[outliers] = fill_value
 
-    return data, np.sum(outliers)
+    return data, outliers, np.sum(outliers)
 
 
 def replace_moving_mean(data, outliers, kernel):

[…]
