### 2022-07-11 403962a
document changes to functions
