### 2022-07-11 b1da04e
More merge conflicts, update NIRISS S3 test ecf
