### 2022-07-11 2c56c24
Minor log tweaks
