### 2022-07-11 e58e90b
Reducing RAM impact of util.check_nans
