### 2022-07-11 54314f5
Avoiding repeated NIRISS extractions
There was an unneccessary for loop around the niriss extraction which
massively slowed things down. That's been fixed and I added some useful
logs while I was at it
