### 2022-07-11 64dc657
Fixing NIRISS spectrum normalization
