### 2022-07-11 5961411
Minor log tweaks
