### 2022-07-11 b7d15ab
Removing useless lines
