### 2022-07-11 627cacf
Fixing S3 NIRISS redirects
