### 2022-07-11 d14cb28
Handling NIRISS's multiple quadrants
