### 2022-07-11 6dfa4ae
Properly computing NIRISS MAD
