### 2022-07-11 aaa2b03
Populating data.optspec and data.opterr
