### 2022-07-11 ef5b8b9
Fixing NIRISS fig 3101 MAD title
