+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = eureka
-version = 0.0.1
+version = 0.2
 author = Eureka! pipeline developers
 author_email = kbstevenson@gmail.com
 description = A data reduction and analysis pipeline intended for time-series observations with the James Webb Space Telescope (JWST).
