### 2022-07-12 d733583
Allowing SRCYPOS to be overridden
Before, if the FITS header had a value for SRCYPOS, it had to be used.
Now users can choose to use that value with their meta.src_pos_type
setting ('header'), but do not have to do so.

This is required to resolve #165
