+++ setup.cfg
@@ -60,9 +60,9 @@ where = src
 
 [options.extras_require]
 jwst =
-    jwst==1.3.3
+    jwst==1.6.0
     stcal
-    asdf>=2.7.1,<2.11.0
+    asdf
 hst =
     # Need the GitHub version as 0.2.6 is required for python>=3.10, but 0.2.6 is not yet on PyPI
     image_registration @ git+https://github.com/keflavich/image_registration@master
