### 2022-07-12 0d4fb8d
minor edits from pytest + flake8 tests
