### 2022-07-12 f1521d0
Allowing white_free or white_fixed
Previously had to change code to switch between free and fixed, but can
now choose the behaviour with the EPFs
