### 2022-07-12 7776c68
Addressing updated stcal version
