+++ setup.cfg
@@ -61,7 +61,7 @@ where = src
 [options.extras_require]
 jwst =
     jwst==1.6.0
-    stcal
+    stcal >= 1.0.0 # Needed for our create_integration_model function
     asdf
 hst =
     # Need the GitHub version as 0.2.6 is required for python>=3.10, but 0.2.6 is not yet on PyPI
