### 2022-07-13 68ff643
Update s4_genLC.py
I think there was just one line causing the merge conflict - hopefully this fixes it!
