### 2022-07-13 2c76b3c
Download JWST data
