### 2022-07-13 bdd34da
move one line
otherwise trying to access None object
