### 2022-07-13 1d8452f
raise exception when ld not in lc
