+++ src/eureka/S5_lightcurve_fitting/s5_fit.py
@@ -144,6 +144,9 @@ def fitlc(eventlabel, ecf_path=None, s4_meta=None):
             # Load limb-darkening coefficients if used from Stage 4
             if hasattr(meta, 'use_generate_ld'):
                 ld_str = meta.use_generate_ld
+                if not hasattr(lc, 'exotic-ld_lin'):
+                    raise Exception("Exotic-ld coefficients have not been" +
+                                   "caluclated in Stage 4")
                 log.writelog("\nUsing generated limb-darkening coefficients" +
                              f"with {ld_str} \n")
                 ld_coeffs = [lc[ld_str + '_lin'].values, 
