### 2022-07-13 aba6be1
change hasattr to not None
