+++ demos/HST/download_data_HST_template.py
@@ -20,6 +20,7 @@ download_dir = '.'
 subgroup = 'IMA'
 
 # MAST API token for accessing data with exclusive access
+# This can be generated at https://auth.mast.stsci.edu/token
 mast_token = None
 
 # Final destination of files after calling mast.consolidate
