### 2022-07-14 ee65562
Tidying fitters.py's save_fit function
