### 2022-07-14 4c51876
Resolving some minor PR edits
