### 2022-07-14 0bc3407
Fixing S5 flux normalization with masks
