+++ src/eureka/S5_lightcurve_fitting/s5_fit.py
@@ -200,15 +200,15 @@ def fitlc(eventlabel, ecf_path=None, s4_meta=None):
                 flux = np.ma.masked_array([])
                 flux_err = np.ma.masked_array([])
                 for channel in range(chanrng):
-                    # FINDME: need to consider optmask
+                    mask = lc.mask.values[channel, :]
+                    flux_temp = np.ma.masked_where(mask,
+                                                   lc.data.values[channel, :])
+                    err_temp = np.ma.masked_where(mask,
+                                                  lc.err.values[channel, :])
                     flux = np.ma.append(flux,
-                                        (lc.data.values[channel, :] /
-                                         np.nanmean(
-                                             lc.data.values[channel, :])))
+                                        flux_temp/np.ma.mean(flux_temp))
                     flux_err = np.ma.append(flux_err,
-                                            (lc.err.values[channel, :] /
-                                             np.nanmean(
-                                                 lc.data.values[channel, :])))
+                                            err_temp/np.ma.mean(flux_temp))
 
                 meta, params = fit_channel(meta, time, flux, 0, flux_err,
                                            eventlabel, params, log,
@@ -226,13 +226,16 @@ def fitlc(eventlabel, ecf_path=None, s4_meta=None):
 
                     # Get the flux and error measurements for
                     # the current channel
-                    flux = lc.data.values[channel, :]
-                    flux_err = lc.err.values[channel, :]
+                    mask = lc.mask.values[channel, :]
+                    flux = np.ma.masked_where(mask,
+                                              lc.data.values[channel, :])
+                    flux_err = np.ma.masked_where(mask,
+                                                  lc.err.values[channel, :])
 
                     # Normalize flux and uncertainties to avoid large
                     # flux values
-                    flux_err = flux_err/np.nanmean(flux)
-                    flux = flux/np.nanmean(flux)
+                    flux_err = flux_err/np.ma.mean(flux)
+                    flux = flux/np.ma.mean(flux)
 
                     meta, params = fit_channel(meta, time, flux, channel,
                                                flux_err, eventlabel, params,
