+++ setup.cfg
@@ -96,6 +96,8 @@ exclude =
     .vscode/*
     # Ignore run_eureka.py files in demos folder
     **/run_eureka*.py
+    # Temporarily ignore NIRISS files
+    **/*niriss*.py
 per-file-ignores =
     # Need to have unused imports in the init files
     **/__init__.py:F401
