+++ src/eureka/S5_lightcurve_fitting/s5_fit.py
@@ -257,7 +257,7 @@ def fitlc(eventlabel, ecf_path=None, s4_meta=None):
 
 
 def fit_channel(meta, time, flux, chan, flux_err, eventlabel, params,
-                log, longparamlist, time_units, paramtitles, chanrng, ldcoeffs
+                log, longparamlist, time_units, paramtitles, chanrng, ldcoeffs,
                 white=False):
     """Run a fit for one channel or perform a shared fit.
 
