### 2022-07-14 d42882c
Fixing wavelength_range issue in generate_LD
