### 2022-07-14 6f0ee92
Undoing generate_LD.py wavelength_range "fix"
