### 2022-07-14 41d7582
Minor log tweaks
