### 2022-07-14 c9feaa6
Testing exotic-ld S4 code
I added the two minimal files needed to test the exotic-ld code without
downloading the ~1GB
