+++ src/eureka/S3_data_reduction/plots_s3.py
@@ -382,11 +382,11 @@ def driftypos(data, meta):
     '''
     plt.figure(3305, figsize=(8, 4))
     plt.clf()
-    plt.plot(np.arange(meta.n_int),data["driftypos"].values, '.')
+    plt.plot(np.arange(meta.n_int), data["driftypos"].values, '.')
     plt.ylabel('Spectrum spatial profile center')
     plt.xlabel('Frame Number')
     plt.tight_layout()
-    fname = 'figs'+os.sep+'fig3601_DriftYPos'+figure_filetype
+    fname = 'figs'+os.sep+'fig3305_DriftYPos'+figure_filetype
     plt.savefig(meta.outputdir+fname, bbox_inches='tight', dpi=300)
     if not meta.hide_plots:
         plt.pause(0.2)
@@ -419,7 +419,7 @@ def driftywidth(data, meta):
     plt.ylabel('Spectrum spatial profile width')
     plt.xlabel('Frame Number')
     plt.tight_layout()
-    fname = 'figs'+os.sep+'fig3602_DriftYWidth'+figure_filetype
+    fname = 'figs'+os.sep+'fig3306_DriftYWidth'+figure_filetype
     plt.savefig(meta.outputdir+fname, bbox_inches='tight', dpi=300)
     if not meta.hide_plots:
         plt.pause(0.2)
