### 2022-07-14 de14380
Allowing iterative sigma clipping
Before, the data was only smoothed with a boxcar filter once and that
sometimes left data unclipped that should have been clipped. This patch
allows iterative clipping to make sure outliers get clipped even if
there are many in a row
