### 2022-07-14 bd0a414
Fixing overly RAM intensive S3 NaN masking
Need to do masking on numpy arrays rather than xarrays to avoid extreme
RAM usage
