### 2022-07-14 df8bc33
Fixing NaNs in source_pos with real data
The real observations have NaNs all over the place from being masked out
in earlier stages, and these were not handled properly in source_pos.py
