+++ src/eureka/S3_data_reduction/source_pos.py
@@ -39,6 +39,9 @@ def source_pos(data, meta, m, integ=0):
         Enable recording of the width if the source is fitted with a Gaussian
         + add an option to fit any integration (not hardcoded to be the first)
     '''
+    # Mask any clipped values
+    flux = np.ma.masked_invalid(data.flux.values)
+
     if meta.src_pos_type == 'header':
         if 'SRCYPOS' not in data.attrs['shdr']:
             raise AttributeError('There is no SRCYPOS in the FITS header. '
@@ -47,16 +50,16 @@ def source_pos(data, meta, m, integ=0):
         src_ypos = data.attrs['shdr']['SRCYPOS'] - meta.ywindow[0]
     elif meta.src_pos_type == 'weighted':
         # find the source location using a flux-weighted mean approach
-        src_ypos = source_pos_FWM(data.flux.values, meta, m, integ=integ)
+        src_ypos = source_pos_FWM(flux, meta, m, integ=integ)
     elif meta.src_pos_type == 'gaussian':
         # find the source location using a gaussian fit
-        src_ypos, src_ywidth = source_pos_gauss(data.flux.values, meta, m,
+        src_ypos, src_ywidth = source_pos_gauss(flux, meta, m,
                                                 integ=integ)
     elif meta.src_pos_type == 'hst':
         src_ypos = data.guess.values[0]
     else:
         # brightest row for source location
-        src_ypos = source_pos_max(data.flux.values, meta, m, integ=integ)
+        src_ypos = source_pos_max(flux, meta, m, integ=integ)
 
     if meta.src_pos_type == 'gaussian':
         return int(round(src_ypos)), src_ypos, src_ywidth
@@ -101,13 +104,12 @@ def source_pos_max(flux, meta, m, integ=0, plot=True):
 
     x_dim = flux.shape[1]
 
-    sum_row = np.sum(flux[integ], axis=1)
-    pos_max = np.argmax(sum_row)
-
-    y_pixels = np.arange(0, x_dim)
+    sum_row = np.ma.sum(flux[integ], axis=1)
+    pos_max = np.ma.argmax(sum_row)
 
     # Diagnostic plot
     if meta.isplots_S3 >= 3 and plot:
+        y_pixels = np.arange(0, x_dim)
         plots_s3.source_position(meta, x_dim, pos_max, m, y_pixels=y_pixels,
                                  sum_row=sum_row)
 
@@ -153,11 +155,11 @@ def source_pos_FWM(flux, meta, m, integ=0):
 
     y_pixels = np.arange(0, x_dim)[pos_max-meta.spec_hw:pos_max+meta.spec_hw]
 
-    sum_row = np.sum(flux[integ],
-                     axis=1)[pos_max-meta.spec_hw:pos_max+meta.spec_hw]
+    sum_row = np.ma.sum(flux[integ],
+                        axis=1)[pos_max-meta.spec_hw:pos_max+meta.spec_hw]
     sum_row -= (sum_row[0]+sum_row[-1])/2
 
-    y_pos = np.sum(sum_row * y_pixels) / np.sum(sum_row)
+    y_pos = np.ma.sum(sum_row * y_pixels)/np.ma.sum(sum_row)
 
     # Diagnostic plot
     if meta.isplots_S3 >= 3:
@@ -237,14 +239,16 @@ def source_pos_gauss(flux, meta, m, integ=0):
     x_dim = flux.shape[1]
 
     # Data cutout around the maximum row
-    pos_max = source_pos_max(flux, meta, m, plot=False)
[…]
