### 2022-07-14 454e258
Using data.mask through S3 to mask bad pixels
