+++ src/eureka/S3_data_reduction/plots_s3.py
@@ -86,10 +86,10 @@ def image_and_background(data, meta, log, m):
     log.writelog('  Creating figures for background subtraction...',
                  mute=(not meta.verbose))
 
-    intstart, subdata, submask, subbg = (data.attrs['intstart'],
-                                         data.flux.values,
-                                         data.mask.values,
-                                         data.bg.values)
+    intstart = data.attrs['intstart']
+    subdata = np.ma.masked_where(~data.mask.values, data.flux.values)
+    subbg = np.ma.masked_where(~data.mask.values, data.bg.values)
+
     xmin, xmax = data.flux.x.min().values, data.flux.x.max().values
     ymin, ymax = data.flux.y.min().values, data.flux.y.max().values
 
@@ -102,15 +102,15 @@ def image_and_background(data, meta, log, m):
         plt.suptitle(f'Integration {intstart + n}')
         plt.subplot(211)
         plt.title('Background-Subtracted Flux')
-        max = np.max(subdata[n] * submask[n])
-        plt.imshow(subdata[n]*submask[n], origin='lower', aspect='auto',
+        max = np.ma.max(subdata[n])
+        plt.imshow(subdata[n], origin='lower', aspect='auto',
                    vmin=0, vmax=max/10, extent=[xmin, xmax, ymin, ymax])
         plt.colorbar()
         plt.ylabel('Detector Pixel Position')
         plt.subplot(212)
         plt.title('Subtracted Background')
-        median = np.median(subbg[n])
-        std = np.std(subbg[n])
+        median = np.ma.median(subbg[n])
+        std = np.ma.std(subbg[n])
         plt.imshow(subbg[n], origin='lower', aspect='auto', vmin=median-3*std,
                    vmax=median+3*std, extent=[xmin, xmax, ymin, ymax])
         plt.colorbar()
