### 2022-07-14 24aede6
Fixing since np.ma.isfinite doesn't exist
