### 2022-07-15 834a328
Fixing WFC3 S3 test
Forgot to move masking earlier in the get_reference_frames function
