### 2022-07-15 90b97de
Enhance data download scripts
