+++ demos/HST/download_data_HST_template.py
@@ -20,7 +20,6 @@ download_dir = '.'
 subgroup = 'IMA'
 
 # MAST API token for accessing data with exclusive access
-# This can be generated at https://auth.mast.stsci.edu/token
 mast_token = None
 
 # Final destination of files after calling mast.consolidate
@@ -35,7 +34,7 @@ for vis in visits:
         # Consolodate and move data into new directory
         md.consolidate(result, final_dir)
         # Sort data into science and calibration folders (scan vs direct image)
-        md.sort(final_dir)
+        md.sortHST(final_dir)
 
 # Delete empty temporary directory structure
 md.cleanup(download_dir)
