+++ src/eureka/lib/mastDownload.py
@@ -23,18 +23,23 @@ def columnNames():
         print(val)
     return
 
-
-def writeTable(table, filename, format='csv'):
-    """Write intermediate results from Astropy Table to an ASCII file
+def writeTable_JWST(proposal_id, observation, visit, filename, format='csv'):
+    """Write all products from specified visit to an ASCII file
 
     Parameters
     ----------
-    table :
-        Astropy Table
-    filename : str
+    proposal_id : string or int
+        JWST proposal/program ID (e.g., 1366).
+    observation : string or int
+        JWST observation number listed on the Visit Status Report (e.g., 2).
+        See https://www.stsci.edu/cgi-bin/get-visit-status?id=XXXXX&observatory=JWST,
+        where XXXXX is the proposal/program ID.
+    visit : string or int
+        JWST visit number listed on the Visit Status Report (e.g., 1).
+        See https://www.stsci.edu/cgi-bin/get-visit-status?id=XXXXX&observatory=JWST,
+        where XXXXX is the proposal/program ID.
+    filename : string
         Name of output file
-    format : str; optional
-        The file format to use. Defaults to 'csv'.
 
     Notes
     -----
@@ -43,9 +48,21 @@ def writeTable(table, filename, format='csv'):
     - July 2022 Kevin Stevenson
         Initial version
     """
-    ascii.write(table, filename, format=format)
-    return
+    # Convert to string
+    if type(proposal_id) is not str:
+        proposal_id = str(proposal_id).zfill(5)
+    if type(observation) is not str:
+        observation = str(observation).zfill(3)
+    if type(visit) is not str:
+        visit = str(visit).zfill(3)
+    # Specify obsid using wildcards
+    obsid = 'jw'+proposal_id+observation+visit+'*'
 
+    # Query MAST for requested visit
+    sci_table = Observations.query_criteria(proposal_id=proposal_id,
+                                            obs_id=obsid)
+    ascii.write(sci_table, filename, format=format)
+    return
 
 def login(mast_token=None):
     """Log into the MAST portal.
@@ -84,26 +101,24 @@ def logout():
 
 
 def downloadHST(proposal_id, visit, inst='WFC3', download_dir='.',
-                subgroup='IMA'):
+             subgroup='IMA'):
     """Download observation visit number from specified proposal ID.
 
     Parameters
     ----------
-    proposal_id : str or int
+    proposal_id : string or int
         HST proposal/program ID (e.g., 13467).
-    visit : str or int
+    visit : string or int
         HST visit number listed on the Visit Status Report (e.g., 60).
         See https://www.stsci.edu/cgi-bin/get-visit-status?id=XXXXX,
         where XXXXX is the proposal/program ID.
-    inst : str; optional
+    inst : string
         HST instrument name, can be upper or lower case.
         Supported options include: WFC3, STIS, COS, or FGS.
-        Defaults to 'WFC3'.
-    download_dir : str; optional
+    download_dir : string (optional)
         Temporary download directory will be 'download_dir'/mastDownload/...
-        Defaults to '.'.
-    subgroup : str; optional
-        FITS file type (usually IMA, sometimes FLT). Defaults to 'IMA'.
+    subgroup : string, (optional)
+        FITS file type (usually IMA, sometimes FLT)
 
     Returns
     -------
@@ -171,17 +186,20 @@ def downloadHST(proposal_id, visit, inst='WFC3', download_dir='.',
                                             download_dir=download_dir)
     return result
 
-
-def downloadJWST(proposal_id, visit, calib_level, subgroup, download_dir='.'):
-    """Download observation visit number from specified proposal ID.
[…]
