+++ demos/JWST/download_data_JWST_template.py
@@ -1,13 +1,17 @@
 #!/usr/bin/env python3
 
+from astroquery.mast import Observations
 import eureka.lib.mastDownload as md
+from astropy.io import ascii
 
 # Proposal/Program ID, can be string or int
 proposal_id = '02734'
 
-# List of one or more visit numbers
-visits = [2]
-# visits = np.arange(1,4)
+# Observation number
+observation = 1
+
+# Visit number
+visit = 1
 
 # Calibration level, list
 # (0 = raw, 1 = uncalibrated, 2 = calibrated, 3 = science product,
@@ -22,27 +26,37 @@ calib_level = [1]
 subgroup = 'UNCAL'
 
 # Temporary download directory will be 'download_dir'/mastDownload/...
-download_dir = '.'
-
-# MAST API token for accessing data with exclusive access
-# This can be generated at https://auth.mast.stsci.edu/token
-mast_token = None
+download_dir = './wasp96b'
 
 # Final destination of files after calling mast.consolidate
-final_dir = './wasp96b/S1'
+final_dir = './wasp96b/S0'
 
+# MAST API token for accessing data with exclusive access
+mast_token = None
 # If data are public, no need to call md.login() or md.logout()
 md.login(mast_token)
-for vis in visits:
-    # Download data from MAST Archive
-    result = md.downloadJWST(proposal_id, vis, calib_level, subgroup,
-                             download_dir)
-    if result is not None:
-        # Consolodate and move data into new directory
-        md.consolidate(result, final_dir)
-        # Sort data into science and calibration folders (scan vs direct image)
-        # md.sort(final_dir)
+
+# Apply standard filters to identify files for download
+table = md.filterJWST(proposal_id, observation, visit, calib_level, subgroup)
+
+# Optional, write out list of products from filtered table
+# ascii.write(table, download_dir+"/output.csv", format='csv')
+# Optional, write out all products from given program, observation, visit
+# md.writeTable_JWST(proposal_id, observation, visit, download_dir+"/output.csv", format='csv')
+
+# Download data products, returns manifest of files downloaded.
+manifest = Observations.download_products(table, curl_flag=False,
+                                          download_dir=download_dir)
+
+# Consolidate and move data into new directory
+md.consolidate(manifest, final_dir)
+
+# Sort files by their type
+md.sortJWST(final_dir, final_dir+'/S0', "uncal.fits")
 
 # Delete empty temporary directory structure
 md.cleanup(download_dir)
+
+# Logout
 md.logout()
+
