### 2022-07-15 7ac4197
Fixing time arrays for real data
We incorrectly guessed that each segment would have the whole time
array. Each segment only has the time values for that segment, so no
need to crop the arrays
