+++ src/eureka/S3_data_reduction/nircam.py
@@ -56,8 +56,7 @@ def read(filename, data, meta, log):
     dq = hdulist['DQ', 1].data
     v0 = hdulist['VAR_RNOISE', 1].data
     wave_2d = hdulist['WAVELENGTH', 1].data
-    int_times = hdulist['INT_TIMES', 1].data[data.attrs['intstart']:
-                                             data.attrs['intend']]
+    int_times = hdulist['INT_TIMES', 1].data
 
     # Record integration mid-times in BJD_TDB
     if (hasattr(meta, 'time_file') and meta.time_file is not None):
