### 2022-07-15 f19b10f
More useful error message if no FITS files found
Because of an update I made to util.readfiles, an unhelpful error
message would be raised if no files were found. This patch reduces
duplicated code and raises a better message if no files are found
