### 2022-07-15 15a22e8
Fixing time array for NIRCam simulated data
The NIRCam simulated data had the full time array for all segments which
led us astray for the real data, but we still need to be able to handle
the simulated data. This tweak allows both
