### 2022-07-15 58699ff
Fixing missing verbose keyword for S1-2
S1-2 have no verbose setting, so the readfiles function broke for those.
The function will now only mute if meta has verbose and verbose is set
to False
