### 2022-07-15 1734c81
Trying out automatic version numbering
Using setuptools_scm to allow automatic version numbering
