+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = eureka
-version = attr: eureka.version.__version__
+version = 0.2.1
 author = Eureka! pipeline developers
 author_email = kbstevenson@gmail.com
 description = A data reduction and analysis pipeline intended for time-series observations with the James Webb Space Telescope (JWST).
