+++ .github/workflows/pytest_testing.yml
@@ -26,7 +26,10 @@ jobs:
           echo "CRDS_PATH=$HOME/crds_cache" >> $GITHUB_ENV
           echo "CRDS_SERVER_URL=https://jwst-crds.stsci.edu" >> $GITHUB_ENV
       - name: Log the current Eureka version
-        run: pip show eureka
+        run: |
+          pip show eureka
+          git describe
+          git log -5
       - name: Run tests, generate coverage report
         run: pytest tests --cov=./src/eureka --cov-report=xml
       - name: Upload coverage reports to Codecov with GitHub Action
