+++ .github/workflows/pytest_testing.yml
@@ -14,7 +14,10 @@ jobs:
       OS: 'ubuntu-latest'
       PYTHON: '3.9'
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+          ref: ${{ github.event.pull_request.head.ref }}
       - uses: actions/setup-python@v1
         with:
           python-version: 3.9
