### 2022-07-15 665f46f
Tidying up testing workflow
