+++ .github/workflows/flake8_linting.yml
@@ -14,9 +14,9 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v3
-        with:
-          fetch-depth: 0
-          ref: ${{ github.event.pull_request.head.ref }}
+      with:
+        fetch-depth: 0
+        ref: ${{ github.event.pull_request.head.ref }}
     - uses: tonybajan/flake8-check-action@v1.4.0
       with:
         select: E1,E2,E3,E4,E5,E7,W6,F
