### 2022-07-15 bd6c45d
Removing all hardcoded eureka version numbers
