+++ setup.cfg
@@ -1,6 +1,5 @@
 [metadata]
 name = eureka
-version = 0.2.1
 author = Eureka! pipeline developers
 author_email = kbstevenson@gmail.com
 description = A data reduction and analysis pipeline intended for time-series observations with the James Webb Space Telescope (JWST).
