### 2022-07-16 14cbdd3
Fixing NIRSpec x-axis trimming
