+++ src/eureka/S2_calibrations/s2_calibrate.py
@@ -27,6 +27,7 @@ from ..lib.plots import figure_filetype
 import jwst.assign_wcs.nirspec
 from functools import partial
 
+
 def calibrateJWST(eventlabel, ecf_path=None, s1_meta=None):
     '''Reduces rateints spectrum or image files ouput from Stage 1 of the JWST
     pipeline into calints and x1dints.
@@ -128,8 +129,8 @@ def calibrateJWST(eventlabel, ecf_path=None, s1_meta=None):
             # wavelength range that you want to trim to.
             jwst.assign_wcs.nirspec.nrs_wcs_set_input = \
                 partial(jwst.assign_wcs.nirspec.nrs_wcs_set_input, 
-                        wavelength_range=[meta.waverange_start,
-                                            meta.waverange_end])
+                    wavelength_range=[meta.waverange_start,
+                                        meta.waverange_end])
     elif telescope == 'HST':
         log.writelog('There is no Stage 2 for HST - skipping.')
         # Clean up temporary folder
