+++ src/eureka/S4_generate_lightcurves/generate_LD.py
@@ -65,4 +65,5 @@ def exotic_ld(meta, spec):
         nonlin_4[i] = \
             sld.compute_4_parameter_non_linear_ld_coeffs(wavelength_range[i], 
                                                          mode)
+
     return lin_c1, quad, nonlin_3, nonlin_4
