### 2022-07-18 a872b05
implemented photometry
