### 2022-07-18 1880233
Fixing flake8 issues
