+++ src/eureka/S2_calibrations/s2_calibrate.py
@@ -129,8 +129,8 @@ def calibrateJWST(eventlabel, ecf_path=None, s1_meta=None):
             # wavelength range that you want to trim to.
             jwst.assign_wcs.nirspec.nrs_wcs_set_input = \
                 partial(jwst.assign_wcs.nirspec.nrs_wcs_set_input, 
-                    wavelength_range=[meta.waverange_start,
-                                        meta.waverange_end])
+                        wavelength_range=[meta.waverange_start,
+                                          meta.waverange_end])
     elif telescope == 'HST':
         log.writelog('There is no Stage 2 for HST - skipping.')
         # Clean up temporary folder
