+++ src/eureka/S3_data_reduction/bright2flux.py
@@ -91,7 +91,8 @@ def dn2electrons(data, meta):
     ystart_trim = ystart-ystart_gain + 1  # 1 indexed, NOT zero                                                         
     xstart_trim = xstart-xstart_gain + 1
 
-    gain = fits.getdata(meta.gainfile)[ystart_trim:ystart_trim+ny, xstart_trim:xstart_trim+nx]
+    gain = fits.getdata(meta.gainfile)[ystart_trim:ystart_trim+ny,
+                                       xstart_trim:xstart_trim+nx]
 
     # Like in the case of MIRI data, the gain file data has to be
     # rotated by 90 degrees
