+++ src/eureka/S3_data_reduction/bright2flux.py
@@ -94,10 +94,11 @@ def dn2electrons(data, meta):
     gain = fits.getdata(meta.gainfile)[ystart_trim:ystart_trim+ny,
                                        xstart_trim:xstart_trim+nx]
 
-    # Like in the case of MIRI data, the gain file data has to be
-    # rotated by 90 degrees
     if data.attrs['shdr']['DISPAXIS'] == 2:
-        gain = np.swapaxes(gain, 0, 1)
+        # In the case of MIRI data, the gain file data has to be
+        # rotated by 90 degrees and mirrored along that new x-axis
+        # so that wavelength increases to the right
+        gain = np.swapaxes(gain, 0, 1)[:, ::-1]
 
     # Gain subarray
     subgain = gain[meta.ywindow[0]:meta.ywindow[1],
