+++ src/eureka/S3_data_reduction/optspex.py
@@ -1,5 +1,6 @@
 import numpy as np
 import matplotlib.pyplot as plt
+from tqdm import tqdm
 from ..lib import gaussian as g
 from ..lib import smooth
 from . import plots_s3
@@ -477,9 +478,54 @@ def profile_gauss(subdata, mask, threshold=10, guess=None, isplots=0):
     return profile
 
 
-def optimize(meta, subdata, mask, bg, spectrum, Q, v0, p5thresh=10,
-             p7thresh=10, fittype='smooth', window_len=21, deg=3,
-             windowtype='hanning', n=0, m=0, meddata=None):
+def optimize(data, meta, log, apdata, apmask, apbg, apv0, gain=1,
+             windowtype='hanning', m=0):
+
+    # Extract optimal spectrum with uncertainties
+    log.writelog("  Performing optimal spectral extraction...",
+                 mute=(not meta.verbose))
+    data['optspec'] = (['time', 'x'], np.zeros(data.stdspec.shape))
+    data['opterr'] = (['time', 'x'], np.zeros(data.stdspec.shape))
+    data['optmask'] = (['time', 'x'], np.ones(data.stdspec.shape))
+    data['optspec'].attrs['flux_units'] = data.flux.attrs['flux_units']
+    data['optspec'].attrs['time_units'] = data.flux.attrs['time_units']
+    data['opterr'].attrs['flux_units'] = data.flux.attrs['flux_units']
+    data['opterr'].attrs['time_units'] = data.flux.attrs['time_units']
+    data['optmask'].attrs['flux_units'] = 'None'
+    data['optmask'].attrs['time_units'] = data.flux.attrs['time_units']
+
+    # Compute median frame
+    data_ma = np.ma.masked_where(data.mask.values == 0, data.flux.values)
+    medflux = np.ma.median(data_ma, axis=0).data
+
+    # Perform optimal extraction on each of the frames
+    iterfn = range(meta.int_start, meta.n_int)
+    if meta.verbose:
+        iterfn = tqdm(iterfn)
+    for n in iterfn:
+        data['optspec'][n], data['opterr'][n], _ = \
+            optimize_one(meta, apdata[n], apmask[n], apbg[n],
+                         data.stdspec[n].values, gain, apv0[n],
+                         p5thresh=meta.p5thresh,
+                         p7thresh=meta.p7thresh,
+                         fittype=meta.fittype,
+                         window_len=meta.window_len,
+                         deg=meta.prof_deg, windowtype=windowtype,
+                         n=n, m=m, meddata=medflux)
+
+    # Mask out NaNs and Infs
+    optspec_ma = np.ma.masked_invalid(data.optspec.values)
+    opterr_ma = np.ma.masked_invalid(data.opterr.values)
+    optmask = np.logical_or(np.ma.getmaskarray(optspec_ma),
+                            np.ma.getmaskarray(opterr_ma)).astype(int)
+    data.optmask.values = optmask
+
+    return data, meta, log
+
+
+def optimize_one(meta, subdata, mask, bg, spectrum, Q, v0, p5thresh=10,
+                 p7thresh=10, fittype='smooth', window_len=21, deg=3,
+                 windowtype='hanning', n=0, m=0, meddata=None):
     '''Extract optimal spectrum with uncertainties.
 
     Parameters
@@ -564,7 +610,7 @@ def optimize(meta, subdata, mask, bg, spectrum, Q, v0, p5thresh=10,
             print("Unknown normalized spatial profile method.")
             return
 
-        if meta.isplots_S3 >= 3:
+        if meta.isplots_S3 >= 3 and n < meta.int_end:
[…]
