+++ src/eureka/S3_data_reduction/source_pos.py
@@ -2,10 +2,12 @@
 
 import numpy as np
 from scipy.optimize import curve_fit
+from tqdm import tqdm
+import multiprocessing as mp
 from . import plots_s3
 
 
-def source_pos(data, meta, m, integ=0):
+def source_pos(data, meta, log, m, integ=0):
     '''Make image+background plot.
 
     Parameters
@@ -16,9 +18,107 @@ def source_pos(data, meta, m, integ=0):
         The metadata object.
     m : int
         The file number.
-    integ : int, optional
+    integ : int or None; optional
+        The integration number. Default is 0 (first integration).
+        If set to None, the source position and width for each frame will be
+        calculated and stored in data.driftypos and data.driftywidth.
+
+    Returns
+    -------
+    src_ypos : int
+        The central position of the star.
+        
+    Notes
+    -----
+    History:
+    
+    - 2022-07-18, Taylor J Bell
+        Changed source_pos to source_pos_one and using this function to
+        allow multiple frames to get source positions in parallel.
+    '''
+    # Mask any clipped values
+    flux = np.ma.masked_where(~data.mask.values, data.flux.values)
+
+    if meta.src_pos_type == 'hst':
+        guess = data.guess.values[0]
+    else:
+        guess = None
+
+    if integ is None:
+        log.writelog("  Recording y position and width for all "
+                     "integrations...", mute=(not meta.verbose))
+        # Get the source position of every frame
+        src_ypos_exact = np.zeros_like(data["time"])
+        src_ypos_width = np.zeros_like(data["time"])
+
+        # Write source_positions
+        def writePos(arg):
+            src_ypos, src_yexact, src_ywidth, n = arg
+            src_ypos_exact[n] = src_yexact
+            src_ypos_width[n] = src_ywidth
+            return
+
+        if meta.ncpu == 1:
+            # Only 1 CPU
+            iterfn = range(meta.int_start, meta.n_int)
+            if meta.verbose:
+                iterfn = tqdm(iterfn)
+            for n in iterfn:
+                writePos(source_pos_one(flux[n], meta, data.attrs['shdr'],
+                                        m, n, False, guess))
+        else:
+            # Multiple CPUs
+            pool = mp.Pool(meta.ncpu)
+            jobs = [pool.apply_async(func=source_pos_one,
+                                     args=(flux[n], meta,
+                                           data.attrs['shdr'], m,
+                                           n, False, guess),
+                                     callback=writePos)
+                    for n in range(meta.int_start, meta.n_int)]
+            pool.close()
+            iterfn = jobs
+            if meta.verbose:
+                iterfn = tqdm(iterfn)
+            for job in iterfn:
+                job.get()
+
+        data['driftypos'] = (['time'], src_ypos_exact)
+        data['driftywidth'] = (['time'], src_ypos_width)
+
+        return data, meta, log
+    else:
+        # Get the source position of frame `integ`
+        log.writelog('  Locating source position...', mute=(not meta.verbose))
+
+        meta.src_ypos = source_pos_one(flux[integ], meta, data.attrs['shdr'],
+                                       m, integ, True, guess)[0]
+
+        log.writelog('    Source position on detector is row '
+                     f'{meta.src_ypos}.', mute=(not meta.verbose))
+
+        return data, meta, log
+
+
[…]
