### 2022-07-19 54092c0
Updating ECF templates and documentation
