### 2022-07-19 92475f8
Updating outdated docstrings
