+++ src/eureka/S3_data_reduction/source_pos.py
@@ -7,7 +7,7 @@ import multiprocessing as mp
 from . import plots_s3
 
 
-def source_pos(data, meta, log, m, integ=0):
+def source_pos_wrapper(data, meta, log, m, integ=0):
     '''Make image+background plot.
 
     Parameters
@@ -39,8 +39,8 @@ def source_pos(data, meta, log, m, integ=0):
     History:
     
     - 2022-07-18, Taylor J Bell
-        Changed source_pos to source_pos_one and using this function to
-        allow multiple frames to get source positions in parallel.
+        Added source_pos_wrapper to allow multiple frames to get
+        source positions in parallel.
     '''
     # Mask any clipped values
     flux = np.ma.masked_where(~data.mask.values, data.flux.values)
@@ -70,12 +70,12 @@ def source_pos(data, meta, log, m, integ=0):
             if meta.verbose:
                 iterfn = tqdm(iterfn)
             for n in iterfn:
-                writePos(source_pos_one(flux[n], meta, data.attrs['shdr'],
-                                        m, n, False, guess))
+                writePos(source_pos(flux[n], meta, data.attrs['shdr'],
+                                    m, n, False, guess))
         else:
             # Multiple CPUs
             pool = mp.Pool(meta.ncpu)
-            jobs = [pool.apply_async(func=source_pos_one,
+            jobs = [pool.apply_async(func=source_pos,
                                      args=(flux[n], meta,
                                            data.attrs['shdr'], m,
                                            n, False, guess),
@@ -96,8 +96,8 @@ def source_pos(data, meta, log, m, integ=0):
         # Get the source position of frame `integ`
         log.writelog('  Locating source position...', mute=(not meta.verbose))
 
-        meta.src_ypos = source_pos_one(flux[integ], meta, data.attrs['shdr'],
-                                       m, integ, True, guess)[0]
+        meta.src_ypos = source_pos(flux[integ], meta, data.attrs['shdr'],
+                                   m, integ, True, guess)[0]
 
         log.writelog('    Source position on detector is row '
                      f'{meta.src_ypos}.', mute=(not meta.verbose))
@@ -105,7 +105,7 @@ def source_pos(data, meta, log, m, integ=0):
         return data, meta, log
 
 
-def source_pos_one(flux, meta, shdr, m, n, plot=True, guess=None):
+def source_pos(flux, meta, shdr, m, n, plot=True, guess=None):
     '''Make image+background plot.
 
     Parameters
@@ -144,7 +144,7 @@ def source_pos_one(flux, meta, shdr, m, n, plot=True, guess=None):
         Enable recording of the width if the source is fitted with a Gaussian
         + add an option to fit any integration (not hardcoded to be the first)
     - 2022-07-18, Taylor J Bell
-        Allowing parallelized code if fitting multiple frames.
+        Tweaked to allow parallelized code if fitting multiple frames.
     '''
     if meta.src_pos_type == 'header':
         if 'SRCYPOS' not in shdr:
