### 2022-07-19 c3fb345
Changing S3 figure numbers and verbosity
