+++ src/eureka/S3_data_reduction/plots_s3.py
@@ -128,7 +128,7 @@ def image_and_background(data, meta, log, m):
 
 
 def drift_2D(data, meta):
-    '''Plot the fitted 2D drift. (Fig 3102)
+    '''Plot the fitted 2D drift. (Fig 3105)
 
     Parameters
     ----------
@@ -137,7 +137,7 @@ def drift_2D(data, meta):
     meta : eureka.lib.readECF.MetaClass
         The metadata object.
     '''
-    plt.figure(3102, figsize=(8, 6))
+    plt.figure(3105, figsize=(8, 6))
     plt.clf()
     plt.subplot(211)
     for p in range(2):
@@ -151,7 +151,7 @@ def drift_2D(data, meta):
     plt.ylabel(f'Drift Along x ({data.drift2D.drift_units})')
     plt.xlabel('Frame Number')
     plt.tight_layout()
-    fname = f'figs{os.sep}fig3102_Drift2D{figure_filetype}'
+    fname = f'figs{os.sep}fig3105_Drift2D{figure_filetype}'
     plt.savefig(meta.outputdir+fname, dpi=300)
     if not meta.hide_plots:
         plt.pause(0.2)
@@ -203,7 +203,7 @@ def optimal_spectrum(data, meta, n, m):
 def source_position(meta, x_dim, pos_max, m, n,
                     isgauss=False, x=None, y=None, popt=None,
                     isFWM=False, y_pixels=None, sum_row=None, y_pos=None):
-    '''Plot source position for MIRI data. (Figs 3303)
+    '''Plot source position for MIRI data. (Figs 3102)
 
     Parameters
     ----------
@@ -247,7 +247,7 @@ def source_position(meta, x_dim, pos_max, m, n,
     - Oct 15, 2021: Taylor Bell
         Tidied up the code a bit to reduce repeated code.
     '''
-    plt.figure(3303)
+    plt.figure(3102)
     plt.clf()
     plt.plot(y_pixels, sum_row, 'o', label='Data')
     if isgauss:
@@ -265,7 +265,7 @@ def source_position(meta, x_dim, pos_max, m, n,
     plt.tight_layout()
     file_number = str(m).zfill(int(np.floor(np.log10(meta.num_data_files))+1))
     int_number = str(n).zfill(int(np.floor(np.log10(meta.n_int))+1))
-    fname = (f'figs{os.sep}fig3303_file{file_number}_int{int_number}' +
+    fname = (f'figs{os.sep}fig3102_file{file_number}_int{int_number}' +
              '_source_pos'+figure_filetype)
     plt.savefig(meta.outputdir+fname, dpi=300)
     if not meta.hide_plots:
@@ -273,7 +273,7 @@ def source_position(meta, x_dim, pos_max, m, n,
 
 
 def profile(meta, profile, submask, n, m):
-    '''Plot weighting profile from optimal spectral extraction routine. (Figs 3304)
+    '''Plot weighting profile from optimal spectral extraction routine. (Figs 3303)
 
     Parameters
     ----------
@@ -300,7 +300,7 @@ def profile(meta, profile, submask, n, m):
     submask = np.ma.masked_where(mask, submask)
     vmax = 0.05*np.ma.max(profile*submask)
     vmin = np.ma.min(profile*submask)
-    plt.figure(3304)
+    plt.figure(3303)
     plt.clf()
     plt.suptitle(f"Profile - Integration {n}")
     plt.imshow(profile*submask, aspect='auto', origin='lower',
@@ -310,7 +310,7 @@ def profile(meta, profile, submask, n, m):
     plt.tight_layout()
[…]
