### 2022-07-19 3f512c5
Missed updating a call to wrapper function in S3
