### 2022-07-19 6912bb6
Limiting the number of CC plots (4301, 4302)
