### 2022-07-19 ef34949
Allow setting nplots to None to plot everything
