### 2022-07-19 08dace5
Updating S2-3 and quickstart figures
