### 2022-07-19 bd6624c
Updating S4-6 plots in outputs.rst
