+++ docs/source/outputs.rst
@@ -84,36 +84,34 @@ Stage 4 Outputs
 ---------------
 
 In Stage 4:
-   - If ``isplots_S4`` = 1: ``Eureka!`` will plot the spectral drift per exposure, and the drift-corrected 2-dimensional lightcurve with extracted bins overlaid.
+   - If ``isplots_S4`` = 1: ``Eureka!`` will plot the spectral drift per exposure, the drift-corrected 2-dimensional lightcurve with extracted bins overlaid, and the 1D light curves.
 
-   .. figure:: ../media/S4_out/fig4101.png
+   .. figure:: ../media/S4_out/fig4101_2D_LC.png
+      :alt: Stage 4 2-dimensional spectrum
+
+      Fig 4101: 2-Dimensional Spectrum
+
+   .. figure:: ../media/S4_out/fig4102_ch0_1D_LC.png
+      :alt: Stage 4 1-dimensional binned spectrum
+
+      Fig 4102: 1-Dimensional Binned Spectrum
+
+   .. figure:: ../media/S4_out/fig4103_DriftXPos.png
       :alt: Stage 4 spectral drift plot
 
-      Stage 4 output: Spectral Drift Plot
+      Fig 4103: Spectral Drift Plot
 
-   .. figure:: ../media/S4_out/fig4201.png
-      :alt: Stage 4 2-dimensional binned spectrum
+   - If ``isplots_S4`` = 3: ``Eureka!`` will plot the cross-correlated reference spectrum with the current spectrum for each integration, and the cross-correlation strength for each integration.
 
-      Stage 4 output: 2-Dimensional Binned Spectrum
-
-   - If ``isplots_S4`` = 3: ``Eureka!`` will plot the spectroscopic lightcurves for each wavelength bin.
-
-   .. figure:: ../media/S4_out/fig4301.png
-      :alt: Stage 4 spectroscopic lightcurve
-
-      Stage 4 output: Spectroscopic Lightcurve
-
-   - If ``isplots_S4`` = 5: ``Eureka!`` will plot the cross-correlated reference spectrum with the current spectrum for each integration, and the cross-correlation strength for each integration.
-
-   .. figure:: ../media/S4_out/fig4401.png
+   .. figure:: ../media/S4_out/fig4301_int00_CC_Spec.png
       :alt: Stage 4 cross correlated reference spectrum
 
-      Stage 4 output: Cross-Correlated Reference Spectrum
+      Fig 4301: Cross-Correlated Reference Spectrum
 
-   .. figure:: ../media/S4_out/fig4501.png
+   .. figure:: ../media/S4_out/fig4302_int00_CC_Vals.png
       :alt: Stage 4 cross correlation strength
 
-      Stage 4 output: Cross-Correlation Strength
+      Fig 4302: Cross-Correlation Strength
 
 
 .. _s5-out:
@@ -122,46 +120,49 @@ Stage 5 Outputs
 ---------------
 
 In Stage 5:
-   - If ``isplots_S5`` = 1: ``Eureka!`` will plot the fitted lightcurve model over the data in each channel.
+   - If ``isplots_S5`` = 1: ``Eureka!`` will plot the fitted lightcurve model over the data in each channel. If fitting with a GP, an additional figure will be made showing the GP component. Finally, an additional plot compares the fits from different fitters.
    
-   .. figure:: ../media/S5_out/fig5101.png
+   .. figure:: ../media/S5_out/fig5101_ch0_lc_emcee.png
       :alt: Stage 5 fit data and lightcurve
 
-      Stage 5 output: Fitted lightcurve
+      Fig 5101: Fitted Lightcurve, Model, and Residual Plot
+
+   .. figure:: ../media/S5_out/fig5102_ch0_lc_GP_emcee.png
[…]
