### 2022-07-19 5889e6c
Protecting against nplots larger than n_ints
