+++ src/eureka/S3_data_reduction/s3_reduce.py
@@ -242,6 +242,9 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
                     meta.int_start = 0
                 if not hasattr(meta, 'nplots') or meta.nplots is None:
                     meta.int_end = meta.n_int
+                elif meta.int_start+meta.nplots > meta.n_int:
+                    # Too many figures requested, so reduce it
+                    meta.int_end = meta.n_int
                 else:
                     meta.int_end = meta.int_start+meta.nplots
 
