### 2022-07-19 aad0053
Making sure eureka.__version__ is always defined
If eureka is run from the git files and not installed, the version.py
file would never have been made
