+++ setup.cfg
@@ -49,6 +49,7 @@ install_requires =
     pandas
     photutils
     scipy >= 1.4.0 # Needed for scipy.fft
+    setuptools_scm
     stdatamodels
     svo_filters
     tqdm
