### 2022-07-19 d35ce10
Fixing ypos and ydrift plotting verbosity levels
