+++ src/eureka/S3_data_reduction/s3_reduce.py
@@ -312,6 +312,10 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
                     data, meta, log = source_pos.source_pos_wrapper(data, meta,
                                                                     log, m,
                                                                     integ=None)
+                    if meta.isplots_S3 >= 1:
+                        # make y position and width plots
+                        plots_s3.driftypos(data, meta)
+                        plots_s3.driftywidth(data, meta)
 
                 # Select only aperture region
                 apdata, aperr, apmask, apbg, apv0 = inst.cut_aperture(data,
@@ -336,10 +340,6 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
                     for n in iterfn:
                         # make optimal spectrum plot
                         plots_s3.optimal_spectrum(data, meta, n, m)
-                    if meta.record_ypos:
-                        # make y position and width plots
-                        plots_s3.driftypos(data, meta)
-                        plots_s3.driftywidth(data, meta)
 
                 if meta.save_output:
                     # Save flux data from current segment
