### 2022-07-19 5295340
Fixing WFC3 get_reference_frames() source_pos call
