### 2022-07-19 95d1e2e
Updating ecf.rst nplots description
