### 2022-07-20 062f7d0
added figure identifiers, changed centroid fitting, fixed squareroot limb darking naming
