+++ src/eureka/lib/imageedit.py
@@ -1,0 +1,268 @@
+# $Author: patricio $
+# $Revision: 301 $
+# $Date: 2010-07-10 03:33:44 -0400 (Sat, 10 Jul 2010) $
+# $HeadURL: file:///home/esp01/svn/code/python/branches/patricio/photpipe/lib/imageedit.py $
+# $Id: imageedit.py 301 2010-07-10 07:33:44Z patricio $
+
+import numpy as np
+"""
+    Name
+    ----
+    Image Edit
+
+    File
+    ----
+    imageedit.py
+
+    Description
+    -----------
+    Routines for editting 2D array images, allows to cut and paste one
+    array into other.
+
+    Package Contents
+    ----------------
+    trimimage(data, (yc,xc), (yr,xr), mask=None, uncd=None, oob=0):
+        Extracts a rectangular area of an image masking out of bound pixels.
+
+    pasteimage(data, subim, (dyc,dxc), (syc,sxc)=(None,None)):
+        Inserts subim array into data, matching the specified coordinates.
+
+    Examples:
+    ---------
+
+    Revisions
+    ---------
+    2010-07-11  patricio  Added documentation.      pcubillos@fulbrightmail.org
+"""
+
+def trimimage(data, c, r, mask=None, uncd=None, oob=0):
+
+  (yc,xc) = c
+  (yr,xr) = r
+
+
+  """
+    Extracts a rectangular area of an image masking out of bound pixels.
+
+    Parameters:
+    ----------
+    data  : 2D ndarray
+            Image from where extract a sub image.
+    yc, xc: 2-elements tuple
+            Position in data, where the extracted image will be centered.
+            yc and xc must have integer values.
+    yr, xr: 2-elements tuple
+            Semi-length of the extracted image. Integer values.
+    mask  : 2D ndarray
+            If specified, this routine will extract the mask subimage
+            as well. out of bound pixels values will have value oob.
+    uncd  : 2D ndarray
+            If specified, this routine will extract the uncd subimage
+            as well.
+    oob   : scalar
+            Value for out of bound pixels in the mask. Default is 0.
+
+    Returns
+    -------
+    Tuple, up to 3 2D-ndarray images containing the extracted image,
+    its mask (if specified) and a uncd array (if specified). The shape
+    of each array is (2*yr+1, 2*xr+1).
+
+    Example:
+    -------
+    >>> from imageedit import *
+    >>> import numpy as np
+
+    >>> # Create  a data image and its mask
+    >>> data  = np.arange(25).reshape(5,5)
+    >>> print(data)
+    [[ 0  1  2  3  4]
+     [ 5  6  7  8  9]
+     [10 11 12 13 14]
+     [15 16 17 18 19]
+     [20 21 22 23 24]]
+    >>> msk  = np.ones(np.shape(data))
+    >>> msk[1:4,2] = 0
+
+    >>> # Extract a subimage centered on (3,1) of shape (3,5)
+    >>> dyc,dxc = 3,1
+    >>> subim, mask = trimimage(data, (dyc,dxc), (1,2), mask=msk)
+    >>> print(subim)
+    [[  0.  10.  11.  12.  13.]
+     [  0.  15.  16.  17.  18.]
+     [  0.  20.  21.  22.  23.]]
+    >>> print(mask)
+    [[ 0.  1.  1.  0.  1.]
+     [ 0.  1.  1.  0.  1.]
+     [ 0.  1.  1.  1.  1.]]
+
[…]
