### 2022-07-20 7ebbccf
updated plots, removed wave_2d for photometry
