+++ src/eureka/lib/apphot.py
@@ -11,587 +11,582 @@ import matplotlib as plt
 from . import interp2d as i2d
 from ..S3_data_reduction import plots_s3
 
-def apphot(iii, mmm, meta, image,  ctr, photap,  skyin, skyout, betahw, targpos,
-           mask = None,   imerr    = None,  skyfrac = 0.0,
-           med  = False,  nochecks = False,
-           expand = 1,    order    = 1,
-           aperr=False,   nappix=False,     skylev=False, skyerr=False,
-           nskypix=False, nskyideal=False,  status=False, isbeta=False, betaper=False):
+
+def apphot(iii, mmm, meta, image, ctr, photap, skyin, skyout, betahw, targpos,
+           mask=None, imerr=None, skyfrac=0.0,
+           med=False, nochecks=False,
+           expand=1, order=1,
+           aperr=False, nappix=False, skylev=False, skyerr=False,
+           nskypix=False, nskyideal=False, status=False, isbeta=False, betaper=False):
+    """
+      Perform aperture photometry on the input image.
+
+      Parameters
+      ----------
+      image    : 2D ndimage
+                 Float array containing object to measure.
+      ctr      : 2 elements tuple
+                 x,y location of object's center.
+      photap   : Scalar
+                 Size of photometry apperture in pixels.
+      skyin    : Scalar
+                 Inner sky annulus edge, in pixels.
+      skyout   : Scalar
+                 Outer sky annulus edge, in pixels.
+
+      betahw   : Scalar
+                 Half-width of box size around centroid for beta calculation.
+
+      targpos  : 2 elements tuple
+                 x,y location of object's center calculated from mean image.
+
+      mask     : 2D ndimage
+                 Byte array giving status of corresponding pixel in
+                 Image: bad pixel=0, good pixel=1.  Default: all pixels
+                 are good. Same shape as image.
+
+      imerr    : 2D ndimage
+                 Error estimate for each pixel in the image.  Suggest
+                 sqrt(image/flat/gain+rdnoise^2), with proper adjustment
+                 if a sky, dark, or bias frame has been removed. Same
+                 shape as image.
+
+      skyfrac  : Scalar
+                 Minimum fraction of sky pixels required to be good.
+                 Must be in range 0 < skyfrac < 1.
+
+      med      : Boolean
+                 If True use median rather than mean in sky level estimation.
+      nochecks : Boolean
+                 Set to True to skip checks of input sanity.
+      expand   : Integer scalar
+                 Positive integer factor by which to blow up image, for
+                 more accurate aperture arithmetic.  If expand=5, each
+                 pixel becomes a 5x5 block of pixels.  If the pixel is
+                 on the edge of an aperture or annulus radius, some of
+                 the 5x5 block will be counted and some will not.
+      order    : Integer scalar
+                 Set order to 0 to do nearest-neighbor interpolation if expand.
+                 Default: 1, bilinear interpolation.
+
+      aperr    : Boolean
+                 Set to True to return flux error.
+
+      nappix   : Boolean
+                 Set to True to return number of total pixels in aperture.
+
+      skylev   : Boolean
+                 Set to True to return the sky level.
+
+      skyerr   : Boolean
+                 Set to True to return the error in the sky level.
+
+      nskypix  : boolean
+                 Set to True to return the number of good pixels in sky annulus.
+
+      nskyideal: Boolean
+                 Set to True to return the number of pixels that should
+                 be in sky annulus.
+
+      status   : Boolean
[…]

[…]
