### 2022-07-20 5737ac0
Cleaned up util.py
