+++ src/eureka/lib/util.py
@@ -93,16 +93,16 @@ def trim(data, meta):
 
 
 def check_nans(data, mask, log, name=''):
-    """Checks where a data array has NaNs or infs.
+    """Checks where a data-like array is invalid (contains NaNs or infs).
 
     Parameters
     ----------
     data : ndarray
-        a data array (e.g. data, err, dq, ...).
+        a data-like array (e.g. data, err, dq, ...).
     mask : ndarray
         Input mask.
     log : logedit.Logedit
-        The open log in which NaNs will be mentioned if existent.
+        The open log in which NaNs/Infs will be mentioned, if existent.
     name : str; optional
         The name of the data array passed in (e.g. SUBDATA, SUBERR, SUBV0).
         Defaults to ''.
@@ -119,7 +119,7 @@ def check_nans(data, mask, log, name=''):
         log.writelog(f"  WARNING: {name} has {num_nans} NaNs/infs. Your "
                      "subregion may be off the edge of the detector "
                      "subarray.\n    Masking NaN region and continuing, "
-                     "but you should really stop and reconsider your"
+                     "but you should really stop and reconsider your "
                      "choices.")
         inan = np.where(np.ma.masked_invalid(data).mask)
         # subdata[inan]  = 0
@@ -342,7 +342,7 @@ def normalize_spectrum(meta, optspec, opterr=None, optmask=None):
     # Normalize the spectrum
     if meta.inst == 'wfc3':
         scandir = np.repeat(meta.scandir, meta.nreads)
-        
+
         for p in range(2):
             iscans = np.where(scandir == p)[0]
             if len(iscans) > 0:
@@ -424,7 +424,7 @@ def get_mad(meta, log, wave_1d, optspec, optmask=None,
             if len(iscans) > 0:
                 mad = np.ma.mean(ediff[iscans])
                 log.writelog(f"Scandir {p} MAD = {int(np.round(mad))} ppm")
-                setattr(meta, f'mad_scandir{p}', mad)   
+                setattr(meta, f'mad_scandir{p}', mad)
 
     return np.ma.mean(ediff)
 
