### 2022-07-20 ad10848
updating plots for photometry
