+++ src/eureka/S3_data_reduction/plots_s3.py
@@ -370,6 +370,9 @@ def phot_bg(meta, data):
     plt.clf()
     plt.suptitle('Photometric background light curve')
     plt.errorbar(data.time, data['skylev'], yerr=data['skyerr'], c='k', fmt='.')
+    plt.ylabel('Flux')
+    plt.xlabel('Time')
+    plt.tight_layout()
     fname = (f'figs{os.sep}fig3305-1D_BG_LC' + figure_filetype)
     plt.savefig(meta.outputdir+fname, dpi=300)
     if not meta.hide_plots:
@@ -385,21 +388,21 @@ def phot_centroid(meta, data):
     plt.suptitle('Centroid positions over time')
 
     plt.subplot(411)
-    plt.plot(range(len(data.centroid_x)), data.centroid_x-np.mean(data.centroid_x))
+    plt.plot(data.time, data.centroid_x-np.mean(data.centroid_x))
     plt.ylabel('Delta x')
 
     plt.subplot(412)
-    plt.plot(range(len(data.centroid_y)), data.centroid_y-np.mean(data.centroid_y))
+    plt.plot(data.time, data.centroid_y-np.mean(data.centroid_y))
     plt.ylabel('Delta y')
 
     plt.subplot(413)
-    plt.plot(range(len(data.centroid_sx)), data.centroid_sy-np.mean(data.centroid_sx))
+    plt.plot(data.time, data.centroid_sy-np.mean(data.centroid_sx))
     plt.ylabel('Delta sx')
 
     plt.subplot(414)
-    plt.plot(range(len(data.centroid_sy)), data.centroid_sy-np.mean(data.centroid_sy))
+    plt.plot(data.time, data.centroid_sy-np.mean(data.centroid_sy))
     plt.ylabel('Delta sy')
-
+    plt.xlabel('Time')
     plt.tight_layout()
     fname = (f'figs{os.sep}fig3306-Centroid' + figure_filetype)
     plt.savefig(meta.outputdir + fname, dpi=250)
@@ -432,48 +435,38 @@ def phot_centroid_fgc(img, x, y, sx, sy, i, m, meta):
         plt.pause(0.2)
 
 
-def phot_2D_frame(meta, m, i, data, image, targpos, skyann, apmask):
+def phot_2D_frame(meta, m, i, data):
     """
     Plots the 2D frame together with the centroid position and apertures.
     """
     plt.figure(3307)
     plt.clf()
-    plt.suptitle('2D frame with Centroid and apertures')
+    plt.suptitle('2D frame with centroid and apertures')
     flux, centroid_x, centroid_y = data.flux[i], data.centroid_x[i], data.centroid_y[i]
-    fig, ax = plt.subplots()
-    ax.imshow(flux, vmax=5e3, origin='lower')
-    ax.scatter(centroid_x, centroid_y, marker='x', s=25, c='r')
-
+    plt.imshow(flux, vmax=5e3, origin='lower')
+    plt.scatter(centroid_x, centroid_y, marker='x', s=25, c='r', label='centroid')
+    #alphas = np.zeros(image.shape)
+    #alphas[np.where(skyann == True)] = 0.9
+    #plt.imshow(~skyann, origin='lower', alpha=alphas, cmap='magma')
+    #alphas = np.zeros(image.shape)
+    #alphas[np.where(apmask == True)] = 0.4
+    circle1 = plt.Circle((centroid_x, centroid_y), meta.photap, color='r', fill=False, lw=3, alpha=0.7, label='target aperture')
+    circle2 = plt.Circle((centroid_x, centroid_y), meta.skyin, color='w', fill=False, lw=4, alpha=0.8, label='sky aperture')
[…]
