+++ src/eureka/lib/centerdriver.py
@@ -87,7 +87,7 @@ def centerdriver(method, data, guess, trim, radius, size,
         sy, sx, y, x = g.fitgaussian(img, yxguess=loc, mask=msk, weights=weights,
                                      fitbg=fitbg, maskg=maskstar)[0][0:4]
         extra = sy, sx  # Gaussian 1-sigma half-widths
-        if meta.isplots_S3 >= 5:
+        if meta.isplots_S3 >= 5 and len(img)==7:
             plots_s3.phot_centroid_fgc(img, x, y, sx, sy, i, m, meta)
 
     elif method == 'col':

[…]
