+++ src/eureka/S4_generate_lightcurves/s4_genLC.py
@@ -20,7 +20,7 @@ import numpy as np
 import scipy.interpolate as spi
 import astraeus.xarrayIO as xrio
 from astropy.convolution import Box1DKernel
-from . import plots_s4, drift, wfc3
+from . import plots_s4, drift, generate_LD, wfc3
 from ..lib import logedit
 from ..lib import readECF
 from ..lib import manageevent as me
@@ -44,6 +44,10 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
 
     Returns
     -------
+    spec : Astreaus object 
+        Data object of wavelength-like arrrays.
+    lc : Astreaus object 
+        Data object of time-like arrrays (light curve).
     meta : eureka.lib.readECF.MetaClass
         The metadata object with attributes added by S4.
 
@@ -57,6 +61,11 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
         Updated to allow for inputs from new S3
     - April 2022 Kevin Stevenson
         Enabled Astraeus
+    - July 2022 Caroline Piaulet
+        Recording of x (computed in S4) and y (computed in S3) pos drifts and 
+        widths in Spec and LC objects
+    - July 2022 Sebastian Zieba
+         Added photometry S4
     '''
     # Load Eureka! control file and store values in Event object
     ecffile = 'S4_' + eventlabel + '.ecf'
@@ -225,12 +234,13 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
             # Do 1D sigma clipping (along time axis) on unbinned spectra
             if meta.sigma_clip:
                 log.writelog('Sigma clipping unbinned optimal spectra along '
-                             'time axis')
+                             'time axis...')
                 outliers = 0
                 for w in range(meta.subnx):
                     spec.optspec[:, w], spec.optmask[:, w], nout = \
                         clipping.clip_outliers(spec.optspec[:, w], log,
-                                               spec.wave_1d[w],
+                                               spec.wave_1d[w].values,
+                                               spec.wave_1d.wave_units,
                                                mask=spec.optmask[:, w],
                                                sigma=meta.sigma,
                                                box_width=meta.box_width,
@@ -246,44 +256,61 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
                              f'wavelength',
                              mute=meta.verbose)
 
-            # Apply 1D drift/jitter correction
-            if meta.correctDrift:
+            if hasattr(meta, 'record_ypos') and meta.record_ypos:
+                lc['driftypos'] = (['time'], spec.driftypos.data)
+                lc['driftywidth'] = (['time'], spec.driftywidth.data)
+            
+            # Record and correct for 1D drift/jitter
+            if meta.recordDrift or meta.correctDrift:
                 # Calculate drift over all frames and non-destructive reads
                 # This can take a long time, so always print this message
-                log.writelog('Applying drift/jitter correction')
+                log.writelog('Computing drift/jitter')
                 # Compute drift/jitter
-                drift1d, driftmask = drift.spec1D(spec.optspec, meta, log,
-                                                  mask=spec.optmask)
+                drift_results = drift.spec1D(spec.optspec, meta, log,
+                                             mask=spec.optmask)
+                drift1d, driftwidth, driftmask = drift_results
                 # Replace masked points with moving mean
                 drift1d = clipping.replace_moving_mean(
                     drift1d, driftmask, Box1DKernel(meta.box_width))
-                lc['drift1d'] = (['time'], drift1d)
+                driftwidth = clipping.replace_moving_mean(
+                    driftwidth, driftmask, Box1DKernel(meta.box_width))
+                lc['driftxpos'] = (['time'], drift1d)
[…]
