### 2022-07-20 kevin218/Eureka#415 @kcrasmus
Errors on installation
### Instrument

_No response_

### What happened?

Tried installing on a clean conda environments with

git clone -b v0.3 https://github.com/kevin218/Eureka.git
cd Eureka
pip install -e '.[jwst]'

getting the following, very long series of warnings/errors:

after that, if I try to run run_eureka.py, it fails with "No module named Eureka"

### Error traceback output

[error.txt](https://github.com/kevin218/Eureka/files/9152704/error.txt)

### What operating system are you using?

Mac OSx

### What version of Python are you running?

3.9.7

### What Python packages do you have installed?

i am trying to install in a brand new conda environment

### Code of Conduct

- [X] I agree to follow this project's Code of Conduct
