### 2022-07-20 2d2e2ef
Fix standard deviation calculation in medstddev by adding axis, also simplify calc
