### 2022-07-20 kevin218/Eureka#415 @taylorbell57
Can you please confirm that running all of the following steps exactly does not work for you:
```
conda create -n eureka python==3.9.7
conda activate eureka
git clone -b v0.3 https://github.com/kevin218/Eureka.git
cd Eureka
pip install -e '.[jwst]'
```

That works for me, and I'm running MacOS 11.6.7 with an Intel CPU. I also tried running the pip install command using the additional `--no-cache-dir` flag to make sure I was downloading all the dependencies from scratch. It seems possible to me based on your txt file that you may not have run the `conda activate eureka` command which may or may not have caused an issue. Can you also confirm you have clang installed using `which clang` in the terminal?
