### 2022-07-20 kevin218/Eureka#415 @taylorbell57
That is very strange, some potential things to try (one at a time):

1. Try leaving out the `-e` when doing the pip install command
2. Try adding `--no-cache-dir` when doing the pip install command to see if your cached installations are broken somehow
3. Try doing `git checkout pyproject_toml` inside that Eureka folder and then doing the pip install commands to see if we just need to add numpy as a dependency to the `pyproject.toml` file.

Related to that last option, your error.txt file did not contain any such "ERROR: Could not build wheels for numpy, which is required to install pyproject.toml-based projects" error. If you're getting a different error message now than you did before (as it seems), can you copy-paste it here (preferrable to attaching a txt file) so that I can investigate more
