### 2022-07-20 kevin218/Eureka#415 @taylorbell57
Can you also confirm your Mac has an Intel processor?
