### 2022-07-20 kevin218/Eureka#415 @kcrasmus
Hello,

    Yeah, h5py is the issue. I found this homebrew workaround—maybe you
could try setting that particular command to pip install h5py=3.2? or do
you know where in the code I could change it myself?

thanks,
kaitlin

On Wed, Jul 20, 2022 at 2:33 PM Taylor James Bell ***@***.***>
wrote:

> Ahh, we've had others report issues with M1 chips. To really simplify the
> test, try
>
> pip install h5py<3.2
>
> to see if it really is just that package fighting with M1.
>
> You can also try removing that version upper limit - I feel like that
> could potentially be outdated (I can't remember why we added the limit and
> didn't write it down anywhere)
>
