### 2022-07-20 kevin218/Eureka#415 @kcrasmus
https://stackoverflow.com/questions/66741778/how-to-install-h5py-needed-for-keras-on-macos-with-m1

this works for me, but pip install h5py=3.2 does not.

kaitlin
