### 2022-07-20 kevin218/Eureka#415 @taylorbell57
You can edit the requirements in the setup.cfg file, but beware that something else may break if h5py isn't below 3.2 - I can't say for sure one way or the other
