### 2022-07-20 kevin218/Eureka#415 @kcrasmus
that one does not work for a different reason—it can't find exotic-ld.

kaitlin

On Wed, Jul 20, 2022 at 2:56 PM Taylor James Bell ***@***.***>
wrote:

> You could also try following our conda environment.yml installation method
> and see if that works for you
>
