### 2022-07-20 kevin218/Eureka#415 @taylorbell57
Ah interesting, that might be a much easier problem to solve - let me try something out in ~30 minutes and I'll see if I can solve that environment.yml method issue
