### 2022-07-20 kevin218/Eureka#415 @taylorbell57
Actually, can you just try moving exotic-ld from near the top of the environment.yml file to nearer the bottom underneath the `pip:` line
