### 2022-07-20 84a518a
Fixing exotic-ld installation in environment.yml
exotic-ld is not hosted on conda, so it must be installed using pip
