### 2022-07-20 kevin218/Eureka#415 @taylorbell57
Great, I'm glad to hear that! Let me know whether or not you end up getting this other error that was previously reported for M1 computers: #367
