+++ src/citations.py
@@ -1,0 +1,23 @@
+CITATIONS = {
+    "eureka": (
+        ("eureka"), 
+        r"""@ARTICLE{eureka,
+                author = {{Bell}, Taylor J. and {Ahrer}, Eva-Maria and {Brande}, Jonathan and {Carter}, Aarynn L. and 
+                {Feinstein}, Adina D. and {Guzman Caloca}, Giannina and {Mansfield}, Megan and {Zieba}, Sebastian and 
+                {Piaulet}, Caroline and {Benneke}, Bj{\"o}rn and {Filippazzo}, Joseph and {May}, Erin M. and 
+                {Roy}, Pierre-Alexis and {Kreidberg}, Laura and {Stevenson}, Kevin B.},
+                    title = "{Eureka!: An End-to-End Pipeline for JWST Time-Series Observations}",
+                journal = {arXiv e-prints},
+                keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics},
+                    year = 2022,
+                    month = jul,
+                    eid = {arXiv:2207.03585},
+                    pages = {arXiv:2207.03585},
+            archivePrefix = {arXiv},
+                eprint = {2207.03585},
+            primaryClass = {astro-ph.IM},
+                adsurl = {https://ui.adsabs.harvard.edu/abs/2022arXiv220703585B},
+                adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+            }"""
+    ),
+}
