### 2022-07-21 3dd702d
Allowing unbinned spectra from S4
