### 2022-07-21 cac1e2f
Fixing nspecchan=None behaviour
This will now allow unbinned spectra out of S4
