### 2022-07-21 5dac005
Fixing bin lines for fig 4101
They'll now only be plotted if there is more than one bin and not one
bin per pixel
