+++ src/citations.py
@@ -1,7 +1,7 @@
 CITATIONS = {
     "eureka": (
         ("eureka"), 
-        r"""@ARTICLE{eureka,
+        [r"""@ARTICLE{eureka,
                 author = {{Bell}, Taylor J. and {Ahrer}, Eva-Maria and {Brande}, Jonathan and {Carter}, Aarynn L. and 
                 {Feinstein}, Adina D. and {Guzman Caloca}, Giannina and {Mansfield}, Megan and {Zieba}, Sebastian and 
                 {Piaulet}, Caroline and {Benneke}, Bj{\"o}rn and {Filippazzo}, Joseph and {May}, Erin M. and 
@@ -18,6 +18,155 @@ CITATIONS = {
             primaryClass = {astro-ph.IM},
                 adsurl = {https://ui.adsabs.harvard.edu/abs/2022arXiv220703585B},
                 adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-            }"""
+            }"""]
+    ),
+    "numpy": (
+        ("numpy"),
+        [r"""@Article{numpy,
+            title         = {Array programming with {NumPy}},
+            author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+                            van der Walt and Ralf Gommers and Pauli Virtanen and David
+                            Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+                            Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+                            and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+                            Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+                            R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                            G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+                            Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+                            Travis E. Oliphant},
+            year          = {2020},
+            month         = sep,
+            journal       = {Nature},
+            volume        = {585},
+            number        = {7825},
+            pages         = {357--362},
+            doi           = {10.1038/s41586-020-2649-2},
+            publisher     = {Springer Science and Business Media {LLC}},
+            url           = {https://doi.org/10.1038/s41586-020-2649-2}
+        }"""]
+    ),
+    "matplotlib": (
+        ("matplotlib"),
+        [r"""@Article{matplotlib,
+            Author    = {Hunter, J. D.},
+            Title     = {Matplotlib: A 2D graphics environment},
+            Journal   = {Computing in Science \& Engineering},
+            Volume    = {9},
+            Number    = {3},
+            Pages     = {90--95},
+            abstract  = {Matplotlib is a 2D graphics package used for Python for
+            application development, interactive scripting, and publication-quality
+            image generation across user interfaces and operating systems.},
+            publisher = {IEEE COMPUTER SOC},
+            doi       = {10.1109/MCSE.2007.55},
+            year      = 2007
+            }"""]
+    ),
+    "scipy": (
+        ("scipy"),
+        [r"""@ARTICLE{scipy,
+            author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
[…]
