### 2022-07-22 c48c150
Allowing manual removal of data in S4
