### 2022-07-23 #4503 @catrionamurray
A question about the conflict of interest policy: I haven't directly worked with any of the authors on the Eureka! paper but I am a coauthor on two papers with Adina Feinstein (and potentially a few more people in the author list). I wasn't sure if I need to disclose this, or if it's an issue?
