### 2022-07-23 a2c16ce
basic code to check imported packages for citations
