+++ src/eureka/S1_detector_processing/s1_process.py
@@ -7,6 +7,9 @@ from jwst.pipeline.calwebb_detector1 import Detector1Pipeline
 
 from eureka.S1_detector_processing.ramp_fitting import Eureka_RampFitStep
 
+import sys
+from citations import CITATIONS
+
 from ..lib import logedit, util
 from ..lib import manageevent as me
 from ..lib import readECF
@@ -96,6 +99,17 @@ def rampfitJWST(eventlabel, ecf_path=None):
     # Calculate total run time
     total = (time_pkg.time() - t0) / 60.
     log.writelog('\nTotal time (min): ' + str(np.round(total, 2)))
+
+    # Store citations to relevant dependencies in the meta file
+    # get currently imported modules (top level only)
+    mods = np.unique([mod.split('.')[0] for mod in sys.modules.keys()])
+    
+    # get modules for which we have citations
+    citemods = np.intersect1d(mods, list(CITATIONS))
+
+    # concatenate the bibitems into a bibliography
+    meta.citations = citemods
+    meta.bibliography = np.concatenate([CITATIONS[entry][1] for entry in citemods])
 
     # Save results
     if not meta.testing_S1:
