+++ src/eureka/S4_generate_lightcurves/s4_genLC.py
@@ -27,6 +27,8 @@ from ..lib import manageevent as me
 from ..lib import util
 from ..lib import clipping
 
+import sys
+from citations import CITATIONS
 
 def genlc(eventlabel, ecf_path=None, s3_meta=None):
     '''Compute photometric flux over specified range of wavelengths.
@@ -96,6 +98,25 @@ def genlc(eventlabel, ecf_path=None, s3_meta=None):
         for bg_hw_val in meta.bg_hw_range:
             meta.run_s4 = util.makedirectory(meta, 'S4', meta.run_s4,
                                              ap=spec_hw_val, bg=bg_hw_val)
+
+    # Store citations to relevant dependencies in the meta file
+    # get currently imported modules (top level only)
+    mods = np.unique([mod.split('.')[0] for mod in sys.modules.keys()])
+            
+    # get modules for which we have citations
+    citemods = np.intersect1d(mods, list(CITATIONS))
+
+    # check if meta has existing list of citations/bibitems, if it does, make sure we have imports from previous stages in our citations
+    try: 
+        hasattr(meta, 'citations')
+        citemods = np.union1d(citemods, meta.citations)
+    except AttributeError:
+        # otherwise, if we're starting from this stage, just use the current list of imports
+        pass
+            
+    # save list of imports and the bibliography to the meta object
+    meta.citations = citemods
+    meta.bibliography = np.concatenate([CITATIONS[entry][1] for entry in citemods])
 
     for spec_hw_val in meta.spec_hw_range:
         for bg_hw_val in meta.bg_hw_range:
