### 2022-07-23 #4503 @kevin218
@catrionamurray, @christinahedges — What's your timeline for providing feedback?  The first ERS paper will likely be ready within a week and it would be great to use the JOSS reference.
