### 2022-07-23 a3d7565
Adjust plot borders
