### 2022-07-24 59b0a16
Convert sigma_clip to clip_unbinned and clip_binned
