### 2022-07-24 d591508
Missed a sigma_clip
