### 2022-07-24 ab8689e
update exotic-ld custom throughput file
