### 2022-07-25 #4503 @dfm
> A question about the conflict of interest policy: I haven't directly worked with any of the authors on the Eureka! paper but I am a coauthor on two papers with Adina Feinstein (and potentially a few more people in the author list). I wasn't sure if this is an issue?

Thanks for bringing this up, @catrionamurray! I believe that this potential conflict is sufficiently minor that, as long as you're confident that you can provide a fair review, I'm happy to waive this apparent conflict of interest. If anyone involved feels that this in inappropriate, please feel free to comment here, or contact me directly via email (thanks @kevin218 for your confirmation already!). Otherwise, I think we should proceed with the review as planned. Thanks again!!
