+++ docs/source/ecf.rst
@@ -496,7 +496,7 @@ Used by exotic-ld if compute_ld=True. The surface gravity in log g.
 
 exotic_ld_direc
 '''''''''''''''
-Used by exotic-ld if compute_ld=True. The fully qualified path to the directory for ancillary files for exotic-ld.
+Used by exotic-ld if compute_ld=True. The fully qualified path to the directory for ancillary files for exotic-ld, download at https://zenodo.org/record/6344946#.Yt66YnaZND8.
 
 exotic_ld_grid
 ''''''''''''''
@@ -504,7 +504,7 @@ Used by exotic-ld if compute_ld=True. 1D or 3D model grid.
 
 exotic_ld_file
 ''''''''''''''
-Used by exotic-ld as throughput input file. If none, exotic-ld uses throughput from ancillary files.
+Used by exotic-ld as throughput input file. If none, exotic-ld uses throughput from ancillary files. Make sure that wavelength is given in Angstrom!
 
 isplots_S4
 ''''''''''
