### 2022-07-25 8506278
add error when .flx ancillary files aren't found
