+++ src/eureka/S4_generate_lightcurves/generate_LD.py
@@ -3,6 +3,7 @@ import numpy as np
 import pandas as pd
 import glob
 
+
 def exotic_ld(meta, spec, log):
     '''Generate limb-darkening coefficients using the exotic_ld package.
 
