### 2022-07-25 kevin218/Eureka#421 @catrionamurray
Installation Error
### Instrument

Other (any stage)

### What happened?

Part of JOSS review (https://github.com/openjournals/joss-reviews/issues/4503). Error when trying to install Eureka following the installation guide in the docs.

### Error traceback output

(eureka) camu5866@CatMsMBP Eureka % pip install -e '.[jwst]'
Obtaining file:///Users/camu5866/Documents/Postdoc/Eureka/Eureka
  Installing build dependencies ... done
  Checking if build backend supports build_editable ... done
  Preparing metadata (pyproject.toml) ... done
Collecting astraeus@ git+https://github.com/kevin218/Astraeus@main
  Cloning https://github.com/kevin218/Astraeus (to revision main) to /private/var/folders/zp/t1q7mq8j2vs68_3mbjscd9qw0000gp/T/pip-install-b4iiuzj9/astraeus_d1b6db9239eb4cb186b23c23aa2ac22d
  Running command git clone --filter=blob:none --quiet https://github.com/kevin218/Astraeus /private/var/folders/zp/t1q7mq8j2vs68_3mbjscd9qw0000gp/T/pip-install-b4iiuzj9/astraeus_d1b6db9239eb4cb186b23c23aa2ac22d
  Resolved https://github.com/kevin218/Astraeus to commit 20373ac916fbfdcb85a4f23dfd0690a21dc2771f
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting astroquery
  Using cached astroquery-0.4.6-py3-none-any.whl (4.5 MB)
Collecting photutils
  Downloading photutils-1.5.0-cp39-cp39-macosx_11_0_arm64.whl (566 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 566.9/566.9 kB 7.2 MB/s eta 0:00:00
Collecting exotic-ld
  Using cached exotic_ld-2.0.0-py3-none-any.whl
Collecting celerite
  Using cached celerite-0.4.2-cp39-cp39-macosx_10_9_universal2.whl (579 kB)
Collecting batman-package
  Using cached batman_package-2.4.9-cp39-cp39-macosx_11_0_arm64.whl
Collecting dynesty>1.0
  Using cached dynesty-1.2.3-py2.py3-none-any.whl (94 kB)
Collecting matplotlib
  Using cached matplotlib-3.5.2-cp39-cp39-macosx_11_0_arm64.whl (7.2 MB)
Collecting crds
  Using cached crds-11.16.6-py3-none-any.whl (1.2 MB)
Collecting tqdm
  Using cached tqdm-4.64.0-py2.py3-none-any.whl (78 kB)
Collecting stdatamodels
  Using cached stdatamodels-0.4.3-py3-none-any.whl (34 kB)
Collecting astropy
  Using cached astropy-5.1-cp39-cp39-macosx_11_0_arm64.whl (6.7 MB)
Collecting svo-filters
  Using cached svo_filters-0.4.3-py2.py3-none-any.whl (1.4 MB)
Collecting emcee>3.0.0
  Using cached emcee-3.1.2-py2.py3-none-any.whl (46 kB)
Collecting pandas
  Using cached pandas-1.4.3-cp39-cp39-macosx_11_0_arm64.whl (10.5 MB)
Collecting corner
  Using cached corner-2.2.1-py3-none-any.whl (15 kB)
Collecting lmfit
  Using cached lmfit-1.0.3.tar.gz (292 kB)
  Preparing metadata (setup.py) ... done
Requirement already satisfied: numpy>=1.20.0 in /Users/camu5866/opt/anaconda3/envs/eureka/lib/python3.9/site-packages (from eureka==0.3) (1.23.1)
Collecting bokeh
  Using cached bokeh-2.4.3-py3-none-any.whl (18.5 MB)
Collecting cython
  Using cached Cython-0.29.30-py2.py3-none-any.whl (985 kB)
Collecting george
  Using cached george-0.4.0.tar.gz (1.0 MB)
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting dask<2022.6.1
  Downloading dask-2022.6.0-py3-none-any.whl (1.1 MB)
Collecting ccdproc
  Downloading ccdproc-2.3.1-py3-none-any.whl (355 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 355.3/355.3 kB 12.8 MB/s eta 0:00:00
Collecting h5py<3.2
  Using cached h5py-3.1.0.tar.gz (371 kB)
  Installing build dependencies ... done
  Installing backend dependencies ... error
  error: subprocess-exited-with-error

  × pip subprocess to install backend dependencies did not run successfully.
  │ exit code: 1
  ╰─> [630 lines of output]
      Collecting pkgconfig
        Using cached pkgconfig-1.5.5-py3-none-any.whl (6.7 kB)
      Collecting numpy==1.19.3
        Using cached numpy-1.19.3.zip (7.3 MB)
        Installing build dependencies: started
        Installing build dependencies: finished with status 'done'
        Getting requirements to build wheel: started
        Getting requirements to build wheel: finished with status 'done'
        Preparing metadata (pyproject.toml): started
        Preparing metadata (pyproject.toml): finished with status 'done'
      Collecting Cython>=0.29.14
        Using cached Cython-0.29.30-py2.py3-none-any.whl (985 kB)
      Building wheels for collected packages: numpy
        Building wheel for numpy (pyproject.toml): started
        Building wheel for numpy (pyproject.toml): finished with status 'error'
        error: subprocess-exited-with-error

        × Building wheel for numpy (pyproject.toml) did not run successfully.
        │ exit code: 1
        ╰─> [604 lines of output]
            Running from numpy source directory.
            numpy/random/_bounded_integers.pxd.in has not changed
            numpy/random/_philox.pyx has not changed
            numpy/random/_bounded_integers.pyx.in has not changed
            numpy/random/_sfc64.pyx has not changed
            numpy/random/_mt19937.pyx has not changed
            numpy/random/bit_generator.pyx has not changed
            Processing numpy/random/_bounded_integers.pyx
            numpy/random/mtrand.pyx has not changed
            numpy/random/_generator.pyx has not changed
            numpy/random/_pcg64.pyx has not changed
            numpy/random/_common.pyx has not changed
            Cythonizing sources
            blas_opt_info:
            blas_mkl_info:
            customize UnixCCompiler
              libraries mkl_rt not found in ['/Users/camu5866/opt/anaconda3/envs/eureka/lib', '/usr/lib']
              NOT AVAILABLE

            blis_info:
              libraries blis not found in ['/Users/camu5866/opt/anaconda3/envs/eureka/lib', '/usr/lib']
              NOT AVAILABLE

            openblas_info:
              libraries openblas not found in ['/Users/camu5866/opt/anaconda3/envs/eureka/lib', '/usr/lib']
              NOT AVAILABLE

            atlas_3_10_blas_threads_info:
            Setting PTATLAS=ATLAS
              libraries tatlas not found in ['/Users/camu5866/opt/anaconda3/envs/eureka/lib', '/usr/lib']
              NOT AVAILABLE

            atlas_3_10_blas_info:
              libraries satlas not found in ['/Users/camu5866/opt/anaconda3/envs/eureka/lib', '/usr/lib']
              NOT AVAILABLE

            atlas_blas_threads_info:
            Setting PTATLAS=ATLAS
              libraries ptf77blas,ptcblas,atlas not found in ['/Users/camu5866/opt/anaconda3/envs/eureka/lib', '/usr/lib']
              NOT AVAILABLE

            atlas_blas_info:
              libraries f77blas,cblas,atlas not found in ['/Users/camu5866/opt/anaconda3/envs/eureka/lib', '/usr/lib']
              NOT AVAILABLE

            accelerate_info:
              libraries accelerate not found in ['/Users/camu5866/opt/anaconda3/envs/eureka/lib', '/usr/lib']
            Library accelerate was not found. Ignoring
              libraries veclib not found in ['/Users/camu5866/opt/anaconda3/envs/eureka/lib', '/usr/lib']
            Library veclib was not found. Ignoring
              FOUND:
                extra_compile_args = ['-faltivec', '-I/System/Library/Frameworks/vecLib.framework/Headers']
                extra_link_args = ['-Wl,-framework', '-Wl,Accelerate']
                define_macros = [('NO_ATLAS_INFO', 3), ('HAVE_CBLAS', None)]

              FOUND:
                extra_compile_args = ['-faltivec', '-I/System/Library/Frameworks/vecLib.framework/Headers']
                extra_link_args = ['-Wl,-framework', '-Wl,Accelerate']
                define_macros = [('NO_ATLAS_INFO', 3), ('HAVE_CBLAS', None)]

            non-existing path in 'numpy/distutils': 'site.cfg'
            lapack_opt_info:
            lapack_mkl_info:
              libraries mkl_rt not found in ['/Users/camu5866/opt/anaconda3/envs/eureka/lib', '/usr/lib']
              NOT AVAILABLE

            openblas_lapack_info:
              libraries openblas not found in ['/Users/camu5866/opt/anaconda3/envs/eureka/lib', '/usr/lib']
              NOT AVAILABLE

            openblas_clapack_info:
              libraries openblas,lapack not found in ['/Users/camu5866/opt/anaconda3/envs/eureka/lib', '/usr/lib']
              NOT AVAILABLE

            flame_info:
              libraries flame not found in ['/Users/camu5866/opt/anaconda3/envs/eureka/lib', '/usr/lib']
              NOT AVAILABLE

            atlas_3_10_threads_info:
            Setting PTATLAS=ATLAS
              libraries lapack_atlas not found in /Users/camu5866/opt/anaconda3/envs/eureka/lib
              libraries tatlas,tatlas not found in /Users/camu5866/opt/anaconda3/envs/eureka/lib
              libraries lapack_atlas not found in /usr/lib
              libraries tatlas,tatlas not found in /usr/lib
            <class 'numpy.distutils.system_info.atlas_3_10_threads_info'>
              NOT AVAILABLE

            atlas_3_10_info:
              libraries lapack_atlas not found in /Users/camu5866/opt/anaconda3/envs/eureka/lib
              libraries satlas,satlas not found in /Users/camu5866/opt/anaconda3/envs/eureka/lib
              libraries lapack_atlas not found in /usr/lib
              libraries satlas,satlas not found in /usr/lib
            <class 'numpy.distutils.system_info.atlas_3_10_info'>
              NOT AVAILABLE

            atlas_threads_info:
            Setting PTATLAS=ATLAS
              libraries lapack_atlas not found in /Users/camu5866/opt/anaconda3/envs/eureka/lib
              libraries ptf77blas,ptcblas,atlas not found in /Users/camu5866/opt/anaconda3/envs/eureka/lib
              libraries lapack_atlas not found in /usr/lib
              libraries ptf77blas,ptcblas,atlas not found in /usr/lib
            <class 'numpy.distutils.system_info.atlas_threads_info'>
              NOT AVAILABLE

            atlas_info:
              libraries lapack_atlas not found in /Users/camu5866/opt/anaconda3/envs/eureka/lib
              libraries f77blas,cblas,atlas not found in /Users/camu5866/opt/anaconda3/envs/eureka/lib
              libraries lapack_atlas not found in /usr/lib
              libraries f77blas,cblas,atlas not found in /usr/lib
            <class 'numpy.distutils.system_info.atlas_info'>
              NOT AVAILABLE

              FOUND:
                extra_compile_args = ['-faltivec', '-I/System/Library/Frameworks/vecLib.framework/Headers']
                extra_link_args = ['-Wl,-framework', '-Wl,Accelerate']
                define_macros = [('NO_ATLAS_INFO', 3), ('HAVE_CBLAS', None)]

            /private/var/folders/zp/t1q7mq8j2vs68_3mbjscd9qw0000gp/T/pip-build-env-uctfku0y/overlay/lib/python3.9/site-packages/setuptools/_distutils/dist.py:275: UserWarning: Unknown distribution option: 'define_macros'
              warnings.warn(msg)
            running bdist_wheel
            running build
            running config_cc
            unifing config_cc, config, build_clib, build_ext, build commands --compiler options
            running config_fc
            unifing config_fc, config, build_clib, build_ext, build commands --fcompiler options
            running build_src
            build_src
            building py_modules sources
            building library "npymath" sources
            Could not locate executable gfortran
            Could not locate executable f95
            Could not locate executable f90
            Could not locate executable f77
            Could not locate executable xlf90
            Could not locate executable xlf
            Could not locate executable ifort
            Could not locate executable ifc
            Could not locate executable g77
            Could not locate executable g95
            Could not locate executable pgfortran
            don't know how to compile Fortran code on platform 'posix'
              adding 'build/src.macosx-11.1-arm64-3.9/numpy/core/src/npymath' to include_dirs.
            None - nothing done with h_files = ['build/src.macosx-11.1-arm64-3.9/numpy/core/src/npymath/npy_math_internal.h']
            building library "npysort" sources
              adding 'build/src.macosx-11.1-arm64-3.9/numpy/core/src/common' to include_dirs.
            None - nothing done with h_files = ['build/src.macosx-11.1-arm64-3.9/numpy/core/src/common/npy_sort.h', 'build/src.macosx-11.1-arm64-3.9/numpy/core/src/common/npy_partition.h', 'build/src.macosx-11.1-arm64-3.9/numpy/core/src/common/npy_binsearch.h']
            building library "npyrandom" sources
            building extension "numpy.core._multiarray_tests" sources
            building extension "numpy.core._multiarray_umath" sources
              adding 'build/src.macosx-11.1-arm64-3.9/numpy/core/src/umath' to include_dirs.
              adding 'build/src.macosx-11.1-arm64-3.9/numpy/core/src/npymath' to include_dirs.
              adding 'build/src.macosx-11.1-arm64-3.9/numpy/core/src/common' to include_dirs.
            numpy.core - nothing done with h_files = ['build/src.macosx-11.1-arm64-3.9/numpy/core/src/umath/funcs.inc', 'build/src.macosx-11.1-arm64-3.9/numpy/core/src/umath/simd.inc', 'build/src.macosx-11.1-arm64-3.9/numpy/core/src/umath/loops.h', 'build/src.macosx-11.1-arm64-3.9/numpy/core/src/umath/matmul.h', 'build/src.macosx-11.1-arm64-3.9/numpy/core/src/umath/clip.h', 'build/src.macosx-11.1-arm64-3.9/numpy/core/src/npymath/npy_math_internal.h', 'build/src.macosx-11.1-arm64-3.9/numpy/core/src/common/templ_common.h', 'build/src.macosx-11.1-arm64-3.9/numpy/core/include/numpy/config.h', 'build/src.macosx-11.1-arm64-3.9/numpy/core/include/numpy/_numpyconfig.h', 'build/src.macosx-11.1-arm64-3.9/numpy/core/include/numpy/__multiarray_api.h', 'build/src.macosx-11.1-arm64-3.9/numpy/core/include/numpy/__ufunc_api.h']
            building extension "numpy.core._umath_tests" sources
            building extension "numpy.core._rational_tests" sources
            building extension "numpy.core._struct_ufunc_tests" sources
            building extension "numpy.core._operand_flag_tests" sources
            building extension "numpy.fft._pocketfft_internal" sources
            building extension "numpy.linalg.lapack_lite" sources
            building extension "numpy.linalg._umath_linalg" sources
            building extension "numpy.random._mt19937" sources
            building extension "numpy.random._philox" sources
            building extension "numpy.random._pcg64" sources
            building extension "numpy.random._sfc64" sources
            building extension "numpy.random._common" sources
            building extension "numpy.random.bit_generator" sources
            building extension "numpy.random._generator" sources
            building extension "numpy.random._bounded_integers" sources
            building extension "numpy.random.mtrand" sources
            building data_files sources
            build_src: building npy-pkg config files
            running build_py
            creating build/lib.macosx-11.1-arm64-3.9
            creating build/lib.macosx-11.1-arm64-3.9/numpy
            copying numpy/conftest.py -> build/lib.macosx-11.1-arm64-3.9/numpy
            copying numpy/version.py -> build/lib.macosx-11.1-arm64-3.9/numpy
            copying numpy/_globals.py -> build/lib.macosx-11.1-arm64-3.9/numpy
            copying numpy/__init__.py -> build/lib.macosx-11.1-arm64-3.9/numpy
            copying numpy/dual.py -> build/lib.macosx-11.1-arm64-3.9/numpy
            copying numpy/_distributor_init.py -> build/lib.macosx-11.1-arm64-3.9/numpy
            copying numpy/setup.py -> build/lib.macosx-11.1-arm64-3.9/numpy
            copying numpy/ctypeslib.py -> build/lib.macosx-11.1-arm64-3.9/numpy
            copying numpy/matlib.py -> build/lib.macosx-11.1-arm64-3.9/numpy
            copying numpy/_pytesttester.py -> build/lib.macosx-11.1-arm64-3.9/numpy
            copying build/src.macosx-11.1-arm64-3.9/numpy/__config__.py -> build/lib.macosx-11.1-arm64-3.9/numpy
            creating build/lib.macosx-11.1-arm64-3.9/numpy/compat
            copying numpy/compat/py3k.py -> build/lib.macosx-11.1-arm64-3.9/numpy/compat
            copying numpy/compat/__init__.py -> build/lib.macosx-11.1-arm64-3.9/numpy/compat
            copying numpy/compat/setup.py -> build/lib.macosx-11.1-arm64-3.9/numpy/compat
            copying numpy/compat/_inspect.py -> build/lib.macosx-11.1-arm64-3.9/numpy/compat
            creating build/lib.macosx-11.1-arm64-3.9/numpy/compat/tests
            copying numpy/compat/tests/__init__.py -> build/lib.macosx-11.1-arm64-3.9/numpy/compat/tests
            copying numpy/compat/tests/test_compat.py -> build/lib.macosx-11.1-arm64-3.9/numpy/compat/tests
            creating build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/umath.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/fromnumeric.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/_dtype.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/_add_newdocs.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/_methods.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/_internal.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/_string_helpers.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/multiarray.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/_asarray.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/records.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/__init__.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/setup_common.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/memmap.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/overrides.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/getlimits.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/_dtype_ctypes.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/defchararray.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/shape_base.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/machar.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/setup.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/numeric.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/function_base.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/einsumfunc.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/umath_tests.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/_ufunc_config.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/_exceptions.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/numerictypes.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/_type_aliases.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/cversions.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/arrayprint.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            copying numpy/core/code_generators/generate_numpy_api.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core
            creating build/lib.macosx-11.1-arm64-3.9/numpy/core/tests
            copying numpy/core/tests/test_numerictypes.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core/tests
            copying numpy/core/tests/test_scalar_methods.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core/tests
            copying numpy/core/tests/test_scalarmath.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core/tests
            copying numpy/core/tests/test_item_selection.py -> build/lib.macosx-11.1-arm64-3.9/numpy/core/tests
[…]
