### 2022-07-25 #4503 @catrionamurray
I had an error installing Eureka! I've opened a new issue here: https://github.com/kevin218/Eureka/issues/421
