### 2022-07-25 kevin218/Eureka#422 @catrionamurray
Mistake in the QuickStart docs?
### Instrument

Other (any stage)

### What happened?

Part of JOSS review (https://github.com/openjournals/joss-reviews/issues/4503). I believe there's a small mistake in the Quickstart docs:
https://eurekadocs.readthedocs.io/en/latest/quickstart.html

Where we set the `nlive` parameter:

> As a compromise, let’s use 256 live points instead to get a fairly nice corner plot but also speed up the fit, so set the following in S5_fit_par_wasp39b.epf:

I think this file should be "S5_wasp39b.ecf", not the .epf file?

### Error traceback output

_No response_

### What operating system are you using?

_No response_

### What version of Python are you running?

_No response_

### What Python packages do you have installed?

_No response_

### Code of Conduct

- [X] I agree to follow this project's Code of Conduct
