### 2022-07-25 kevin218/Eureka#423 @catrionamurray
Missing Science References in JOSS Paper
### Instrument

_No response_

### What happened?

In the JOSS paper (https://github.com/openjournals/joss-reviews/issues/4503) there are missing science references to JWST (and HST) and a lack of context. It would also be nice to have a very brief (1-2 line) introduction to JWST and its science for those who are unfamiliar given that this is an interdisciplinary journal.

### Error traceback output

_No response_

### What operating system are you using?

_No response_

### What version of Python are you running?

_No response_

### What Python packages do you have installed?

_No response_

### Code of Conduct

- [X] I agree to follow this project's Code of Conduct
