### 2022-07-25 #4503 @catrionamurray
There are missing references in the text to JWST, HST etc., issue here: https://github.com/kevin218/Eureka/issues/423
