### 2022-07-25 kevin218/Eureka#423 @catrionamurray
The `jwst` pipeline is also mentioned before it is referenced in the Stage Outline step, I would add the reference at its first mention.
