### 2022-07-25 kevin218/Eureka#421 @taylorbell57
Ah yes, we just realized that there is this issue with installing on M1 computers that is caused by h5py's pip installation not working for M1 Macs (#415). If you try the conda environment.yml installation method, I have heard that it does work though. I have yet to hear back from any M1 users as to whether the code actually then runs successfully, so any feedback on that end would be much appreciated! Closing this issue for now as a duplicate of #415
