### 2022-07-25 a5313d1
Minor docs updates
Noting to use environment.yml on M1 processors: resolves #415 and #421

Fixing small typo in quickstart: resolves #422
