### 2022-07-25 e8da5bd
Simplify quickstart by avoiding exotic-ld by default
I also added documentation there on what to change if you do want to use
exotic-ld instead.

Resolves #407
