### 2022-07-25 263ab67
Moving manual_clip to S5 and testing in MIRI test
