+++ src/eureka/S4_generate_lightcurves/generate_LD.py
@@ -2,6 +2,7 @@ from exotic_ld import StellarLimbDarkening
 import numpy as np
 import pandas as pd
 import glob
+import os
 
 
 def exotic_ld(meta, spec, log):
@@ -33,8 +34,8 @@ def exotic_ld(meta, spec, log):
                  mute=(not meta.verbose))
     
     # Check if exotic-ld directory includes csv files
-    exotic_ld_files = glob.glob(meta.exotic_ld_direc + '/**/*.flx',
-                                recursive=True)
+    exotic_ld_files = glob.glob(meta.exotic_ld_direc + os.sep + '**' + os.sep +
+                                '*.flx', recursive=True)
     if not exotic_ld_files:
         raise AssertionError('Unable to find ancillary files.' +
                              'Have you downloaded them? (see Zenodo link)')
