### 2022-07-26 331b02d
Updating meta.clip to meta.manual_clip
Fixes the broken pytest tests
