### 2022-07-26 2b692c2
should be passing instrument/citations properly now
