+++ src/eureka/S3_data_reduction/s3_reduce.py
@@ -124,25 +124,6 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
             meta.run_s3 = util.makedirectory(meta, 'S3', meta.run_s3,
                                              ap=spec_hw_val, bg=bg_hw_val)
 
-    # Store citations to relevant dependencies in the meta file
-    # get currently imported modules (top level only)
-    mods = np.unique([mod.split('.')[0] for mod in sys.modules.keys()])
-            
-    # get modules for which we have citations
-    citemods = np.intersect1d(mods, list(CITATIONS))
-
-    # check if meta has existing list of citations/bibitems, if it does, make sure we have imports from previous stages in our citations
-    try: 
-        hasattr(meta, 'citations')
-        citemods = np.union1d(citemods, meta.citations)
-    except AttributeError:
-        # otherwise, if we're starting from this stage, just use the current list of imports
-        pass
-            
-    # save list of imports and the bibliography to the meta object
-    meta.citations = citemods
-    meta.bibliography = np.concatenate([CITATIONS[entry][1] for entry in citemods])
-
     # begin process
     for spec_hw_val in meta.spec_hw_range:
         for bg_hw_val in meta.bg_hw_range:
@@ -415,6 +396,25 @@ def reduce(eventlabel, ecf_path=None, s2_meta=None):
                 # 2D light curve without drift correction
                 plots_s3.lc_nodriftcorr(meta, spec.wave_1d, spec.optspec,
                                         optmask=spec.optmask)
+
+            # Store citations to relevant dependencies in the meta file
+            # get currently imported modules (top level only)
+            mods = np.unique([mod.split('.')[0] for mod in sys.modules.keys()])
+                        
+            # get modules for which we have citations
+            citemods = np.intersect1d(mods, list(CITATIONS))
+
+            # check if meta has existing list of citations/bibitems, if it does, make sure we include imports from previous stages in our citations
+            if hasattr(meta, 'citations'):
+                citemods = np.union1d(citemods, meta.citations)
+                            
+                    
+            # save list of imports and the bibliography to the meta object
+            if meta.inst in citemods:
+                meta.citations = citemods
+            else:
+                meta.citations = np.append(citemods, meta.inst)
+            meta.bibliography = np.concatenate([CITATIONS[entry] for entry in citemods])
 
             # Save results
             if meta.save_output:
