+++ src/citations.py
@@ -1,6 +1,5 @@
 CITATIONS = {
-    "eureka": (
-        ("eureka"), 
+    "eureka": 
         [r"""@ARTICLE{eureka,
                 author = {{Bell}, Taylor J. and {Ahrer}, Eva-Maria and {Brande}, Jonathan and {Carter}, Aarynn L. and 
                 {Feinstein}, Adina D. and {Guzman Caloca}, Giannina and {Mansfield}, Megan and {Zieba}, Sebastian and 
@@ -18,10 +17,8 @@ CITATIONS = {
             primaryClass = {astro-ph.IM},
                 adsurl = {https://ui.adsabs.harvard.edu/abs/2022arXiv220703585B},
                 adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-            }"""]
-    ),
-    "numpy": (
-        ("numpy"),
+            }"""],
+    "numpy": 
         [r"""@Article{numpy,
             title         = {Array programming with {NumPy}},
             author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
@@ -43,10 +40,8 @@ CITATIONS = {
             doi           = {10.1038/s41586-020-2649-2},
             publisher     = {Springer Science and Business Media {LLC}},
             url           = {https://doi.org/10.1038/s41586-020-2649-2}
-        }"""]
-    ),
-    "matplotlib": (
-        ("matplotlib"),
+        }"""],
+    "matplotlib": 
         [r"""@Article{matplotlib,
             Author    = {Hunter, J. D.},
             Title     = {Matplotlib: A 2D graphics environment},
@@ -60,10 +55,8 @@ CITATIONS = {
             publisher = {IEEE COMPUTER SOC},
             doi       = {10.1109/MCSE.2007.55},
             year      = 2007
-            }"""]
-    ),
-    "scipy": (
-        ("scipy"),
+            }"""],
+    "scipy": 
         [r"""@ARTICLE{scipy,
             author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
                         Haberland, Matt and Reddy, Tyler and Cournapeau, David and
@@ -86,10 +79,8 @@ CITATIONS = {
             pages   = {261--272},
             adsurl  = {https://rdcu.be/b08Wh},
             doi     = {10.1038/s41592-019-0686-2},
-        }"""]
-    ),
-    "astropy": (
-        ("astropy"),
+        }"""],
+    "astropy": 
         [r"""@article{astropy2013,
             Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
             Adsurl = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A},
@@ -167,6 +158,140 @@ CITATIONS = {
         primaryClass = {astro-ph.IM},
             adsurl = {https://ui.adsabs.harvard.edu/abs/2018AJ....156..123A},
             adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-        }"""]
-    ),
+        }"""],
+    "nirspec": 
+        ["""@ARTICLE{birkmann2022,
[…]

[…]
