### 2022-07-28 10dd6bc
Speeding up S5 by simplifying Model.update
Previously, a new Parameter was made at each step, but now the
original parameter's value (and value only) is updated. This cuts the
fraction of time spent on Model.update by about 2x

The next step will be to update the _parse_coeffs which are currently
taking longer than they likely need to be
