+++ src/eureka/S5_lightcurve_fitting/models/Model.py
@@ -27,6 +27,7 @@ class Model:
         self._time = None
         self._flux = None
         self._units = q.day
+        self.freenames = None
         self._parameters = Parameters()
         self.components = None
         self.modeltype = None
@@ -105,24 +106,19 @@ class Model:
 
         return interp_flux
 
-    def update(self, newparams, names, **kwargs):
+    def update(self, newparams, **kwargs):
         """Update the model with new parameter values.
 
         Parameters
         ----------
         newparams : ndarray
             New parameter values.
-        names : list
-            Parameter names.
         **kwargs : dict
             Unused by the base
             eureka.S5_lightcurve_fitting.models.Model class.
         """
-        for ii, arg in enumerate(names):
-            if hasattr(self.parameters, arg):
-                val = getattr(self.parameters, arg).values[1:]
-                val[0] = newparams[ii]
-                setattr(self.parameters, arg, val)
+        for val, arg in zip(newparams, self.freenames):
+            getattr(self.parameters, arg).value = val
         self._parse_coeffs()
         return
 
@@ -410,7 +406,7 @@ class CompositeModel(Model):
         else:
             new_time = self.time
 
-        flux = np.ones(len(self.time)*self.nchan)
+        flux = np.ones(len(new_time)*self.nchan)
 
         # Evaluate flux at each model
         for model in self.components:
@@ -423,20 +419,18 @@ class CompositeModel(Model):
                     flux *= model.eval(**kwargs)
         return flux, new_time
 
-    def update(self, newparams, names, **kwargs):
+    def update(self, newparams, **kwargs):
         """Update parameters in the model components.
 
         Parameters
         ----------
         newparams : ndarray
             New parameter values.
-        names : list
-            Parameter names.
         **kwargs : dict
             Additional parameters to pass to
             eureka.S5_lightcurve_fitting.models.Model.update().
         """
         # Evaluate flux at each model
         for model in self.components:
-            model.update(newparams, names, **kwargs)
+            model.update(newparams, **kwargs)
         return
