### 2022-07-28 kevin218/Eureka#415 @catrionamurray
Having this issue with installation from conda .yml file, does it not contain `exotic-ld`?:
```
(base) camu5866@CatMsMBP Eureka % conda env create --file environment.yml --force
Solving environment: failed

ResolvePackageNotFound:
  - exotic-ld
```
