### 2022-07-28 kevin218/Eureka#415 @taylorbell57
Ah yes, you'll need to checkout the main branch of the GitHub repo as we've yet to issue a new version since fixing that bug
