### 2022-07-28 e2a19b3
Fixing the bug with pytest
Missed one old instance of meta.clip instead of meta.manual_clip
