### 2022-07-29 beb8f24
removed not needed deps from apphot
