### 2022-07-29 c4d3943
added interp grid and flag bad DQ to util.py
