### 2022-07-29 36e820c
added multiple new plots for photometry
