+++ src/eureka/S3_data_reduction/plots_s3.py
@@ -6,6 +6,8 @@ from .source_pos import gauss
 from ..lib import util
 from ..lib.plots import figure_filetype
 import scipy.stats as stats
+from matplotlib.colors import LogNorm
+from mpl_toolkits import axes_grid1
 
 
 def lc_nodriftcorr(meta, wave_1d, optspec, optmask=None):
@@ -455,7 +457,7 @@ def phot_bg(meta, data):
     plt.ylabel('Flux')
     plt.xlabel('Time')
     plt.tight_layout()
-    fname = (f'figs{os.sep}fig3305-1D_BG_LC' + figure_filetype)
+    fname = (f'figs{os.sep}fig3305-1D_LC_BG' + figure_filetype)
     plt.savefig(meta.outputdir+fname, dpi=300)
     if not meta.hide_plots:
         plt.pause(0.2)
@@ -465,90 +467,32 @@ def phot_centroid(meta, data):
     """
     Plots the (x, y) centroids and (sx, sy) the Gaussian 1-sigma half-widths as a function of time.
     """
-    plt.figure(3306)
+    plt.figure(3104)
     plt.clf()
     plt.suptitle('Centroid positions over time')
 
-    plt.subplot(411)
-    plt.plot(data.time, data.centroid_x-np.mean(data.centroid_x))
+    cx = data.centroid_x
+    cx_rms = np.sqrt(np.mean(((cx - np.mean(cx)) / cx) ** 2))
+    cy = data.centroid_y
+    cy_rms = np.sqrt(np.mean(((cy - np.mean(cy)) / cy) ** 2))
+
+    ax = plt.subplot(411)
+    plt.plot(data.time, data.centroid_x-np.mean(data.centroid_x), label='scatter in x = {0} pixels'.format(cx_rms))
     plt.ylabel('Delta x')
-
-    plt.subplot(412)
-    plt.plot(data.time, data.centroid_y-np.mean(data.centroid_y))
+    plt.subplot(412, sharex=ax)
+    plt.plot(data.time, data.centroid_y-np.mean(data.centroid_y), label='scatter in y = {0} pixels'.format(cy_rms))
     plt.ylabel('Delta y')
-
-    plt.subplot(413)
+    plt.subplot(413, sharex=ax)
     plt.plot(data.time, data.centroid_sy-np.mean(data.centroid_sx))
     plt.ylabel('Delta sx')
-
-    plt.subplot(414)
+    plt.subplot(414, sharex=ax)
     plt.plot(data.time, data.centroid_sy-np.mean(data.centroid_sy))
     plt.ylabel('Delta sy')
     plt.xlabel('Time')
-    plt.tight_layout()
-    fname = (f'figs{os.sep}fig3306-Centroid' + figure_filetype)
-    plt.savefig(meta.outputdir + fname, dpi=250)
-    if not meta.hide_plots:
-        plt.pause(0.2)
-
-
-def phot_centroid_fgc(img, x, y, sx, sy, i, m, meta):
-    plt.figure(3502)
-    plt.clf()
-    plt.suptitle('Centroid gaussian fit')
-    fig, ax = plt.subplots(2,2, figsize=(8,8))
-    fig.delaxes(ax[1,1])
-    ax[0,0].imshow(img, vmax=5e3, origin='lower', aspect='auto')
-
-    ax[1,0].plot(range(len(np.sum(img, axis=0))), np.sum(img, axis=0))
-    x_plot = np.linspace(0, len(np.sum(img, axis=0)))
-    ax[1,0].plot(x_plot, stats.norm.pdf(x_plot, x, sx)/np.max(stats.norm.pdf(x_plot, x, sx))*np.max(np.sum(img, axis=0)))
-
-    ax[0,1].plot(np.sum(img, axis=1), range(len(np.sum(img, axis=1))))
[…]
