### 2022-07-29 904e374
Fixing flake8 issues
There is also some strange bug where the Batman model's ephemeris ends
up getting fit terribly in this branch but not in the main branch. Not
sure why yet...
