### 2022-07-29 4dd4c36
Add Fig3304, which plots the median, BG-subtracted frame to study the residual BG region and aperture/BG sizes
