### 2022-07-29 0b77844
Fixing bug with poor fits from faster code
The Parameters.dict values are not Parameter objects, so updating the
Parameter object did not update the dict values which some models make
use of. It'll likely be worth re-evaluating whether we want to have
duplicate copies of the parameters values
