### 2022-07-29 ee31ccd
Skip figure 3304 when using WFC3
