### 2022-07-29 705dbd8
Adding DQ array
