### 2022-07-29 7238248
Make vmin,vmax optional parameters
