### 2022-07-30 c3e0694
Fixing Fig4101 for unbinned spectra
Before the plot would break for unbinned spectra if wave_min and
wave_max were not set to the full wavekength extent
