### 2022-07-30 e13dc60
Replace masked pixels in medflux using interpolation
