+++ src/eureka/S3_data_reduction/plots_s3.py
@@ -466,7 +466,7 @@ def residualBackground(data, meta, m, vmin=-200, vmax=1000):
         First version
     '''
     # Median flux of segment
-    # Don't us masked arrays so that we can see flux in masked areas
+    # Don't use masked arrays so that we can see flux in masked areas
     subdata = np.ma.masked_where(~data.mask.values, data.flux.values)
     flux = np.ma.median(subdata, axis=0)
     # flux = np.median(data.flux, axis=0)
@@ -480,7 +480,7 @@ def residualBackground(data, meta, m, vmin=-200, vmax=1000):
     cmap = mpl.cm.get_cmap("plasma").copy()
     cmap.set_bad('k', 1.)
 
-    plt.figure(3304)
+    plt.figure(3304, figsize=(8, 3.5))
     plt.clf()
     fig, (a0, a1) = plt.subplots(1, 2, gridspec_kw={'width_ratios': [3, 1]},
                                  num=3304, figsize=(8, 3.5))
