### 2022-08-01 #4503 @christinahedges
Hi all, my review is on-going as I have been having installation struggles with this tool. I'll keep working on it, but for now I've got this feedback:

## Installation
* I have had some installation issues. Given that some people don’t use conda, I tried to use a different environment/package manager. I had some of the same troubles people seem to be having installing this tool on an M1 Mac. I believe there are some M1 specific installation instructions that are going live on the docs, though these are to install via conda. I'm figuring out ways to get this to install.

* The installation instructions point to a github repo that seems to be being updated at the moment with bug fixes and new functionality. Can authors highlight what version is their stable version for the purposes of this review?

## Documentation

* **Key feedback:** The documentation (paper and repo documentation) needs to make more obvious clear a) some of these functions are wrappers to the `jwst` package, b)  that Eureka has alternatives for `jwst` package routines (e.g. ramp fitting) and c) why/when users should choose one or the other. A good place to do this might be a docs page explaining “what is Eureka”, which would explain what the package is and what users can expect to achieve with it, with some discussion on these points.

* There are API docs, but some of the API documentation is missing, for example:

Several of the docstrings look like this, and so it’s not clear what these functions do. If these functions are not part of the stable release for this review, please can authors discuss here?

* The API docs could benefit from some description of a demo workflow. As it stands, because the workflow is behind some scripts like `run_eureka.py` or functions e.g. `calibrateJWST` it’s not immediately clear to the reader what is happening under the hood, or what functions are being called in what order.

* Inside the documentation there’s not a discussion of the methods behind the functions (which is called out in the JOSS checklist). In the quickstart documentation there is a list of what is happening (e.g. ramp fitting) but not a detailed description of the methods/assumptions made in each step. It may be that this is not where the authors intend to put this information, in which case could they point me towards the right documentation about the methods?

### Example Usage

* There is an example usage in the docs online, and example config files in the ‘demos` folder. It could be made more clear how to use this example. The `demos` folder would benefit from a readme to explain how to use this example.

* Authors might want to include the schematic from the paper inside their docs.

## Tests

* There are tests, and they are run automatically with CI. Looking at the tests, they are designed to import the relevant modules, run scripts on test data, and check that the resultant files are produced and that figures are produced. These tests do check that the tools run and produce an output, but don’t seem to check that the output is valid or correct. This does meet a minimum test that files are produced by the routines, but it does not thoroughly test the routines.

## High Level Feedback

The Eureka tool works based on config files, and has multiple "stages" that are imported. The tool is run with scripts, as a pipeline. Because of this, the API can be quite hard to read to understand what’s going on, and for the user the tool can seem to be a bit of a black box, for example this is a workflow in the tests:

```python
s2_meta = s2.calibrateJWST(meta.eventlabel, ecf_path=ecf_path)
s3_spec, s3_meta = s3.reduce(meta.eventlabel, ecf_path=ecf_path,
                                 s2_meta=s2_meta)
s4_spec, s4_lc, s4_meta = s4.genlc(meta.eventlabel, ecf_path=ecf_path,
                                       s3_meta=s3_meta)
s5.fitlc(meta.eventlabel, ecf_path=ecf_path, s4_meta=s4_meta)
```

It's not easily understandable to the reader what's happening in these steps. This is a common workflow in astronomy, and I am not advocating that you change any of these design decisions. However I wanted to highlight this, because I believe we need to ensure the reader has some information in the docs about what steps are happening inside these functions, and what methods are being used. This has lead to my feedback above.

I'll keep working on this review, but wanted to give some feedback as I go.
