### 2022-08-01 #4503 @taylorbell57
A quick response to a couple of your points - I'll try to address a couple of these today and the rest throughout the week.

> * I have had some installation issues. Given that some people don’t use conda, I tried to use a different environment/package manager. I had some of the same troubles people seem to be having installing this tool on an M1 Mac. I believe there are some M1 specific installation instructions that are going live on the docs, though these are to install via conda. I'm figuring out ways to get this to install.

Yeah, unfortunately none of our development team has an M1 Mac (Linux and Intel Macs only), so there's nothing we can do personally to test on M1 chips and we're left to relying on user-submitted issues which we've tried to resolve quickly. I think I and others avoided M1 chips for the time being knowing that a lot of software and packages would have issues with the new processor for a year or few. At this point, there are no known issues with our code itself on M1, but a couple of our dependencies have issues with M1 and there's little we can do ourselves about that. For several of the known installation issues, the problems can be resolved using the conda installation methods as some of the packages do seem to have been built with M1 knowledge on conda but not on pip.

> * The installation instructions point to a github repo that seems to be being updated at the moment with bug fixes and new functionality. Can authors highlight what version is their stable version for the purposes of this review?

Yeah, the code has been under pretty rapid development with the release of the first JWST observations. Things are stabilizing more with time though, so let's use the latest release from yesterday (v0.4; https://github.com/kevin218/Eureka/tree/v0.4; https://eurekadocs.readthedocs.io/en/v0.4/). The `joss` branch is now up-to-date with that v0.4 but also has the current paper draft which is not in the main branch. Note that changes to the paper draft to address your comments are only going to be present in the `joss` branch (which editorialbot uses here to build the paper), and for the docs/code issues you've raised the changes will be made on the `main` branch (https://github.com/kevin218/Eureka; https://eurekadocs.readthedocs.io/en/latest/) and occasionally released as stable versions.

> * There are API docs, but some of the API documentation is missing

Ah yeah, several of these are functions that could potentially be used some day for WFC3 analyses but are not used and aren't likely to be used anytime soon, so I'll just remove them from the repo shortly.

> * The API docs could benefit from some description of a demo workflow. As it stands, because the workflow is behind some scripts like `run_eureka.py` or functions e.g. `calibrateJWST` it’s not immediately clear to the reader what is happening under the hood, or what functions are being called in what order.

Agreed, we've discussed making flowcharts for each stage (https://github.com/kevin218/Eureka/issues/113), but we didn't want to start on that until we'd run some real JWST data through the pipeline as we knew things were subject to significant change. Now that things are settling down, making such flowcharts would certainly be beneficial.

> * Inside the documentation there’s not a discussion of the methods behind the functions (which is called out in the JOSS checklist). In the quickstart documentation there is a list of what is happening (e.g. ramp fitting) but not a detailed description of the methods/assumptions made in each step. It may be that this is not where the authors intend to put this information, in which case could they point me towards the right documentation about the methods?

This will take some more thinking on our end - I will get back to you. The flowcharts will help some, but not completely.

> * There is an example usage in the docs online, and example config files in the `demos` folder. It could be made more clear how to use this example. The`demos` folder would benefit from a readme to explain how to use this example.

Would a simple README.md which points the users to the quickstart tutorial be adequate for what you're thinking?

> * Authors might want to include the schematic from the paper inside their docs.

Ah, indeed - I guess we forgot to include that in the docs.

> * There are tests, and they are run automatically with CI. Looking at the tests, they are designed to import the relevant modules, run scripts on test data, and check that the resultant files are produced and that figures are produced. These tests do check that the tools run and produce an output, but don’t seem to check that the output is valid or correct. This does meet a minimum test that files are produced by the routines, but it does not thoroughly test the routines.

Agreed, we've discussed this too (in issue https://github.com/kevin218/Eureka/issues/140 and during video calls) - the hard part is spoofing data for each of the functions for which we know the expected result. We're unlikely to write such tests for Stage 2 (as there's basically no code in there) and writing them for Stage 1 would be very challenging (due to the large file sizes of the input uncal files for even ~3 integrations), but the functions in Stages 3-6 are good candidates for detailed unit testing. As it stands, the already existing tests have been very helpful as they test end-to-end and the code typically crashes somewhere along the way if something is off, but I totally agree that a lack of crashes does not imply the correct result has been produced.

> ... Because of this, the API can be quite hard to read to understand what’s going on, and for the user the tool can seem to be a bit of a black box ...

Agreed, and hopefully this will be significantly ameliorated with the flowcharts I described above. We've also talked about making it easier for the user to call each of the steps in the stages themselves to give a better understanding of the flow and procedure (and make customization easier)
