### 2022-08-01 544d444
Adding requested science references from #423
