+++ joss/paper.bib
@@ -1,3 +1,115 @@
+@ARTICLE{JWST:2006,
+       author = {{Gardner}, Jonathan P. and {Mather}, John C. and {Clampin}, Mark and {Doyon}, Rene and {Greenhouse}, Matthew A. and {Hammel}, Heidi B. and {Hutchings}, John B. and {Jakobsen}, Peter and {Lilly}, Simon J. and {Long}, Knox S. and {Lunine}, Jonathan I. and {McCaughrean}, Mark J. and {Mountain}, Matt and {Nella}, John and {Rieke}, George H. and {Rieke}, Marcia J. and {Rix}, Hans-Walter and {Smith}, Eric P. and {Sonneborn}, George and {Stiavelli}, Massimo and {Stockman}, H.~S. and {Windhorst}, Rogier A. and {Wright}, Gillian S.},
+        title = "{The James Webb Space Telescope}",
+      journal = {\ssr},
+     keywords = {galaxies: formation, infrared: general, planetary systems, space vehicles: instruments, stars: formation, Astrophysics},
+         year = 2006,
+        month = apr,
+       volume = {123},
+       number = {4},
+        pages = {485-606},
+          doi = {10.1007/s11214-006-8315-7},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/0606175},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2006SSRv..123..485G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{MIRI:2015,
+       author = {{Rieke}, G.~H. and {Wright}, G.~S. and {B{\"o}ker}, T. and {Bouwman}, J. and {Colina}, L. and {Glasse}, Alistair and {Gordon}, K.~D. and {Greene}, T.~P. and {G{\"u}del}, Manuel and {Henning}, Th. and {Justtanont}, K. and {Lagage}, P. -O. and {Meixner}, M.~E. and {N{\o}rgaard-Nielsen}, H. -U. and {Ray}, T.~P. and {Ressler}, M.~E. and {van Dishoeck}, E.~F. and {Waelkens}, C.},
+        title = "{The Mid-Infrared Instrument for the James Webb Space Telescope, I: Introduction}",
+      journal = {\pasp},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2015,
+        month = jul,
+       volume = {127},
+       number = {953},
+        pages = {584},
+          doi = {10.1086/682252},
+archivePrefix = {arXiv},
+       eprint = {1508.02294},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2015PASP..127..584R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@INPROCEEDINGS{NIRCam:2004,
+       author = {{Horner}, Scott D. and {Rieke}, Marcia J.},
+        title = "{The near-infrared camera (NIRCam) for the James Webb Space Telescope (JWST)}",
+    booktitle = {Optical, Infrared, and Millimeter Space Telescopes},
+         year = 2004,
+       editor = {{Mather}, John C.},
+       series = {Society of Photo-Optical Instrumentation Engineers (SPIE) Conference Series},
+       volume = {5487},
+        month = oct,
+        pages = {628-634},
+          doi = {10.1117/12.552281},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2004SPIE.5487..628H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@INPROCEEDINGS{NIRSpec:2007,
[…]
