### 2022-08-01 9aae032
Adding context on JWST science for #423
