### 2022-08-01 bbdf830
Differentiating from jwst pipeline for #423
