### 2022-08-01 kevin218/Eureka#423 @taylorbell57
My edits just now should have resolved your above requests, so I'm marking this as closed for now, but please reopen if you feel there is anything missing still
