### 2022-08-03 #4503 @taylorbell57
Following up on the above, I just made a PR (https://github.com/kevin218/Eureka/pull/442) which adds to the documentation website some more flowcharts which describe the workflow for Eureka!'s Stages 3 and 4 (visible here for now https://eurekadocs.readthedocs.io/en/joss_docs/stages.html). I didn't bother to make flowcharts for Stages 5 and 6 as they're not that complicated or deeply nested. I added links to the documentation for Stages 1 and 2 made by the STScI as any edits we currently allow for Stage 1 are fully experimental.

That PR should also resolve the issues you pointed out about undocumented or poorly documented functions in the API.

Still unaddressed are comments:
1. Add unit tests rather than just relying on end-to-end tests
2. The demos folder would benefit from a readme to explain how to use this example.
    a. Please clarify if a README file with a link to the quickstart guide is all that is needed here
3. Inside the documentation there’s not a discussion of the methods behind the functions
    a. Can you comment on how detailed you are thinking here? With so many functions, this could pretty quickly get out of hand.

Let me know if you feel that any of your other comments have not been addressed now (either with edits in https://github.com/kevin218/Eureka/pull/442 or in the discussion above).
