### 2022-08-03 #4503 @christinahedges
Hi @taylorbell57

The new flow charts look good!

> Please clarify if a README file with a link to the quickstart guide is all that is needed here

Yes a README is what's needed, with a brief explanation of what the demos are and a pointer to the quickstart sounds fine.

>  Can you comment on how detailed you are thinking here? With so many functions, this could pretty quickly get out of hand.

I think it would be worth having @dfm weigh in here. Currently, as a user, I think there's not a place that I can go to find the methods this pipeline is using under the hood. @dfm, to meet the documentation of methods here, what should the authors be aiming for?
