### 2022-08-03 #4503 @dfm
I think that overall the state of the docs looks good, but I think your point is a good one @christinahedges. In many cases the docstrings do leave something to be desired in terms of narrative discussion of what the method is doing in. I wouldn't consider this a showstopper for the JOSS review, but I also think that the user experience could be substantially improved with some expansion of the docstrings for the key user-facing methods, or more narrative structure in the high-level summary pages.

In summary: from my perspective the minimum requirements are certainly satisfied, but I think that some effort in adding some detail to some key subset of docstrings would go a long way!
