### 2022-09-19 #4503 @taylorbell57
Aha, found the typo - the STScI had a missing } in the bibtex snippet they gave me. Should compile fine for you now!
