### 2022-09-19 #4503 @catrionamurray
Thanks Taylor!

From my end the minimum checklist has been completed, I just reran through the (M1 chip) installation and Quickstart using the v0.5 of Eureka! and I'm happy that it installs and runs as it should. A small note is that the Quickstart plots are probably from an older version, so the plots/MAD values etc. I got following the tutorial were slightly different. Not a big deal, but I thought it was worth mentioning.

The docs are also much improved since the start of this review process. But I definitely agree with @christinahedges points about having more informative method descriptions in the API section (or at least links to relevant docs) where you describe the inputs/outputs/methods for each step compared to the one-line summaries that are there currently (e.g. "Fit sky background with out-of-spectra data."). Also the point about tests that don't just check if the code ran without crashing (or basic quality checks, e.g. are the final values physical?). However, these changes may take a bit more time and thought and are probably beyond the scope of this JOSS review, as per Dan's comment.

Thanks for the work you've all put in, I'm sure this will be (and already is) an invaluable tool for analyzing the JWST data!!
