### 2022-09-20 #4503 @taylorbell57
Amazing, I'm very glad to hear that it is now working on M1!! Good point on the quickstart figures - I'll try to remember to update those when I get the chance.

For the docs, how much of this has been addressed by the flowcharts found [here](https://eurekadocs.readthedocs.io/en/latest/stages.html). Are you suggesting a link to that page from the top of the API page or from within some key functions, or are you meaning that you want more of a text description of each function in the API? A detailed paragraph for every function honestly isn't going to be feasible (especially in the short term), but adding one to some of the higher level functions that users are most likely to directly interact with is probably doable. I can try to do this fairly soon, but please clarify whether this is something that needs to be done before JOSS acceptance.

For unit tests, this is definitely on our to-do list, but will take substantial amounts of time and I don't feel it is required at this point (although I agree it would be ideal to have them by now). We're just all at or near the limits of what we can take on with all this ERS, GO, and GTO data flooding in, and none of us are getting paid to work on Eureka so we're limited in how we can spend our time.

Thanks, it has been very rewarding seeing Eureka! used in each of the ERS papers that are in prep - exciting seeing all this teamwork paying off!
