### 2022-09-20 #4503 @catrionamurray
I don't think the improvements on either the docs or the tests are necessary for JOSS acceptance (though maybe @dfm has more insight into this). However, I think it would be great for the user if there should be something along the lines of what exists for the JWST pipeline (https://jwst-pipeline.readthedocs.io/en/latest/jwst/background_step/description.html) at least for the major steps beyond Stage 2/differences from the JWST pipeline. Especially since this tool is designed to replace several major stages in the JWST pipeline, which is very well documented.

As an example of what I'm referring to, in the Eureka overview docs I see descriptions like "Stage 3: Using Stage 2 outputs, performs background subtraction and optimal spectral extraction". It would be nice to have more detailed paragraph for what major methods like **background subtraction** and **optimal spectral extraction** mean. Say I was having some weird residuals in my light curves and I wanted to check/make changes to the background subtraction step. If I look at the flowchart I see:

![Screen Shot 2022-09-20 at 9 52 15 AM](https://user-images.githubusercontent.com/34891147/191305775-6788ddaa-4936-4ca6-b1bf-58abf27a2369.png)

Then I go to the API to see what inst.fit_bg() does, but I can only (at least easily) find:

![Screen Shot 2022-09-20 at 9 56 47 AM](https://user-images.githubusercontent.com/34891147/191306776-5d4a1aee-c29c-4fe7-948d-90ac6841a09d.png)
![Screen Shot 2022-09-20 at 9 57 24 AM](https://user-images.githubusercontent.com/34891147/191306892-009783ce-f3d1-4b3e-85b8-2eb95baf6b84.png)
![Screen Shot 2022-09-20 at 10 11 50 AM](https://user-images.githubusercontent.com/34891147/191310079-b7bc4b71-ce86-474c-9e25-558ba2d0a982.png)

Which doesn't really tell me what these stages are doing without digging into the code itself. I understand that it may not be feasible for all functions that Eureka! runs in the short-term but I think it would definitely improve the user experience to have more detail (e.g. the JWST Pipeline docs) for methods that can significantly affect the final results. Even if it was limited to brief descriptions of the 5-10 methods mentioned in the overview for now I think this would be a great step!

I hope that makes sense?
