### 2022-09-23 #4503 @dfm
> I don't think the improvements on either the docs or the tests are necessary for JOSS acceptance

Agreed! We want to provide constructive feedback to improve the usability and reliability of scientific software, but we don't want to be too hard line. I agree with @catrionamurray and @christinahedges that these documentation improvements should be high priority for the development team to increase the impact of Eureka, but a complete overhaul is not necessary for this submission as far as I'm concerned!
