### 2022-10-10 #4503 @dfm
@taylorbell57 — I've gone through the checklist and happily checked all but 1 of the boxes. You may have also noticed the issue and PR that I opened. After you take a look at those, I'd say that the only remaining issue here is the "State of the field" in the software paper section of the checklist. I know that some alternative tools are listed in the paper, but I think it would be great to provide a few more words of context about where Eureka fits within this ecosystem. Personally I think this should be in *both* the paper and the docs.
