### 2022-10-11 kevin218/Eureka#415 @kcrasmus
I have the M1 chip, does that make a difference?

thanks,
kaitlin

On Wed, Jul 20, 2022 at 2:08 PM Taylor James Bell ***@***.***>
wrote:

> Can you also confirm your Mac has an Intel processor?
>
