### 2022-10-11 kevin218/Eureka#415 @taylorbell57
Are you still having issues installing Eureka? If so, have you followed the latest installation instructions, as I have heard from others that the conda installation instructions now work for them on an M1 mac
