### 2022-10-11 kevin218/Eureka#472 @taylorbell57
Making myself a terse checklist to keep track of what I've already taken care of
- [ ] Include a high-level overview of how Eureka works either at the beginning of the quickstart document, or on the index page, or both. Clarify that users will continue to use this same format of interacting with the pipeline
- [x] Remove (sys.path.insert(0, f'..{os.sep}..{os.sep}')) from run_eureka.py
- [ ] Clarify in the quickstart that inputdir and outputdir are both relative to topdir.
- [ ] Investigate deprecation warnings and sqrt of negative number warnings, and reassure users in quickstart if they can. be ignored.
- [ ] Update quickstart figures so that users get the same results as those in the tutorial (to offer a kind of validation test)

<br/><br/>
As for your comment:
> ... I'd say that overall this interface feels clunky: is there a reason why topdir can't be defined once and then reused in the other stages with the subdirectory names can be set automatically?

- I get your feeling, but there are good reasons that things are this way. If someone renames a folder or moves it to a different hard drive, then assuming too much about the file structure and its permanence breaks things. And someone may want to email me their Stage 3 output folder and let me try running Stages 4-5 myself, and then the topdir and subdirectories will change. Or if I rename a particular run to some other name manually to remind me that this folder is my best reduction, then I need to be able to pass that renamed folder's name into Eureka. Previously I had written things to have a lot of assumptions about folder names, and we ran into a tonne of issues. This level of specificity and automation seems to have worked well though. In general, you set the same topdir in each of your ECFs, set the inputdir and outputdirs to just Stage N-1 and Stage N, and then you're good to go. If you want to analyze an old Stage3 output for example, then you can change inputdir in Stage 4 to "Stage3/S3_2022-10-01_ wasp39b_run1" or something like that to specify that the code should not use the most recent folder but instead that specific folder.
