### 2022-10-18 #4503 @taylorbell57
Okay @dfm, I believe that all your quickstart related comments are addressed by PR https://github.com/kevin218/Eureka/pull/474 and your "State of the field" comment is addressed by PR https://github.com/kevin218/Eureka/pull/477.
