### 2022-10-24 #4503 @dfm
@taylorbell57 — Looks great to me! I'll need the last few steps from you now to finish the processing:

1. Take one last read through the generated manuscript (linked above ☝️) to make sure that you're happy with it (it's harder to make changes later!), especially the names and affiliations. I've taken a pass and it looks good to me!
