### 2022-11-03 #4503 @taylorbell57
@dfm

1. Done!
2. Done - the version number is: v0.6
3. Done - the DOI is: 10.5281/zenodo.7278300
