### 2022-11-03 #4503 @dfm
@catrionamurray, @christinahedges — Many thanks for your constructive reviews! JOSS is a volunteer effort and we couldn't do it without people like you.

@taylorbell57 — Your paper has now been published in JOSS! Thanks for your submission, and your responses to all the reviewer comments.
